==> 00_load_data_libraries.Rmd <==
---
title: "test_browsing"
author: "anne"
date: "31 oktober 2018"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r}
library(readr)
browse <- read_delim("rawdata/Beitetrykk og forstyrrelse  i 9 og 10 takst pa artsgruppeniva.csv", ";", escape_double = FALSE, locale = locale(decimal_mark = ","), na = "NA", trim_ws = TRUE)
browse <- as.data.frame(browse)

browse0.1 <- read_delim("rawdata/Beitetrykk_og_forstyrrelse _i_9_og_10_takst_pa_artsgruppeniva_2019-01-31.csv", 
+     ";", escape_double = FALSE, locale = locale(decimal_mark = ",", 
+         grouping_mark = "."), na = "null", 
+     trim_ws = TRUE)

browse0.1 <- read_csv("rawdata/Beitetrykk_og_forstyrrelse _i_9_og_10_takst_pa_artsgruppeniva_2019-01-31.csv", na = c("null", "NA"))


# 
# browse <- read.csv2("rawdata/Beitetrykk og forstyrrelse i 9 og 10 takst pa artsgruppeniva.csv", sep = ";", dec = ",")

browse0.1 <- as.data.frame(browse0.1)

save(browse0.1, file = "browse0.1.rda")


load("rawdata/browse0.1.rda")




## ungskogspleie
browse0.2 <- read_delim("rawdata/Beitetrykk_og_forstyrrelse_i_9_10_og_11_takst_pa_artsgruppeniva.csv", ";", escape_double = FALSE, na = "NA", trim_ws = TRUE)

names(browse0.2)
browse0.2_ung <- browse0.2[, c(1,5,55,94,108,110,116,118)]
names(browse0.2_ung)

save(browse0.2_ung, file = "processdata/browse0.2_ung.rda")


```



```{r}
# library("readr", lib.loc="C:/Programs/R/R-3.5.1/library")
library("lme4", lib.loc="C:/Programs/R/R-3.5.1/library")
library("ggplot2", lib.loc="C:/Programs/R/R-3.5.1/library")
library("plyr", lib.loc="C:/Programs/R/R-3.5.1/library")
library("mgcv", lib.loc="C:/Programs/R/R-3.5.1/library")
# library("AICcmodavg", lib.loc="C:/Programs/R/R-3.5.1/library")
library("visreg", lib.loc="C:/Programs/R/R-3.5.1/library")
library("lattice", lib.loc="C:/Programs/R/R-3.5.1/library")
# library("latticeExtra", lib.loc="C:/Programs/R/R-3.5.1/library")
# library("car", lib.loc="C:/Programs/R/R-3.5.1/library")
# library("pscl", lib.loc="C:/Programs/R/R-3.5.1/library")
library("effects", lib.loc="C:/Programs/R/R-3.5.1/library")
# library("lmerTest", lib.loc="C:/Programs/R/R-3.5.1/library")
library("dplyr", lib.loc="C:/Programs/R/R-3.5.1/library")
library("lmtest", lib.loc="C:/Programs/R/R-3.5.1/library")
library("betareg", lib.loc="C:/Programs/R/R-3.5.1/library")
library("glmmTMB", lib.loc="C:/Programs/R/R-3.5.1/library")


```
