==> 01_cleaning_browse_data.Rmd <==
---
title: "01_cleaning_browse_data"
author: "anne"
date: "31 oktober 2018"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



```{r}
load("processdata/browse.rda")
load("processdata/browse_sub.rda")
```

## Cleaning data
```{r}
class(browse)
str(browse)
head(browse, n = 3)
names(browse)

str(browse_sub)
head(browse_sub, n = 3)
names(browse-sub)

```

Create a subset of browse with less variables
```{r}
browse_sub  <- browse[,c(1:5,7:9,15:24,27:45,51,59,61:67,69,74,76:83,85,87:90,92:97,99:105,108:122)]
head(browse_sub, n = 3)

save(browse_sub, file = "browse_sub.rda")


load("processdata/browse_sub.rda")
head(browse_sub)
names(browse_sub)
```



##### change column names, structure (numeric, integer, factor), and level names for factors
```{r}
# new column


# col. names

# colnames(fdata1)[colnames(fdata1)=="ï..ID"] <- "ID"


# structure 
browse_sub$FYLNR <- as.integer(browse_sub$FYLNR)
browse_sub$KOMNR <- as.integer(browse_sub$KOMNR)
browse_sub$elgbROSprs10 <- as.numeric(browse_sub$elgbROSprs10)
browse_sub$elgbFURUprs10 <- as.numeric(browse_sub$elgbFURUprs10)
browse_sub$elgbBJORKprs10 <- as.numeric(browse_sub$elgbBJORKprs10)
browse_sub$ROS_0_PRHA <- as.integer(browse_sub$ROS_0_PRHA)
browse_sub$ROS_0_25_PRHA <- as.integer(browse_sub$ROS_0_25_PRHA)
browse_sub$ROS_25_49_PRHA <- as.integer(browse_sub$ROS_25_49_PRHA)
browse_sub$Bjork_0_PRHA <- as.integer(browse_sub$Bjork_0_PRHA)
browse_sub$Bjork_0_25_PRHA <- as.integer(browse_sub$Bjork_0_25_PRHA)
browse_sub$Bjork_25_49_PRHA <- as.integer(browse_sub$Bjork_25_49_PRHA)
browse_sub$elgFURUha10 <- as.numeric(browse_sub$elgFURUha10)
browse_sub$elgROSha10 <- as.numeric(browse_sub$elgROSha10)
browse_sub$elgBJORKha10 <- as.numeric(browse_sub$elgBJORKha10)
browse_sub$elgLAUVha10 <- as.numeric(browse_sub$elgLAUVha10)
browse_sub$treartgruppe <- as.factor(browse_sub$treartgruppe)
browse_sub$skogkategori <- as.factor(browse_sub$skogkategori)
browse_sub$kant <- as.factor(browse_sub$kant)
browse_sub$region <- as.factor(browse_sub$region)
browse_sub$landsdel <- as.factor(browse_sub$landsdel)
browse_sub$tretetthet0 <- as.numeric(browse_sub$tretetthet0)
browse_sub$tretetthet025 <- as.numeric(browse_sub$tretetthet025)
browse_sub$tretetthet2549 <- as.numeric(browse_sub$tretetthet2549)
browse_sub$tretetthetpluss50 <- as.numeric(browse_sub$tretetthetpluss50)
browse_sub$tretetthetminus25 <- as.numeric(browse_sub$tretetthetminus25)

```



## Subsetting

Subset of "browse_sub" (several) on 9th and 10th takst
```{r}
browse_910 <- split(browse_sub, browse_sub$takst)

browse_9 <- browse_910[['9']]
browse_10 <- browse_910[['10']]


save(browse_9, file = "browse_9.rda")
save(browse_10, file = "browse_10.rda")


browse_9 <- browse_9[,c(1:8,10:18,26:28,32:39,41:46,48:52,58:61,68:74,79:82,89)]

browse9bro <- browse_9[,c(1:8, 28, 35, 40:55)] # for browsing pressure response
browse9tree <- browse_9[,c(1:8, 10:12, 28, 32, 35:36, 40:54)] # for tree density/ number response

# browse_9$elgROSprs9 <- NULL
# browse_9$elgLAUVprs9 <- NULL
# browse_9$elgFURUprs9 <- NULL


```

Creating subsets for different data
```{r}

browse9bro_disthus <- browse9bro[which(browse9bro$distvei1 < 1501 & browse9bro$disthus > 499), ] # longest distance to any road (that is further than 500m from a house) 1500m

browse9bro_dist <- browse9bro[browse9bro$distvei1 < 1501, ] # longest distance to any public road 1500m


# 
# browse2000 <- browse_sub[browse_sub$disthouse < 2001, ] # longest distance to house 2000m
# browse1500 <- browse_sub[browse_sub$distroad500 < 1501, ] # longest distance to any road (that is further than 500m from a house) 1500m
# browse1500pub <- browse_sub[browse_sub$distpubroad500 < 1501, ] # longest distance to public roads (that are further than 500m from a house) 1500m
# 
# browseros <- browse2000[browse2000$treegroup=="raw", ]
# browserosdistroad <- browse1500[browse1500$treegroup=="raw", ]
# browserosdistpubroad <- browse1500pub[browse1500pub$treegroup=="raw", ]
# 
# 
# 
# browse2000 <- browse_sub[browse_sub$disthouse < 2001, ] # longest distance to house 2000m
# browse1500 <- browse_sub[browse_sub$distroad500 < 1501, ] # longest distance to any road (that is further than 500m from a house) 1500m
# # browse1500pub <- browse_sub[browse_sub$distpubroad500 < 1501, ] # longest distance to public roads (that are further than 500m from a house) 1500m
# 
# browseroad1500 <- browse_sub[browse_sub$distroad1 < 1501, ]
# broad1500house500 <- browse_sub[which(browse_sub$distroad1 < 1501 & browse_sub$disthouse > 499), ]
# 
# browse9_dist <- browse_9[which(browse_9$distvei1 < 1501 & browse_9$disthus > 499), ] # longest distance to any road (that is further than 500m from a house) 1500m
# browse10_dist <- browse_10[which(browse_10$distvei1 < 1501 & browse_10$disthus > 499), ] # longest distance to any road (that is further than 500m from a house) 1500m
# 
# browse9_dist2 <- browse_9[which(browse_9$distvei1 < 1501 & browse_9$disthus > 499 & browse_9$elgbROSprs9 < 101), ] # longest distance to any road (that is further than 500m from a house) 1500m and browsingpressure is under 101%
# 
# # browseros <- browse2000[browse2000$treegroup=="raw", ]
# # browserosdistroad <- browse1500[browse1500$treegroup=="raw", ]
# # browserosdistpubroad <- browse1500pub[browse1500pub$treegroup=="raw", ]

names(browseroad1500)
head(broad1500house500)

save(browseroad1500, file = "browseroad1500.rda")

```



## For browse_9

##### Explore NAs 
```{r}

nrow(browse_9)
nrow(na.omit(browse_9))

head(browse_9, n = 10)
tail(browse_9, n = 10)
# unique(unlist(lapply(browse_9, function (x) which(is.na(x)))))
# fdata1[102, ]

browse_9$distvei1[is.na(browse_9$distvei1)] <- 10000
# browse_9$distvei2[is.na(browse_9$distvei2)] <- 10000
# browse_9$distvei3[is.na(browse_9$distvei3)] <- 10000

browse_9 <- na.omit(browse_9)

save(browse_9, file = "browse_9.rda")
load("processdata/browse_9.rda")

```

##### Data structure -> factor levels
```{r}

ggplot(data = browse_9, aes(landsdel)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9, aes(treartgruppe)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9, aes(hohkatb)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9, aes(region)) +
  geom_bar()+
  theme_classic()

```

```{r}

browse_9$beitetrykkr <- as.integer(round(browse_9$beitetrykk))
browse_9$bt100 <- browse_9$beitetrykkr/100
```

```{r}
# browse_9$landsdel <- as.factor(browse_9$landsdel)
# browse_9$hohkatb <- as.factor(browse_9$hohkatb)
# browse_9$landsdel <- factor(browse_9$landsdel, levels = c("Sørlandet", "Østlandet-Øst", "Midt-Norge", "Nord-Norge", "Vestlandet", "Østlandet-vest"))
browse_9$skogkategori <- factor(browse_9$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse_9$kant <- factor(browse_9$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse_9$treartgruppe <- factor(browse_9$treartgruppe, levels = c("LAUV", "BJORK", "FURU", "ROS"))
# browse_9$hohkatb <- factor(browse_9$hohkatb, levels = c("100", "0", "200", "300", "400", "500", "600", "700", "800"))
browse_9$region <- factor(browse_9$region, levels = c("Agder", "Buskerud", "Hordaland", "Hedmark", "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))

```

```{r}
save(browse_9, file = "browse_9.rda")
load("processdata/browse_9.rda")
```


##### Subset
```{r}

browse_9$disthus_cut210 <- ifelse(browse_9$disthus >= 210, 210, browse_9$disthus)
browse_9$disthus_cut210 <- ifelse(browse_9$disthus >= 210, 210, browse_9$disthus)

browse_9_dist <- browse_9[browse_9$distvei1 > 200, ] # longest distance to any public road 1500m
head(browse_9_dist)
save(browse_9_dist, file = "browse_9_dist.rda")

browse_9_distbin <- browse_9[browse_9$disthus > 200, ] 
browse_9_distbeta <- browse_9[browse_9$disthus > 210, ] 
# head(browse_9_dist)
save(browse_9_distbin, file = "browse_9_distbin.rda")
save(browse_9_distbeta, file = "browse_9_distbeta.rda")

```





## For browse9bro

##### Explore NAs 
```{r}

nrow(browse9bro)
nrow(na.omit(browse9bro))

head(browse9bro, n = 10)
tail(browse9bro, n = 10)
# unique(unlist(lapply(browse9bro, function (x) which(is.na(x)))))
# fdata1[102, ]

# browse9bro$distvei1[is.na(browse9bro$distvei1)] <- 10000
# browse9bro$distvei2[is.na(browse9bro$distvei2)] <- 10000
# browse9bro$distvei3[is.na(browse9bro$distvei3)] <- 10000

browse9bro <- na.omit(browse9bro)

save(browse9bro, file = "browse9bro.rda")
load("browse9bro.rda")

```

##### Data structure -> factor levels
```{r}

ggplot(data = browse9bro, aes(landsdel)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(treartgruppe)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(hohkatb)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(region)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(vegetasjon)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro, aes(veitype1)) +
  geom_bar()+
  theme_classic()

```



```{r}
browse9bro$beitetrykkr <- round(browse9bro$beitetrykk)
```

```{r}
browse9bro$landsdel <- as.factor(browse9bro$landsdel)
browse9bro$hohkatb <- as.factor(browse9bro$hohkatb)
browse9bro$landsdel <- factor(browse9bro$landsdel, levels = c("Vestlandet", "Sørlandet", "Østlandet-Øst", "Midt-Norge", "Nord-Norge", "Østlandet-vest"))
browse9bro$skogkategori <- factor(browse9bro$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9bro$kant <- factor(browse9bro$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9bro$treartgruppe <- factor(browse9bro$treartgruppe, levels = c("LAUV", "BJORK", "FURU", "ROS"))
browse9bro$hohkatb <- factor(browse9bro$hohkatb, levels = c("100", "0", "200", "300", "400", "500", "600", "700", "800"))
browse9bro$region <- factor(browse9bro$region, levels = c("Agder", "Buskerud", "Hordaland", "Hedmark", "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))
browse9bro$vegetasjon <- as.factor(browse9bro$vegetasjon)
browse9bro$vegetasjon <- factor(browse9bro$vegetasjon, levels = c("Blaabaerskog", "Alm-lindeskog", "Baerlyngskog", "Berg-grus",  "Blokkebaerskog", "Flommarkskog", "Graaorskog", "Hagemarkskog", "Hogstaudeskog", "Kystlynghei", "Laagurtskog", "Lavskog", "Myr-myrskog", "Or-askeskog", "Smaabregneskog", "Storbregneskog", "Viersumpskog"))
browse9bro$veitype1 <- as.factor(browse9bro$veitype1)
browse9bro$veitype1 <- factor(browse9bro$veitype1, levels = c("fylkesveg", "europaogriksveg", "kommuneveg"))
```

```{r}
save(browse9bro, file = "browse9bro.rda")
```




# For browse9bro_dist

##### Data structure -> factor levels
```{r}

ggplot(data = browse9bro_dist, aes(landsdel)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(treartgruppe)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(hohkatb)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(region)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(vegetasjon)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(veitype1)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9bro_dist, aes(helling)) +
  geom_bar()+
  theme_classic()

```

```{r}
browse9bro_dist$landsdel <- as.factor(browse9bro_dist$landsdel)
browse9bro_dist$hohkatb <- as.factor(browse9bro_dist$hohkatb)
browse9bro_dist$landsdel <- factor(browse9bro_dist$landsdel, levels = c("Vestlandet", "Sørlandet", "Østlandet-Øst", "Midt-Norge", "Nord-Norge", "Østlandet-vest"))
browse9bro_dist$skogkategori <- factor(browse9bro_dist$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9bro_dist$kant <- factor(browse9bro_dist$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9bro_dist$treartgruppe <- factor(browse9bro_dist$treartgruppe, levels = c("LAUV", "BJORK", "FURU", "ROS"))
browse9bro_dist$hohkatb <- factor(browse9bro_dist$hohkatb, levels = c("100", "0", "200", "300", "400", "500", "600", "700", "800"))
browse9bro_dist$region <- factor(browse9bro_dist$region, levels = c("Agder", "Buskerud", "Hordaland", "Hedmark", "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))
browse9bro_dist$vegetasjon <- as.factor(browse9bro_dist$vegetasjon)
browse9bro_dist$vegetasjon <- factor(browse9bro_dist$vegetasjon, levels = c("Blaabaerskog", "Alm-lindeskog", "Baerlyngskog", "Berg-grus",  "Blokkebaerskog", "Flommarkskog", "Graaorskog", "Hagemarkskog", "Hogstaudeskog", "Kystlynghei", "Laagurtskog", "Lavskog", "Myr-myrskog", "Or-askeskog", "Smaabregneskog", "Storbregneskog", "Viersumpskog"))
browse9bro_dist$veitype1 <- as.factor(browse9bro_dist$veitype1)
browse9bro_dist$veitype1 <- factor(browse9bro_dist$veitype1, levels = c("fylkesveg", "europaogriksveg", "kommuneveg"))
browse9bro_dist$helling <- as.factor(browse9bro_dist$helling)
```

```{r}
save(browse9bro_dist, file = "browse9bro_dist.rda")
```


For binary logistic regression
```{r}
browse9bro_dist$nobro <- 100-browse9bro_dist$beitetrykkr
browse9bro_dist$nobro <- as.integer(browse9bro_dist$nobro)
browse9bro_dist$beitetrykkr <- as.integer(browse9bro_dist$beitetrykkr)
```


For beta regression
```{r}
browse9bro_dist$bt100 <- browse9bro_dist$beitetrykkr/100

browse9bro_disthus$bt100 <- browse9bro_disthus$beitetrykk/100



browse9bro_dist$bt100[browse9bro_dist$bt100 == 0.00] <- 0.01   # ---- use this to run a normal beta regression, which cannot handle values of 0.0

str(browse9bro_dist$bt100)


```


For binomial
```{r}
browse9bro_dist$btbinary <- browse9bro_dist$beitetrykkr

browse9bro_dist$btbinary[browse9bro_dist$btbinary > 0] <- 1

str(browse9bro_dist$btbinary)


```


## For browse_9_dist

##### Data structure -> factor levels
```{r}

ggplot(data = browse_9_dist, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9_dist, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9_dist, aes(treartgruppe)) +
  geom_bar()+
  theme_classic()


ggplot(data = browse_9_dist, aes(region)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse_9_dist, aes(KOMNR)) +
  geom_bar()+
  theme_classic()



```

```{r}
str(browse_9_dist$skogkategori)
str(browse_9_dist$kant)
str(browse_9_dist$treartgruppe)
str(browse_9_dist$region)

```

```{r}
browse_9_dist$skogkategori <- as.factor(browse_9_dist$skogkategori)
levels(browse_9_dist$skogkategori)
browse_9_dist$skogkategori <- factor(browse_9_dist$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
levels(browse_9_dist$skogkategori)

browse_9_dist$kant <- as.factor(browse_9_dist$kant)
levels(browse_9_dist$kant)
browse_9_dist$kant <- factor(browse_9_dist$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
levels(browse_9_dist$kant)

browse_9_dist$treartgruppe <- as.factor(browse_9_dist$treartgruppe)
levels(browse_9_dist$treartgruppe)
browse_9_dist$treartgruppe <- factor(browse_9_dist$treartgruppe, levels = c("LAUV", "FURU", "ROS")) 
levels(browse_9_dist$treartgruppe)

browse_9_dist$region <- as.factor(browse_9_dist$region)
levels(browse_9_dist$region)

browse_9_dist$KOMNR <- as.factor(browse_9_dist$KOMNR)
levels(browse_9_dist$KOMNR)

```


```{r}
save(browse_9_dist, file = "browse_9_dist.rda")
```








# For browse9.2 (for browsing pressure response)

##### Data structure -> factor levels

```{r}
load("processdata/browse9.1.rda")
load("processdata/browse9.2.rda")
```


```{r}

str(browse0.1)
head(browse0.1, n = 3)
names(browse0.1)

```

##### Subsetting on 9th and 10th cycle
```{r}
browse0.1_910 <- split(browse0.1, browse0.1$takst)

browse9.1 <- browse0.1_910[['9']]
# browse10.1 <- browse0.1_910[['10']]
summary(browse9.1)

save(browse9.1, file = "browse9.1.rda")
# save(browse_10, file = "browse_10.rda")


browse9.1 <- browse9.1[,c(1:5,7:9,15:24,34:36,40:45,51,59,62:67,74,76:79,88:90,99:105,112:115)]
browse9.3 <- browse9.1
browse9.3 <- subset(browse9.3, select = -c(9:27))
browse9.3 <- subset(browse9.3, select = -c(11:16, 18:21))
# browse9.2$x <- NULL
summary(browse9.2)

browse9.5 <- browse9.1
browse9.5 <- browse9.5[,c(1:3,5:8,19:24,36,41,43,48,50:53)]

```


##### Explore NAs 
```{r}

nrow(browse9.2)
nrow(na.omit(browse9.2))

nrow(browse9.3)
nrow(na.omit(browse9.3))

head(browse_9, n = 10)
tail(browse_9, n = 10)
# unique(unlist(lapply(browse_9, function (x) which(is.na(x)))))
# fdata1[102, ]

sum(is.na(browse9.1$beitetrykk9))

browse9.2$distvei1[is.na(browse9.2$distvei1)] <- 10000
browse9.2$distvei2[is.na(browse9.2$distvei2)] <- 10000
browse9.2$distvei3[is.na(browse9.2$distvei3)] <- 10000
browse9.2$disthus[is.na(browse9.2$disthus)] <- 10000


browse9.3$distvei1[is.na(browse9.3$distvei1)] <- 10000
browse9.3$distvei2[is.na(browse9.3$distvei2)] <- 10000
browse9.3$distvei3[is.na(browse9.3$distvei3)] <- 10000
browse9.3$disthus[is.na(browse9.3$disthus)] <- 10000

browse9.2 <- na.omit(browse9.2)
browse9.3 <- na.omit(browse9.3)

save(browse9.2, file = "browse9.2.rda")
load("processdata/browse9.2.rda")


```


```{r}
library(ggplot2)
# ggplot(data = browse9bro_dist, aes(landsdel)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.2, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.4, aes(treartgruppe9)) +
  geom_bar()+
  theme_classic()

# ggplot(data = browse9.2, aes(treartgruppe)) +
#   geom_bar()+
#   theme_classic()

# ggplot(data = browse9bro_dist, aes(hohkatb)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.2, aes(region)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(vegetasjon)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2_dist, aes(veitype2)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2_dist, aes(veitype3)) +
  geom_bar()+
  theme_classic()


```

```{r}

browse9.2$beitetrykkr <- as.integer(round(browse9.2$beitetrykk9))
browse9.2$bt100 <- browse9.2$beitetrykkr/100
```

```{r}
# browse9.2$landsdel <- as.factor(browse9.2$landsdel)
# browse9.2$hohkatb <- as.factor(browse9.2$hohkatb)
# browse9.2$landsdel <- factor(browse9.2$landsdel, levels = c("Sørlandet", "Østlandet-Øst", "Midt-Norge", "Nord-Norge", "Vestlandet", "Østlandet-vest"))
browse9.2$skogkategori <- factor(browse9.2$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9.2$kant <- factor(browse9.2$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9.2$treartgruppe <- factor(browse9.2$treartgruppe9, levels = c("LAUV", "FURU", "ROS"))
# browse9.2$hohkatb <- factor(browse9.2$hohkatb, levels = c("100", "0", "200", "300", "400", "500", "600", "700", "800"))
browse9.2$region <- factor(browse9.2$region, levels = c("Hedmark", "Agder", "Buskerud", "Hordaland",  "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))

browse9.2_dist$veitype2 <- as.character(browse9.2_dist$veitype2)
browse9.2_dist$veitype2[browse9.2_dist$veitype2 == ""] <- "europaogriksveg"
browse9.2_dist$veitype2 <- factor(browse9.2_dist$veitype2, levels = c("privatveg", "europaogriksveg", "fylkesveg", "kommuneveg"))


```

```{r}
save(browse9.2, file = "browse9.2.rda")
save(browse9.2_dist, file = "browse9.2_dist.rda")
load("processdata/browse9.2.rda")
```


##### Subset
```{r}

# subset on disthus and distvei2 > 3000
# browse9.3 <- browse9.2[ which(browse9.2$disthus < 3000 & browse9.2$distvei2 < 3000),]
# save(browse9.3, file = "browse9.3.rda")
# 
# browse_9$disthus_cut210 <- ifelse(browse_9$disthus >= 210, 210, browse_9$disthus)
# 
# browse_9_dist <- browse_9[browse_9$distvei1 > 200, ] # longest distance to any public road 1500m
# head(browse_9_dist)
# save(browse_9_dist, file = "browse_9_dist.rda")
# 
# browse_9_distbin <- browse_9[browse_9$disthus > 200, ] 
# browse_9_distbeta <- browse_9[browse_9$disthus > 210, ] 
# # head(browse_9_dist)
# save(browse_9_distbin, file = "browse_9_distbin.rda")
# save(browse_9_distbeta, file = "browse_9_distbeta.rda")



browse9.2$disthus_trunc <- ifelse(browse9.2$disthus >= 3000, 3000, browse9.2$disthus)
browse9.2$distvei2_trunc <- ifelse(browse9.2$distvei2 >= 2218, 2218, browse9.2$distvei2)
browse9.2$tretetthet9_trunc <- ifelse(browse9.2$tretetthet9 >= 3960, 3960, browse9.2$tretetthet9)


# browse9.2_dist$disthus_trunc <- ifelse(browse9.2_dist$disthus >= 3000, 3000, browse9.2_dist$disthus)
# browse9.2_dist$distvei2_trunc <- ifelse(browse9.2_dist$distvei2 >= 3000, 3000, browse9.2_dist$distvei2)
# browse9.2_dist$tretetthet9_trunc <- ifelse(browse9.2_dist$tretetthet9 >= 3960, 3960, browse9.2_dist$tretetthet9)

browse9.2_distnew <- browse9.2[ which(browse9.2$disthus > 200), ]
head(browse9.2_dist)



save(browse9.2_dist, file = "browse9.2_dist.rda")

```

##### threshold cuts disthus
```{r}
browse9.2$disthus_cut1000 <- ifelse(browse9.2$disthus >= 1000, 1000, browse9.2$disthus)
browse9.2$disthus_cut950 <- ifelse(browse9.2$disthus >= 950, 950, browse9.2$disthus)
browse9.2$disthus_cut900 <- ifelse(browse9.2$disthus >= 900, 900, browse9.2$disthus)
browse9.2$disthus_cut850 <- ifelse(browse9.2$disthus >= 850, 850, browse9.2$disthus)
browse9.2$disthus_cut800 <- ifelse(browse9.2$disthus >= 800, 800, browse9.2$disthus)
browse9.2$disthus_cut750 <- ifelse(browse9.2$disthus >= 750, 750, browse9.2$disthus)
browse9.2$disthus_cut700 <- ifelse(browse9.2$disthus >= 700, 700, browse9.2$disthus)
browse9.2$disthus_cut650 <- ifelse(browse9.2$disthus >= 650, 650, browse9.2$disthus)
browse9.2$disthus_cut600 <- ifelse(browse9.2$disthus >= 600, 600, browse9.2$disthus)
browse9.2$disthus_cut500 <- ifelse(browse9.2$disthus >= 500, 500, browse9.2$disthus)
browse9.2$disthus_cut400 <- ifelse(browse9.2$disthus >= 400, 400, browse9.2$disthus)
browse9.2$disthus_cut300 <- ifelse(browse9.2$disthus >= 300, 300, browse9.2$disthus)
browse9.2$disthus_cut200 <- ifelse(browse9.2$disthus >= 200, 200, browse9.2$disthus)
browse9.2$disthus_cut150 <- ifelse(browse9.2$disthus >= 150, 150, browse9.2$disthus)
browse9.2$disthus_cut175 <- ifelse(browse9.2$disthus >= 175, 175, browse9.2$disthus)
browse9.2$disthus_cut160 <- ifelse(browse9.2$disthus >= 160, 160, browse9.2$disthus)
browse9.2$disthus_cut190 <- ifelse(browse9.2$disthus >= 190, 190, browse9.2$disthus)
browse9.2$disthus_trunc_cut1000 <- ifelse(browse9.2$disthus_trunc >= 1000, 1000, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut900 <- ifelse(browse9.2$disthus_trunc >= 900, 900, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut800 <- ifelse(browse9.2$disthus_trunc >= 800, 800, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut700 <- ifelse(browse9.2$disthus_trunc >= 700, 700, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut600 <- ifelse(browse9.2$disthus_trunc >= 600, 600, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut500 <- ifelse(browse9.2$disthus_trunc >= 500, 500, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut400 <- ifelse(browse9.2$disthus_trunc >= 400, 400, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut300 <- ifelse(browse9.2$disthus_trunc >= 300, 300, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut200 <- ifelse(browse9.2$disthus_trunc >= 200, 200, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut150 <- ifelse(browse9.2$disthus_trunc >= 150, 150, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut175 <- ifelse(browse9.2$disthus_trunc >= 175, 175, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut160 <- ifelse(browse9.2$disthus_trunc >= 160, 160, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut190 <- ifelse(browse9.2$disthus_trunc >= 190, 190, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut100 <- ifelse(browse9.2$disthus_trunc >= 100, 100, browse9.2$disthus_trunc)



```



##### Scale variables
```{r}

browse9.2$HOH_scale = scale(browse9.2$HOH)
browse9.2$tretetthet9_scale = scale(browse9.2$tretetthet9)
browse9.2$helling_scale = scale(browse9.2$helling)
browse9.2$disthus_scale = scale(browse9.2$disthus)
browse9.2$distvei2_scale = scale(browse9.2$distvei2)
browse9.2$ac_scale = scale(browse9.2$ac)
browse9.2$disthus_cut1000_scale = scale(browse9.2$disthus_cut1000)
browse9.2$disthus_cut900_scale = scale(browse9.2$disthus_cut900)
browse9.2$disthus_cut800_scale = scale(browse9.2$disthus_cut800)
browse9.2$disthus_cut700_scale = scale(browse9.2$disthus_cut700)
browse9.2$disthus_cut600_scale = scale(browse9.2$disthus_cut600)
browse9.2$disthus_cut500_scale = scale(browse9.2$disthus_cut500)
browse9.2$disthus_cut400_scale = scale(browse9.2$disthus_cut400)
browse9.2$disthus_cut300_scale = scale(browse9.2$disthus_cut300)
browse9.2$disthus_cut200_scale = scale(browse9.2$disthus_cut200)
browse9.2$disthus_cut150_scale = scale(browse9.2$disthus_cut150)
browse9.2$disthus_cut175_scale = scale(browse9.2$disthus_cut175)
browse9.2$disthus_cut160_scale = scale(browse9.2$disthus_cut160)
browse9.2$disthus_cut190_scale = scale(browse9.2$disthus_cut190)
browse9.2$disthus_trunc_cut1000_scale = scale(browse9.2$disthus_trunc_cut1000)
browse9.2$disthus_trunc_cut900_scale = scale(browse9.2$disthus_trunc_cut900)
browse9.2$disthus_trunc_cut800_scale = scale(browse9.2$disthus_trunc_cut800)
browse9.2$disthus_trunc_cut700_scale = scale(browse9.2$disthus_trunc_cut700)
browse9.2$disthus_trunc_cut600_scale = scale(browse9.2$disthus_trunc_cut600)
browse9.2$disthus_trunc_cut500_scale = scale(browse9.2$disthus_trunc_cut500)
browse9.2$disthus_trunc_cut400_scale = scale(browse9.2$disthus_trunc_cut400)
browse9.2$disthus_trunc_cut300_scale = scale(browse9.2$disthus_trunc_cut300)
browse9.2$disthus_trunc_cut200_scale = scale(browse9.2$disthus_trunc_cut200)
browse9.2$disthus_trunc_cut150_scale = scale(browse9.2$disthus_trunc_cut150)
browse9.2$disthus_trunc_cut175_scale = scale(browse9.2$disthus_trunc_cut175)
browse9.2$disthus_trunc_cut160_scale = scale(browse9.2$disthus_trunc_cut160)
browse9.2$disthus_trunc_cut190_scale = scale(browse9.2$disthus_trunc_cut190)
browse9.2$tretetthet9_trunc_scale = scale(browse9.2$tretetthet9_trunc)
browse9.2$distvei2_trunc_scale = scale(browse9.2$distvei2_trunc)
browse9.2$disthus_trunc_scale = scale(browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut100_scale = scale(browse9.2$disthus_trunc_cut100)


browse9.2_dist$tretetthet9_trunc_scale = scale(browse9.2_dist$tretetthet9_trunc)
browse9.2_dist$distvei2_trunc_scale = scale(browse9.2_dist$distvei2_trunc)
browse9.2_dist$disthus_trunc_scale = scale(browse9.2_dist$disthus_trunc)
browse9.2_dist$helling_scale = scale(browse9.2_dist$helling)
browse9.2_dist$helling_scale <- browse9.2_dist$helling_scale
browse9.2_dist$helling_scale <- NULL

browse9.3$HOH_scale = scale(browse9.3$HOH)
browse9.3$tretetthet9_scale = scale(browse9.3$tretetthet9)
browse9.3$helling_scale = scale(browse9.3$helling)
browse9.3$disthus_scale = scale(browse9.3$disthus)
browse9.3$distvei2_scale = scale(browse9.3$distvei2)
browse9.3$ac_scale = scale(browse9.3$ac)
browse9.3$disthus_cut1000_scale = scale(browse9.3$disthus_cut1000)
browse9.3$disthus_cut950_scale = scale(browse9.3$disthus_cut950)
browse9.3$disthus_cut900_scale = scale(browse9.3$disthus_cut900)
browse9.3$disthus_cut850_scale = scale(browse9.3$disthus_cut850)
browse9.3$disthus_cut800_scale = scale(browse9.3$disthus_cut800)
browse9.3$disthus_cut750_scale = scale(browse9.3$disthus_cut750)
browse9.3$disthus_cut700_scale = scale(browse9.3$disthus_cut700)
browse9.3$disthus_cut650_scale = scale(browse9.3$disthus_cut650)

save(browse9.2, file = "browse9.2.rda")
save(browse9.2_distnew, file = "browse9.2_distnew.rda")

```






## For browse9.3 (Treedensity)
##### Explore NAs 
```{r}
# 
# nrow(browse9.2)
# nrow(na.omit(browse9.2))

nrow(browse9.3)
nrow(na.omit(browse9.3))


# sum(is.na(browse9.1$beitetrykk9))
# 
# browse9.2$distvei1[is.na(browse9.2$distvei1)] <- 10000
# browse9.2$distvei2[is.na(browse9.2$distvei2)] <- 10000
# browse9.2$distvei3[is.na(browse9.2$distvei3)] <- 10000
# browse9.2$disthus[is.na(browse9.2$disthus)] <- 10000


browse9.3$distvei1[is.na(browse9.3$distvei1)] <- 10000
browse9.3$distvei2[is.na(browse9.3$distvei2)] <- 10000
browse9.3$distvei3[is.na(browse9.3$distvei3)] <- 10000
browse9.3$disthus[is.na(browse9.3$disthus)] <- 10000

# browse9.2 <- na.omit(browse9.2)
browse9.3 <- na.omit(browse9.3)
# 
# save(browse9.2, file = "browse9.2.rda")
# load("processdata/browse9.2.rda")

save(browse9.3, file = "browse9.3.rda")
load("processdata/browse9.3.rda")


```

##### Frequency plots categorical variables
```{r}
library(ggplot2)
# ggplot(data = browse9bro_dist, aes(landsdel)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.3, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(treartgruppe9)) +
  geom_bar()+
  theme_classic()

# ggplot(data = browse9.2, aes(treartgruppe)) +
#   geom_bar()+
#   theme_classic()

# ggplot(data = browse9bro_dist, aes(hohkatb)) +
#   geom_bar()+
#   theme_classic()

# ggplot(data = browse9.2, aes(region)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.3, aes(vegetasjon)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(veitype2)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(veitype3)) +
  geom_bar()+
  theme_classic()


```

##### Data structure

```{r}

browse9.3$skogkategori <- factor(browse9.3$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9.3$kant <- factor(browse9.3$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9.3$treartgruppe <- factor(browse9.3$treartgruppe9, levels = c("LAUV", "FURU", "ROS"))

browse9.3$vegetasjon <- factor(browse9.3$vegetasjon, levels = c("Blaabaerskog", "", "Alm-lindeskog", "Baerlyngskog", "Berg-grus", "Blokkebaerskog", "Flommarkskog", "Graaorskog", "Hagemarkskog", "Hogstaudeskog", "Kystlynghei", "Laagurtskog", "Lavskog", "Myr-myrskog", "Or-askeskog", "Smaabregneskog", "Storbregneskog", "Viersumpskog"))

browse9.3$veitype2 <- as.character(browse9.3$veitype2)
browse9.3$veitype2[browse9.3$veitype2 == ""] <- "europaogriksveg"
browse9.3$veitype2 <- factor(browse9.3$veitype2, levels = c("privatveg", "europaogriksveg", "fylkesveg", "kommuneveg"))

browse9.3$vegetasjon <- as.character(browse9.3$vegetasjon)
browse9.3$vegetasjon[browse9.3$vegetasjon == ""] <- "Blokkebaerskog"
browse9.3$vegetasjon <- factor(browse9.3$vegetasjon, levels = c("Blaabaerskog", "Alm-lindeskog", "Baerlyngskog", "Berg-grus", "Blokkebaerskog", "Flommarkskog", "Graaorskog", "Hagemarkskog", "Hogstaudeskog", "Kystlynghei", "Laagurtskog", "Lavskog", "Myr-myrskog", "Or-askeskog", "Smaabregneskog", "Storbregneskog", "Viersumpskog"))

browse9.3$beitetrykkr <- as.integer(round(browse9.3$beitetrykk9))

browse9.3$disthus_trunc_scale = scale(browse9.3$disthus_trunc)
browse9.3$distvei2_trunc_scale = scale(browse9.3$distvei2_trunc)
browse9.3$ac_scale = scale(browse9.3$ac)

browse9.3$disthus_trunc_cut200 <- ifelse(browse9.3$disthus_trunc >= 200, 200, browse9.3$disthus_trunc)
browse9.3$disthus_trunc_cut100 <- ifelse(browse9.3$disthus_trunc >= 100, 100, browse9.3$disthus_trunc)
browse9.3$disthus_trunc_cut300 <- ifelse(browse9.3$disthus_trunc >= 300, 300, browse9.3$disthus_trunc)
browse9.3$disthus_trunc_cut250 <- ifelse(browse9.3$disthus_trunc >= 250, 250, browse9.3$disthus_trunc)
browse9.3$disthus_trunc_cut200_scale = scale(browse9.3$disthus_trunc_cut200)
browse9.3$disthus_trunc_cut100_scale = scale(browse9.3$disthus_trunc_cut100)
browse9.3$disthus_trunc_cut300_scale = scale(browse9.3$disthus_trunc_cut300)
browse9.3$disthus_trunc_cut250_scale = scale(browse9.3$disthus_trunc_cut250)


```

```{r}
save(browse9.3, file = "browse9.3.rda")

```


##### Subset and censoring
```{r}


browse9.3$disthus_trunc <- ifelse(browse9.3$disthus >= 3000, 3000, browse9.3$disthus)
browse9.3$distvei2_trunc <- ifelse(browse9.3$distvei2 >= 2218, 2218, browse9.3$distvei2)
browse9.3$tretetthet9_trunc <- ifelse(browse9.3$tretetthet9 >= 3960, 3960, browse9.3$tretetthet9)

browse9.3_dist <- browse9.3[ which(browse9.3$disthus > 250), ]

save(browse9.3, file = "browse9.3.rda")
save(browse9.3_dist, file = "browse9.3_dist.rda")

```








## For browse9.5 (tree height classes, not tree density)

```{r}
load("browse9.5.rda")
```

##### Subsetting on 9th and 10th cycle
```{r}
browse0.1_910 <- split(browse0.1, browse0.1$takst)

browse9.1 <- browse0.1_910[['9']]
summary(browse9.1)

# save(browse9.1, file = "browse9.1.rda")
load("processdata/browse9.1.rda")


browse9.1 <- browse9.1[,c(1:5,7:9,15:24,34:36,40:45,51,59,62:67,74,76:79,88:90,99:105,112:115)]

# browse9.2$x <- NULL


browse9.5 <- browse9.1
browse9.5 <- browse9.5[,c(1:3,5:8,19:24,36,41,43,48,50:54)]
browse9.5 <- subset(browse9.5, select = -c(27:32,34:36))


save(browse9.5, file = "browse9.5.rda")

```

##### Explore NAs 
```{r}

nrow(browse9.5)
nrow(na.omit(browse9.5))


browse9.5$distvei2[is.na(browse9.5$distvei2)] <- 2218
browse9.5$disthus[is.na(browse9.5$disthus)] <- 3000

browse9.5 <- na.omit(browse9.5)


# save(browse9.5, file = "browse9.5.rda")
load("processdata/browse9.5.rda")


```


```{r}
library(ggplot2)

ggplot(data = browse9.5, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5m, aes(treartgruppe9)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5, aes(region)) +
  geom_bar()+
  theme_classic()


```


```{r}

browse9.5$skogkategori <- factor(browse9.5$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9.5$kant <- factor(browse9.5$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9.5m$treartgruppe9 <- factor(browse9.5m$treartgruppe9, levels = c("LAUV", "FURU"))

browse9.5$region <- factor(browse9.5$region, levels = c("Hedmark", "Agder", "Buskerud", "Hordaland",  "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))



```

```{r}

browse9.5$beitetrykkr <- as.integer(round(browse9.5$beitetrykk9))
browse9.5$bt100 <- browse9.5$beitetrykkr/100
```

```{r}
save(browse9.5, file = "browse9.5.rda")

save(browse9.6, file = "browse9.6.rda")
```


##### Subsets and new variable(s)
```{r}

# Creating new variable proportion large trees "proplt" 
load("processdata/browse9.5.rda")
browse9.5 <- as.data.frame(browse9.5)

tmp1 <- browse9.5[browse9.5$treartgruppe9 == "FURU", ]
tmp2 <- browse9.5[browse9.5$treartgruppe9 == "LAUV", ]
tmp3 <- browse9.5[browse9.5$treartgruppe9 == "ROS", ]
browse9.5m <-  browse9.5[browse9.5$treartgruppe9 != "FURU", ]
browse9.5m <-  browse9.5m[!duplicated(browse9.5m$FLATEID), ]
browse9.5m$treartgruppe9 <-  "LAUV"
browse9.5m <- rbind(browse9.5m, tmp1)
browse9.5m$rosTD <- tmp3$tretetthet9[match(browse9.5m$FLATEID, tmp3$FLATEID)]
browse9.5m$rosBP <- tmp3$beitetrykkr [match(browse9.5m$FLATEID, tmp3$FLATEID)]
browse9.5m$lauvTD <- tmp2$tretetthet9[match(browse9.5m$FLATEID, tmp2$FLATEID)]
browse9.5m$lauvBP <- tmp2$beitetrykkr [match(browse9.5m$FLATEID, tmp2$FLATEID)]

browse9.5m$rosTD <- ifelse(is.na(browse9.5m$rosTD), 0, browse9.5m$rosTD)
browse9.5m$rosBP <- ifelse(is.na(browse9.5m$rosBP), 0, browse9.5m$rosBP)
browse9.5m$lauvTD <- ifelse(is.na(browse9.5m$lauvTD), 0, browse9.5m$lauvTD)
browse9.5m$lauvBP <- ifelse(is.na(browse9.5m$lauvBP), 0, browse9.5m$lauvBP)
  
browse9.5m$browspres_merged <- browse9.5m$beitetrykkr
browse9.5m$browspres_merged[browse9.5m$treartgruppe9 == "LAUV"] <- (browse9.5m$lauvTD[browse9.5m$treartgruppe9 == "LAUV"] * browse9.5m$lauvBP[browse9.5m$treartgruppe9 == "LAUV"] + browse9.5m$rosTD[browse9.5m$treartgruppe9 == "LAUV"] * browse9.5m$rosBP[browse9.5m$treartgruppe9 == "LAUV"])/ (browse9.5m$lauvTD[browse9.5m$treartgruppe9 == "LAUV"] + browse9.5m$rosTD[browse9.5m$treartgruppe9 == "LAUV"])

nrow(browse9.5m)
nrow(na.omit(browse9.5m))

browse9.5m <- as.data.frame(browse9.5m)

browse9.5m <- browse9.5m[,c(1:24,29)]


browse9.5m$proplt <- ifelse(browse9.5m$treartgruppe9 == "FURU", (browse9.5m$Furu_0_25_PRHA + browse9.5m$Furu_25_49_PRHA)/(browse9.5m$Furu_0_PRHA + browse9.5m$Furu_0_25_PRHA + browse9.5m$Furu_25_49_PRHA), (browse9.5m$Lauv_0_25_PRHA + browse9.5m$Lauv_25_49_PRHA)/ (browse9.5m$Lauv_0_PRHA + browse9.5m$Lauv_0_25_PRHA + browse9.5m$Lauv_25_49_PRHA))


nrow(browse9.5m)
nrow(na.omit(browse9.5m))

browse9.5m <- na.omit(browse9.5m)


browse9.5msc <- browse9.5msc[,c(1,3,7,15:17,19:20,25:26,33,40)]
save(browse9.5msc, file = "browse9.5msc.rda")



# Subset on different cuts for disthus
browse9.6 <- browse9.5m[ which(browse9.5m$disthus > 400), ]
save(browse9.6, file = "browse9.6.rda")

browse9.7pos <- browse9.6pos[ which(browse9.6pos$disthus > 200), ]
save(browse9.7pos, file = "browse9.7pos.rda")




save(browse9.5, file = "browse9.5.rda")

save(browse9.5m, file = "browse9.5m.rda")

```




## browse9.5m

##### threshold cuts disthus
```{r}
browse9.5m$disthus_1000 <- ifelse(browse9.5m$disthus >= 1000, 1000, browse9.5m$disthus)
# browse9.5m$disthus_cut950 <- ifelse(browse9.5m$disthus >= 950, 950, browse9.5m$disthus)
browse9.5m$disthus_900 <- ifelse(browse9.5m$disthus >= 900, 900, browse9.5m$disthus)
# browse9.5m$disthus_cut850 <- ifelse(browse9.5m$disthus >= 850, 850, browse9.5m$disthus)
browse9.5m$disthus_800 <- ifelse(browse9.5m$disthus >= 800, 800, browse9.5m$disthus)
# browse9.5m$disthus_cut750 <- ifelse(browse9.5m$disthus >= 750, 750, browse9.5m$disthus)
browse9.5m$disthus_700 <- ifelse(browse9.5m$disthus >= 700, 700, browse9.5m$disthus)
# browse9.5m$disthus_cut650 <- ifelse(browse9.5m$disthus >= 650, 650, browse9.5m$disthus)
browse9.5m$disthus_600 <- ifelse(browse9.5m$disthus >= 600, 600, browse9.5m$disthus)
browse9.5m$disthus_500 <- ifelse(browse9.5m$disthus >= 500, 500, browse9.5m$disthus)
browse9.5m$disthus_400 <- ifelse(browse9.5m$disthus >= 400, 400, browse9.5m$disthus)
browse9.5m$disthus_350 <- ifelse(browse9.5m$disthus >= 350, 350, browse9.5m$disthus)
browse9.5m$disthus_300 <- ifelse(browse9.5m$disthus >= 300, 300, browse9.5m$disthus)
browse9.5m$disthus_200 <- ifelse(browse9.5m$disthus >= 200, 200, browse9.5m$disthus)
# browse9.5m$disthus_150 <- ifelse(browse9.5m$disthus >= 150, 150, browse9.5m$disthus)
browse9.5m$disthus_175 <- ifelse(browse9.5m$disthus >= 175, 175, browse9.5m$disthus)
browse9.5m$disthus_160 <- ifelse(browse9.5m$disthus >= 160, 160, browse9.5m$disthus)
# browse9.5m$disthus_190 <- ifelse(browse9.5m$disthus >= 190, 190, browse9.5m$disthus)
# browse9.2$disthus_trunc_cut1000 <- ifelse(browse9.2$disthus_trunc >= 1000, 1000, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut900 <- ifelse(browse9.2$disthus_trunc >= 900, 900, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut800 <- ifelse(browse9.2$disthus_trunc >= 800, 800, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut700 <- ifelse(browse9.2$disthus_trunc >= 700, 700, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut600 <- ifelse(browse9.2$disthus_trunc >= 600, 600, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut500 <- ifelse(browse9.2$disthus_trunc >= 500, 500, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut400 <- ifelse(browse9.2$disthus_trunc >= 400, 400, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut300 <- ifelse(browse9.2$disthus_trunc >= 300, 300, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut200 <- ifelse(browse9.2$disthus_trunc >= 200, 200, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut150 <- ifelse(browse9.2$disthus_trunc >= 150, 150, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut175 <- ifelse(browse9.2$disthus_trunc >= 175, 175, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut160 <- ifelse(browse9.2$disthus_trunc >= 160, 160, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut190 <- ifelse(browse9.2$disthus_trunc >= 190, 190, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut100 <- ifelse(browse9.2$disthus_trunc >= 100, 100, browse9.2$disthus_trunc)

save(browse9.5m, file = "browse9.5m.rda")

```
