==> 01_cleaning_data_Browsing_pressure.Rmd <==
---
title: "01_cleaning_browse_data_BP"
author: "anne"
date: "15 januar 2020"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
# Old analysis
## For browse9.2 (for browsing pressure response)

##### Data structure -> factor levels

```{r}
load("rawdata/browse0.1.rda")
load("processdata/browse9.1.rda")
load("processdata/browse9.2.rda")
```


```{r}

str(browse0.1)
head(browse0.1, n = 3)
names(browse0.1)

```

##### Subsetting on 9th and 10th cycle
```{r}
browse0.1_910 <- split(browse0.1, browse0.1$takst)

browse9.1 <- browse0.1_910[['9']]
# browse10.1 <- browse0.1_910[['10']]
summary(browse9.1)

save(browse9.1, file = "browse9.1.rda")
# save(browse_10, file = "browse_10.rda")
load("processdata/browse9.1.rda")

browse9.1 <- browse9.1[,c(1:5,7:9,15:24,34:36,40:45,51,59,62:67,74,76:79,88:90,99:105,112:115)]
browse9.2 <- browse9.1
# browse9.3 <- subset(browse9.3, select = -c(9:27))
# browse9.3 <- subset(browse9.3, select = -c(11:16, 18:21))
# browse9.2$x <- NULL
summary(browse9.2)

browse9.5 <- browse9.1
browse9.5 <- browse9.5[,c(1:3,5:8,19:24,36,41,43,48,50:53)]
str(browse9.5)

```


##### Explore NAs 
```{r}

nrow(browse9.2)
nrow(na.omit(browse9.2))

# nrow(browse9.3)
# nrow(na.omit(browse9.3))

# head(browse_9, n = 10)
# tail(browse_9, n = 10)
# sum(unique(unlist(lapply(browse9.3, function (x) which(is.na(x))))))
# fdata1[102, ]

# browse9.3 <- browse9.2[,c(1:3,5,7:9,88:90,102:105,112:115,126:131)]
# 
# sum(is.na(browse9.3$beitetrykk9))
# sum(is.na(browse9.3$tretetthet9))

browse9.2$distvei1[is.na(browse9.2$distvei1)] <- 10000
browse9.2$distvei2[is.na(browse9.2$distvei2)] <- 10000
browse9.2$distvei3[is.na(browse9.2$distvei3)] <- 10000
browse9.2$disthus[is.na(browse9.2$disthus)] <- 10000


# browse9.3$distvei1[is.na(browse9.3$distvei1)] <- 10000
# browse9.3$distvei2[is.na(browse9.3$distvei2)] <- 10000
# browse9.3$distvei3[is.na(browse9.3$distvei3)] <- 10000
# browse9.3$disthus[is.na(browse9.3$disthus)] <- 10000

browse9.2 <- na.omit(browse9.2)
# browse9.3 <- na.omit(browse9.3)
# 
# save(browse9.2, file = "browse9.2.rda")
load("processdata/browse9.2.rda")


```


```{r}
library(ggplot2)
# ggplot(data = browse9bro_dist, aes(landsdel)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.2, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.4, aes(treartgruppe9)) +
  geom_bar()+
  theme_classic()

# ggplot(data = browse9.2, aes(treartgruppe)) +
#   geom_bar()+
#   theme_classic()

# ggplot(data = browse9bro_dist, aes(hohkatb)) +
#   geom_bar()+
#   theme_classic()

ggplot(data = browse9.2, aes(region)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.3, aes(vegetasjon)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2_dist, aes(veitype2)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.2_dist, aes(veitype3)) +
  geom_bar()+
  theme_classic()


```

```{r}

browse9.2$beitetrykkr <- as.integer(round(browse9.2$beitetrykk9))
browse9.2$bt100 <- browse9.2$beitetrykkr/100
```

```{r}
# browse9.2$landsdel <- as.factor(browse9.2$landsdel)
# browse9.2$hohkatb <- as.factor(browse9.2$hohkatb)
# browse9.2$landsdel <- factor(browse9.2$landsdel, levels = c("Sørlandet", "Østlandet-Øst", "Midt-Norge", "Nord-Norge", "Vestlandet", "Østlandet-vest"))
browse9.2$skogkategori <- factor(browse9.2$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9.2$kant <- factor(browse9.2$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9.2$treartgruppe <- factor(browse9.2$treartgruppe9, levels = c("LAUV", "FURU", "ROS"))
# browse9.2$hohkatb <- factor(browse9.2$hohkatb, levels = c("100", "0", "200", "300", "400", "500", "600", "700", "800"))
browse9.2$region <- factor(browse9.2$region, levels = c("Hedmark", "Agder", "Buskerud", "Hordaland",  "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))

browse9.2_dist$veitype2 <- as.character(browse9.2_dist$veitype2)
browse9.2_dist$veitype2[browse9.2_dist$veitype2 == ""] <- "europaogriksveg"
browse9.2_dist$veitype2 <- factor(browse9.2_dist$veitype2, levels = c("privatveg", "europaogriksveg", "fylkesveg", "kommuneveg"))


```

```{r}
save(browse9.2, file = "browse9.2.rda")
save(browse9.2_dist, file = "browse9.2_dist.rda")
load("processdata/browse9.2.rda")
```


##### Subset
```{r}

# subset on disthus and distvei2 > 3000
# browse9.3 <- browse9.2[ which(browse9.2$disthus < 3000 & browse9.2$distvei2 < 3000),]
# save(browse9.3, file = "browse9.3.rda")
# 
# browse_9$disthus_cut210 <- ifelse(browse_9$disthus >= 210, 210, browse_9$disthus)
# 
# browse_9_dist <- browse_9[browse_9$distvei1 > 200, ] # longest distance to any public road 1500m
# head(browse_9_dist)
# save(browse_9_dist, file = "browse_9_dist.rda")
# 
# browse_9_distbin <- browse_9[browse_9$disthus > 200, ] 
# browse_9_distbeta <- browse_9[browse_9$disthus > 210, ] 
# # head(browse_9_dist)
# save(browse_9_distbin, file = "browse_9_distbin.rda")
# save(browse_9_distbeta, file = "browse_9_distbeta.rda")



browse9.2$disthus_trunc <- ifelse(browse9.2$disthus >= 3000, 3000, browse9.2$disthus)
browse9.2$distvei2_trunc <- ifelse(browse9.2$distvei2 >= 2218, 2218, browse9.2$distvei2)
browse9.2$tretetthet9_trunc <- ifelse(browse9.2$tretetthet9 >= 3960, 3960, browse9.2$tretetthet9)


# browse9.2_dist$disthus_trunc <- ifelse(browse9.2_dist$disthus >= 3000, 3000, browse9.2_dist$disthus)
# browse9.2_dist$distvei2_trunc <- ifelse(browse9.2_dist$distvei2 >= 3000, 3000, browse9.2_dist$distvei2)
# browse9.2_dist$tretetthet9_trunc <- ifelse(browse9.2_dist$tretetthet9 >= 3960, 3960, browse9.2_dist$tretetthet9)

browse9.2_distnew <- browse9.2[ which(browse9.2$disthus > 200), ]
head(browse9.2_dist)



save(browse9.2_dist, file = "browse9.2_dist.rda")

```

##### threshold cuts disthus
```{r}
browse9.2$disthus_cut1000 <- ifelse(browse9.2$disthus >= 1000, 1000, browse9.2$disthus)
browse9.2$disthus_cut950 <- ifelse(browse9.2$disthus >= 950, 950, browse9.2$disthus)
browse9.2$disthus_cut900 <- ifelse(browse9.2$disthus >= 900, 900, browse9.2$disthus)
browse9.2$disthus_cut850 <- ifelse(browse9.2$disthus >= 850, 850, browse9.2$disthus)
browse9.2$disthus_cut800 <- ifelse(browse9.2$disthus >= 800, 800, browse9.2$disthus)
browse9.2$disthus_cut750 <- ifelse(browse9.2$disthus >= 750, 750, browse9.2$disthus)
browse9.2$disthus_cut700 <- ifelse(browse9.2$disthus >= 700, 700, browse9.2$disthus)
browse9.2$disthus_cut650 <- ifelse(browse9.2$disthus >= 650, 650, browse9.2$disthus)
browse9.2$disthus_cut600 <- ifelse(browse9.2$disthus >= 600, 600, browse9.2$disthus)
browse9.2$disthus_cut500 <- ifelse(browse9.2$disthus >= 500, 500, browse9.2$disthus)
browse9.2$disthus_cut400 <- ifelse(browse9.2$disthus >= 400, 400, browse9.2$disthus)
browse9.2$disthus_cut300 <- ifelse(browse9.2$disthus >= 300, 300, browse9.2$disthus)
browse9.2$disthus_cut200 <- ifelse(browse9.2$disthus >= 200, 200, browse9.2$disthus)
browse9.2$disthus_cut150 <- ifelse(browse9.2$disthus >= 150, 150, browse9.2$disthus)
browse9.2$disthus_cut175 <- ifelse(browse9.2$disthus >= 175, 175, browse9.2$disthus)
browse9.2$disthus_cut160 <- ifelse(browse9.2$disthus >= 160, 160, browse9.2$disthus)
browse9.2$disthus_cut190 <- ifelse(browse9.2$disthus >= 190, 190, browse9.2$disthus)
browse9.2$disthus_trunc_cut1000 <- ifelse(browse9.2$disthus_trunc >= 1000, 1000, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut900 <- ifelse(browse9.2$disthus_trunc >= 900, 900, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut800 <- ifelse(browse9.2$disthus_trunc >= 800, 800, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut700 <- ifelse(browse9.2$disthus_trunc >= 700, 700, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut600 <- ifelse(browse9.2$disthus_trunc >= 600, 600, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut500 <- ifelse(browse9.2$disthus_trunc >= 500, 500, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut400 <- ifelse(browse9.2$disthus_trunc >= 400, 400, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut300 <- ifelse(browse9.2$disthus_trunc >= 300, 300, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut200 <- ifelse(browse9.2$disthus_trunc >= 200, 200, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut150 <- ifelse(browse9.2$disthus_trunc >= 150, 150, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut175 <- ifelse(browse9.2$disthus_trunc >= 175, 175, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut160 <- ifelse(browse9.2$disthus_trunc >= 160, 160, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut190 <- ifelse(browse9.2$disthus_trunc >= 190, 190, browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut100 <- ifelse(browse9.2$disthus_trunc >= 100, 100, browse9.2$disthus_trunc)



```



##### Scale variables
```{r}

browse9.2$HOH_scale = scale(browse9.2$HOH)
browse9.2$tretetthet9_scale = scale(browse9.2$tretetthet9)
browse9.2$helling_scale = scale(browse9.2$helling)
browse9.2$disthus_scale = scale(browse9.2$disthus)
browse9.2$distvei2_scale = scale(browse9.2$distvei2)
browse9.2$ac_scale = scale(browse9.2$ac)
browse9.2$disthus_cut1000_scale = scale(browse9.2$disthus_cut1000)
browse9.2$disthus_cut900_scale = scale(browse9.2$disthus_cut900)
browse9.2$disthus_cut800_scale = scale(browse9.2$disthus_cut800)
browse9.2$disthus_cut700_scale = scale(browse9.2$disthus_cut700)
browse9.2$disthus_cut600_scale = scale(browse9.2$disthus_cut600)
browse9.2$disthus_cut500_scale = scale(browse9.2$disthus_cut500)
browse9.2$disthus_cut400_scale = scale(browse9.2$disthus_cut400)
browse9.2$disthus_cut300_scale = scale(browse9.2$disthus_cut300)
browse9.2$disthus_cut200_scale = scale(browse9.2$disthus_cut200)
browse9.2$disthus_cut150_scale = scale(browse9.2$disthus_cut150)
browse9.2$disthus_cut175_scale = scale(browse9.2$disthus_cut175)
browse9.2$disthus_cut160_scale = scale(browse9.2$disthus_cut160)
browse9.2$disthus_cut190_scale = scale(browse9.2$disthus_cut190)
browse9.2$disthus_trunc_cut1000_scale = scale(browse9.2$disthus_trunc_cut1000)
browse9.2$disthus_trunc_cut900_scale = scale(browse9.2$disthus_trunc_cut900)
browse9.2$disthus_trunc_cut800_scale = scale(browse9.2$disthus_trunc_cut800)
browse9.2$disthus_trunc_cut700_scale = scale(browse9.2$disthus_trunc_cut700)
browse9.2$disthus_trunc_cut600_scale = scale(browse9.2$disthus_trunc_cut600)
browse9.2$disthus_trunc_cut500_scale = scale(browse9.2$disthus_trunc_cut500)
browse9.2$disthus_trunc_cut400_scale = scale(browse9.2$disthus_trunc_cut400)
browse9.2$disthus_trunc_cut300_scale = scale(browse9.2$disthus_trunc_cut300)
browse9.2$disthus_trunc_cut200_scale = scale(browse9.2$disthus_trunc_cut200)
browse9.2$disthus_trunc_cut150_scale = scale(browse9.2$disthus_trunc_cut150)
browse9.2$disthus_trunc_cut175_scale = scale(browse9.2$disthus_trunc_cut175)
browse9.2$disthus_trunc_cut160_scale = scale(browse9.2$disthus_trunc_cut160)
browse9.2$disthus_trunc_cut190_scale = scale(browse9.2$disthus_trunc_cut190)
browse9.2$tretetthet9_trunc_scale = scale(browse9.2$tretetthet9_trunc)
browse9.2$distvei2_trunc_scale = scale(browse9.2$distvei2_trunc)
browse9.2$disthus_trunc_scale = scale(browse9.2$disthus_trunc)
browse9.2$disthus_trunc_cut100_scale = scale(browse9.2$disthus_trunc_cut100)


browse9.2_dist$tretetthet9_trunc_scale = scale(browse9.2_dist$tretetthet9_trunc)
browse9.2_dist$distvei2_trunc_scale = scale(browse9.2_dist$distvei2_trunc)
browse9.2_dist$disthus_trunc_scale = scale(browse9.2_dist$disthus_trunc)
browse9.2_dist$helling_scale = scale(browse9.2_dist$helling)
browse9.2_dist$helling_scale <- browse9.2_dist$helling_scale
browse9.2_dist$helling_scale <- NULL

browse9.3$HOH_scale = scale(browse9.3$HOH)
browse9.3$tretetthet9_scale = scale(browse9.3$tretetthet9)
browse9.3$helling_scale = scale(browse9.3$helling)
browse9.3$disthus_scale = scale(browse9.3$disthus)
browse9.3$distvei2_scale = scale(browse9.3$distvei2)
browse9.3$ac_scale = scale(browse9.3$ac)
browse9.3$disthus_cut1000_scale = scale(browse9.3$disthus_cut1000)
browse9.3$disthus_cut950_scale = scale(browse9.3$disthus_cut950)
browse9.3$disthus_cut900_scale = scale(browse9.3$disthus_cut900)
browse9.3$disthus_cut850_scale = scale(browse9.3$disthus_cut850)
browse9.3$disthus_cut800_scale = scale(browse9.3$disthus_cut800)
browse9.3$disthus_cut750_scale = scale(browse9.3$disthus_cut750)
browse9.3$disthus_cut700_scale = scale(browse9.3$disthus_cut700)
browse9.3$disthus_cut650_scale = scale(browse9.3$disthus_cut650)

save(browse9.2, file = "browse9.2.rda")
save(browse9.2_distnew, file = "browse9.2_distnew.rda")

```


## New dataset with only variables we use in the models
```{r}
browse9.4_house <- browse9.2[,c(8,10,12,17,19:21,23:25,35,49,92)]
browse9.4_house[,"disthus"] <- NULL
browse9.4_house[,"HOH"] <- NULL
browse9.4_house[,"tretetthet9"] <- NULL
browse9.4_house$tretetthet9_trunc <- browse9.2$tretetthet9_trunc
browse9.4_house$KOMNR <- browse9.2$KOMNR
browse9.4_house$treartgruppe9 <- factor(browse9.4_house$treartgruppe9, levels = c("LAUV", "FURU", "ROS"))



save(browse9.4_house, file = "browse9.4_house.rda")
browse9.4 <- browse9.2_dist[,c(3,10,12,17,19:23,25,35,36,49,69,92,97)]
browse9.4 <- subset(browse9.4, select = -c(4:5,9,12))
browse9.4$tretetthet9_trunc <- browse9.2_dist$tretetthet9_trunc
browse9.4[,"beitetrykk9"] <- NULL
browse9.4$beitetrykkr <- browse9.2_dist$beitetrykkr
browse9.4[,"tretetthet9_trunc_scale"] <- NULL

browse9.4$treartgruppe9 <- factor(browse9.4$treartgruppe9, levels = c("LAUV", "FURU", "ROS"))

save(browse9.4, file = "browse9.4.rda")




```


## Add moose density (mean over 5 year period to datasets and models)
```{r}
library(readr)
moose_mean_density <- read_csv("processdata_copy/five_y_mean_density.csv", na = "NA")
str(moose_mean_density)
library(tidyverse)
moose_mean_density <- moose_mean_density %>%
  rename(KOMNR = knr2017) %>%
  rename(moose_density = five_y_mean_density)
str(moose_mean_density)

```

#### Browsing pressure house
```{r}
## HOUSE
load("processdata_copy/browse9.4_house.rda")
browse9.4_house_test <- left_join(browse9.4_house, moose_mean_density, by = "KOMNR")
str(browse9.4_house_test$moose_density)
sum(is.na(browse9.4_house_test$moose_density))
which(is.na(browse9.4_house_test$moose_density))
browse9.4_house_test[90, ]
browse9.4_house_test[170, ]
browse9.4_house_test$moose_density <- ifelse(is.na(browse9.4_house_test$moose_density), 0, browse9.4_house_test$moose_density)
sum(is.na(browse9.4_house_test$moose_density))

browse9.4_house_test_pos <- browse9.4_house_test[browse9.4_house_test$bt100 > 0, ]



library(lme4)
library(glmmTMB)

## BIN
load("output_copy/m1kbin2.rda")
bh_bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + scale(tretetthet9_trunc) + scale(moose_density) + (1|KOMNR),
            family = binomial(),
            data = browse9.4_house_test)

bh_bin_int <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + scale(tretetthet9_trunc) + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
            family = binomial(),
            data = browse9.4_house_test)

library(AICcmodavg)
candmods_BP_hbin <- list(bh_bin, m1kbin2, bh_bin_int)
aictab(candmods_BP_hbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)
#       K     AICc Delta_AICc AICcWt Cum.Wt        LL
# Mod1 19 21504.26       0.00   0.58   0.58 -10733.11
# Mod3 20 21504.93       0.67   0.42   1.00 -10732.45
# Mod2 18 21522.13      17.87   0.00   1.00 -10743.05

summary(bh_bin) # moose density positive ***


# library(sjPlot)
# library(ggplot2)
# p1 <- plot_model(bh_bin, type = c("eff"), transform = NULL, terms = c("moose_density"),
#            grid = TRUE)
# p1 <- p1 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("5 year mean moose density\n") +
#   ylab("Browsing pressure (%)\n")


## BETA
load("output_copy/m1ibeta2.rda")
bh_beta <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + (1|KOMNR),
            data = browse9.4_house_test_pos,
            family = beta_family())

bh_beta_int <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
            data = browse9.4_house_test_pos,
            family = beta_family())

library(bbmle)
AIC(bh_beta, m1ibeta2, bh_beta_int)
ICtab(bh_beta, m1ibeta2, bh_beta_int, type = "AIC", weights = TRUE, delta = TRUE)

summary(bh_beta) # moose density not sign.

```


#### Browsing presure roads
```{r}
## ROADS
load("processdata_copy/browse9.4.rda")
browse9.4_test <- left_join(browse9.4, moose_mean_density, by = "KOMNR")
str(browse9.4_test$moose_density)
sum(is.na(browse9.4_test$moose_density))
which(is.na(browse9.4_test$moose_density))
browse9.4_test[78, ]
browse9.4_test[175, ]
browse9.4_test$moose_density <- ifelse(is.na(browse9.4_test$moose_density), 0, browse9.4_test$moose_density)
sum(is.na(browse9.4_test$moose_density))

browse9.4_test_pos <- browse9.4_test[browse9.4_test$bt100 > 0, ]


library(lme4)
library(glmmTMB)


## BIN
load("output_copy/m4abin2.rda")
br_bin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + (1|KOMNR),
            family = binomial(),
            data = browse9.4_test)

br_bin_int <- glmer(factor(beitetrykkr > 0) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + moose_density*log(distvei2_trunc+1) + (1|KOMNR),
            family = binomial(),
            data = browse9.4_test)

library(AICcmodavg)
candmods_BP_hbeta <- list(br_bin, m4abin2, br_bin_int)
aictab(candmods_BP_hbeta, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)
#       K     AICc Delta_AICc AICcWt Cum.Wt       LL
# Mod1 18 19057.81       0.00   0.73   0.73 -9510.89
# Mod3 19 19059.76       1.94   0.27   1.00 -9510.86
# Mod2 17 19077.52      19.71   0.00   1.00 -9521.75

summary(br_bin) # moose density positiv ***


## BETA
load("output_copy/m4abeta2.rda")
br_beta <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + (1|KOMNR),
            data = browse9.4_test_pos,
            family = beta_family())

br_beta_int <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + moose_density + moose_density*log(distvei2_trunc+1) + (1|KOMNR),
            data = browse9.4_test_pos,
            family = beta_family())

library(bbmle)
AIC(br_beta, m4abeta2, br_beta_int)
ICtab(br_beta, m4abeta2, br_beta_int, type = "AIC", weights = TRUE, delta = TRUE)

summary(br_beta) # moose density not sign.

```




#### run the final models with final datset (only variables that are in the model)
```{r}

## house
m1kbin2 <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + (1|KOMNR),
            family = binomial(),
            data = browse9.4_house)
save(m1kbin2, file = "m1kbin2.rda")
load("output/m1kbin2.rda")


m1ibeta2 <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + (1|KOMNR),
            data = browse9.4_house_pos,
            family = beta_family())
save(m1ibeta2, file = "m1ibeta2.rda")
load("output/m1ibeta2.rda")



## Roads
m4abin2 <- glmer(factor(beitetrykkr > 0) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + (1|KOMNR),
            family = binomial(),
            data = browse9.4)
load("output/m4abin2.rda")


m4abeta2 <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe9 + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc + (1|KOMNR),
            data = browse9.4pos,
            family = beta_family())
load("output/m4abeta2.rda")

```





# ############################################
#         NEW data set                     ##
#############################################



## Prepare data set for browsing pressure
```{r}

load("rawdata/browse0.1.rda")
names(browse0.1)


browse_full <- browse0.1[, c(1,3,5,7:9,62:67,74,88,90,103,105,112:115)]
str(browse_full)
names(browse_full)


browse_910 <- split(browse_full, browse_full$takst)

browse_9 <- browse_910[['9']]
# summary(browse_9)
names(browse_9)


library(dplyr)
browse <- distinct(browse_9, FLATEID, .keep_all = TRUE)



head(browse)


```




## Make a new recruits dataset

#### Duplicate rows and make column treartgruppe9
```{r}
# browse is made in: ## Prepare data set for browsing pressure

str(browse)
table(as.factor(browse$treartgruppe9))

browse <- browse %>% slice(rep(1:n(), each = 3))
head(browse)

x <- factor(c("ROS","FURU","LAUV"))
browse$treartgruppe9 <- rep(x, 14726) # nrows browse/3 = 14726, also just nrows recruits before replication
head(browse$treartgruppe9)
head(browse)

```


#### Make column tretetthet9
```{r}
browse$tretetthet9 <- ifelse(browse$treartgruppe9 == "ROS", browse$elgROSha9, NA)
browse$tretetthet9 <- ifelse(browse$treartgruppe9 == "FURU", browse$elgFURUha9, browse$tretetthet9)
browse$tretetthet9 <- ifelse(browse$treartgruppe9 == "LAUV", browse$elgLAUVha9, browse$tretetthet9)

head(browse$tretetthet9, n = 12)
sum(is.na(browse$tretetthet9)) #68, meaning that there were no trees to count? But then it should have been 0, no?


```


#### Make column beitetrykk9
```{r}
browse$beitetrykk9 <- ifelse(browse$treartgruppe9 == "ROS", browse$elgbROSprs9, NA)
browse$beitetrykk9 <- ifelse(browse$treartgruppe9 == "FURU", browse$elgbFURUprs9, browse$beitetrykk9)
browse$beitetrykk9 <- ifelse(browse$treartgruppe9 == "LAUV", browse$elgbLAUVprs9, browse$beitetrykk9)

head(browse$beitetrykk9, n = 12)
sum(is.na(browse$beitetrykk9)) #21725, meaning that there could not have been browsing pressure, because there were no trees to browse on 


## NA the ones where tretetthet is 0, but we have beitetrykk values
browse$beitetrykk9 <- ifelse(browse$tretetthet9 != 0, browse$beitetrykk9, NA)
sum(is.na(browse$beitetrykk9)) #21834, meaning that there could not have been browsing pressure, because there were no trees to browse on  

browse$beitetrykk9 <- as.integer(round(browse$beitetrykk9))
browse$bt100 <- browse$beitetrykk9/100

```


#### Make new variable moose density (mean over 5 year period)
```{r}
library(readr)
moose_mean_density <- read_csv("processdata/five_y_mean_density_2021_02_01.csv", na = "NA")
str(moose_mean_density)
library(tidyverse)
moose_mean_density <- moose_mean_density %>%
  rename(KOMNR = knr2017) %>%
  rename(moose_density = five_y_mean_density)
str(moose_mean_density)

browse <- left_join(browse, moose_mean_density, by = "KOMNR")
str(browse$moose_density)
sum(is.na(browse$moose_density)) # 6807 --> Either "elgareal" was 0 or NA
which(is.na(browse$moose_density))
browse$moose_density <- ifelse(is.na(browse$moose_density), 0, browse$moose_density)
sum(is.na(browse$moose_density))

```

#### Reduce data set to variables we are going to use further
```{r}
names(browse)
browse <- browse[,-c(7:12)] 
names(browse)
```



#### Stuff like... NAs, level order, etc.
```{r}

nrow(browse)
nrow(na.omit(browse))

sum(is.na(browse$distvei2))
sum(is.na(browse$disthus))
sum(is.na(browse$HOH))
sum(is.na(browse$region))
sum(is.na(browse$skogkategori))
sum(is.na(browse$kant))
sum(is.na(browse$helling)) #6867, what does it mean when there is NA in helling? helling from 1-100, so I guess NA is when there is no helling, so 0?
sum(is.na(browse$treartgruppe9))
sum(is.na(browse$beitetrykk9)) # 21834, because there are 0 trees (in the 9th cycle), hence no BP --> still, there are browse in the 10th/11th... 
sum(is.na(browse$tretetthet9)) #68, meaning no trees to count? Should these not just be 0 then? Or delete?
sum(is.na(browse$recruitment))
sum(is.na(browse$recruitment_ratio))


browse$distvei2[is.na(browse$distvei2)] <- 2218 # max distance to roads
browse$disthus[is.na(browse$disthus)] <- 3000# max distance to house
browse$helling[is.na(browse$helling)] <- 0 # no helling
browse$tretetthet9 <- ifelse(browse$tretetthet9 >= 3960, 3960, browse$tretetthet9)

nrow(browse) 
nrow(na.omit(browse)) # only NAs left are the ones in beitetrykk and tretetthet
browse <- na.omit(browse)


library(ggplot2)
ggplot(data = browse, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse, aes(kant)) +
  geom_bar()+
  theme_classic()

browse$skogkategori <- factor(browse$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse$kant <- factor(browse$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 



```



## Modelling spatial autocorrelation (https://onlinelibrary.wiley.com/doi/full/10.1111/j.2007.0906-7590.05171.x)
Change the projection and make coordinates and a spatial point data frame to make a shapefile

```{r}
library(tidyr)
browse$UTMmerge <- paste(browse$UTM_OV_33, browse$UTM_SN_33, sep="_")
sum(duplicated(browse$UTMmerge))


```


```{r}

library(sp)
library(rgdal)

crs <- CRS("+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs")
coords <- cbind(browse$UTM_OV_33, browse$UTM_SN_33)
data_spdf <- SpatialPointsDataFrame(coords, browse, proj4string = crs)
# plot(data_spdf)
writeOGR(data_spdf, dsn = "output", layer = "spac", driver = "ESRI Shapefile")

```


```{r}

library(spdep)
library(ape)


# spdep package:

# prepare neighbour lists for spatial autocorrelation analysis
nb.list <- dnearneigh(as.matrix(browse[browse$beitetrykk9, c("UTM_OV_33", "UTM_SN_33")]), 0, 5000)
nb.weights <- nb2listw(nb.list, zero.policy = TRUE)
# moran.test(browse$beitetrykk9, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)


# Make a matrix of coordinates - use when chunk above was not run
coords <- as.matrix(cbind(browse$UTM_OV_33, browse$UTM_SN_33))

# compute the autocovariate based on the above distance and weight
ac <- autocov_dist(browse$beitetrykk9, coords, nbs = 5000, type="inverse", zero.policy = TRUE)
browse$ac <- ac



```


#### Threshold distance to house
```{r}
browse$disthus_600 <- ifelse(browse$disthus >= 600, 600, browse$disthus)
browse$disthus_500 <- ifelse(browse$disthus >= 500, 500, browse$disthus)
browse$disthus_400 <- ifelse(browse$disthus >= 400, 400, browse$disthus)
browse$disthus_350 <- ifelse(browse$disthus >= 350, 350, browse$disthus)
browse$disthus_300 <- ifelse(browse$disthus >= 300, 300, browse$disthus)
browse$disthus_200 <- ifelse(browse$disthus >= 200, 200, browse$disthus)
browse$disthus_150 <- ifelse(browse$disthus >= 150, 150, browse$disthus)

```



#### Reduce data set to variables used further
```{r}
names(browse)
browse <- browse[,-c(3:5,18)]
names(browse)

save(browse, file = "browse_2021_02_01.rda")


```



#### Threshold distance to road
```{r}
load("processdata/browse_2021_02_01.rda")
names(browse)

browse$distvei_600 <- ifelse(browse$distvei2 >= 600, 600, browse$distvei2)
browse$distvei_500 <- ifelse(browse$distvei2 >= 500, 500, browse$distvei2)
browse$distvei_400 <- ifelse(browse$distvei2 >= 400, 400, browse$distvei2)
browse$distvei_350 <- ifelse(browse$distvei2 >= 350, 350, browse$distvei2)
browse$distvei_300 <- ifelse(browse$distvei2 >= 300, 300, browse$distvei2)
browse$distvei_200 <- ifelse(browse$distvei2 >= 200, 200, browse$distvei2)
browse$distvei_150 <- ifelse(browse$distvei2 >= 150, 150, browse$distvei2)
browse$distvei_100 <- ifelse(browse$distvei2 >= 100, 100, browse$distvei2)
browse$distvei_50 <- ifelse(browse$distvei2 >= 50, 50, browse$distvei2)


sum(is.na(browse)) # 0

save(browse, file = "processdata/browse_2021_04_15.rda")

```


<!-- ## Roads data set -->
<!-- ```{r} -->
<!-- load("processdata/browse_2021_02_01.rda") -->
<!-- names(browse) -->
<!-- browse_roads <- browse[,-c(4,16:20,22)] -->
<!-- names(browse_roads) -->
<!-- browse_roads <- browse_roads[ which(browse_roads$disthus > 200), ] -->
<!-- summary(browse_roads$disthus) -->

<!-- str(browse_roads) -->


<!-- save(browse_roads, file = "browse_roads_2021_02_02.rda") -->

<!-- ``` -->
