==> 01_cleaning_data_Recruits.Rmd <==
---
title: "01_cleaning_data_browse_vers2"
author: "anne"
date: "20 desember 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Old analysis
## For browse9.5 (tree height classes, not tree density)

```{r}
load("processdata/browse9.5.rda")
load("rawdata/browse0.1.rda")
```

##### Subsetting on 9th and 10th cycle
```{r}
browse0.1_910 <- split(browse0.1, browse0.1$takst)

browse9.1 <- browse0.1_910[['9']]
summary(browse9.1)

# save(browse9.1, file = "browse9.1.rda")
load("processdata/browse9.1.rda")
load("processdata/browse9.5.rda")

browse9.1 <- browse9.1[,c(1:5,7:9,15:24,34:36,40:45,51,59,62:67,74,76:79,88:90,99:105,112:115)]

# browse9.2$x <- NULL


browse9.5 <- browse9.1
browse9.5 <- browse9.5[,c(1:3,5:8,19:24,36,41:43,48,50:54)]
browse9.5 <- subset(browse9.5, select = -c(27:32,34:36))


save(browse9.5, file = "browse9.5.rda")

```

##### Explore NAs 
```{r}

nrow(browse9.5)
nrow(na.omit(browse9.5))

sum(is.na(browse9.5$beitetrykk9))
sum(is.na(browse9.5$tretetthet9))

summary(browse9.5)

browse9.5$distvei2[is.na(browse9.5$distvei2)] <- 2218
browse9.5$disthus[is.na(browse9.5$disthus)] <- 3000

browse9.5 <- na.omit(browse9.5)


# save(browse9.5, file = "browse9.5.rda")
load("processdata/browse9.5.rda")
load("processdata/browse9.5msc.rda")


# browse9.5$FLATEID <- as.factor(browse9.5$FLATEID)
# str(levels(browse9.5$FLATEID)) 
# unique(browse9.5$FLATEID)
# 
browse9.5m$FLATEID <- as.factor(browse9.5m$FLATEID)
str(levels(browse9.5m$FLATEID))
unique(browse9.5m$FLATEID)

```


```{r}
library(ggplot2)

ggplot(data = browse9.5, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5, aes(kant)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5m, aes(treartgruppe9)) +
  geom_bar()+
  theme_classic()

ggplot(data = browse9.5, aes(region)) +
  geom_bar()+
  theme_classic()


```


```{r}

browse9.5$skogkategori <- factor(browse9.5$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
browse9.5$kant <- factor(browse9.5$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 
browse9.5$treartgruppe9 <- factor(browse9.5$treartgruppe9, levels = c("LAUV", "FURU"))

browse9.5$region <- factor(browse9.5$region, levels = c("Hedmark", "Agder", "Buskerud", "Hordaland",  "More og Romsdal", "Nord-Trondelag", "Nordland", "Oppland", "Ostfold-Akershus", "Rogaland", "Sogn og Fjordane", "Sor-Trondelag", "Troms", "Vestfold-Telemark"))



```

```{r}

browse9.5$beitetrykkr <- as.integer(round(browse9.5$beitetrykk9))
browse9.5$bt100 <- browse9.5$beitetrykkr/100
```

```{r}
save(browse9.5, file = "browse9.5.rda")

save(browse9.6, file = "browse9.6.rda")
```


##### Subsets and new variable(s)
```{r}

## Creating new variable "browsed merged"
load("processdata/browse9.5.rda")
browse9.5 <- as.data.frame(browse9.5)

tmp1 <- browse9.5[browse9.5$treartgruppe9 == "FURU", ]
tmp2 <- browse9.5[browse9.5$treartgruppe9 == "LAUV", ]
tmp3 <- browse9.5[browse9.5$treartgruppe9 == "ROS", ]
browse9.5m <-  browse9.5[browse9.5$treartgruppe9 != "FURU", ]
browse9.5m <-  browse9.5m[!duplicated(browse9.5m$FLATEID), ]
browse9.5m$treartgruppe9 <-  "LAUV"
browse9.5m <- rbind(browse9.5m, tmp1)
browse9.5m$rosTD <- tmp3$tretetthet9[match(browse9.5m$FLATEID, tmp3$FLATEID)]
browse9.5m$rosBP <- tmp3$beitetrykkr [match(browse9.5m$FLATEID, tmp3$FLATEID)]
browse9.5m$lauvTD <- tmp2$tretetthet9[match(browse9.5m$FLATEID, tmp2$FLATEID)]
browse9.5m$lauvBP <- tmp2$beitetrykkr [match(browse9.5m$FLATEID, tmp2$FLATEID)]

browse9.5m$rosTD <- ifelse(is.na(browse9.5m$rosTD), 0, browse9.5m$rosTD)
browse9.5m$rosBP <- ifelse(is.na(browse9.5m$rosBP), 0, browse9.5m$rosBP)
browse9.5m$lauvTD <- ifelse(is.na(browse9.5m$lauvTD), 0, browse9.5m$lauvTD)
browse9.5m$lauvBP <- ifelse(is.na(browse9.5m$lauvBP), 0, browse9.5m$lauvBP)
  
browse9.5m$browspres_merged <- browse9.5m$beitetrykkr
browse9.5m$browspres_merged[browse9.5m$treartgruppe9 == "LAUV"] <- (browse9.5m$lauvTD[browse9.5m$treartgruppe9 == "LAUV"] * browse9.5m$lauvBP[browse9.5m$treartgruppe9 == "LAUV"] + browse9.5m$rosTD[browse9.5m$treartgruppe9 == "LAUV"] * browse9.5m$rosBP[browse9.5m$treartgruppe9 == "LAUV"])/ (browse9.5m$lauvTD[browse9.5m$treartgruppe9 == "LAUV"] + browse9.5m$rosTD[browse9.5m$treartgruppe9 == "LAUV"])

browse9.5m$distvei2[is.na(browse9.5m$distvei2)] <- 2218
browse9.5m$disthus[is.na(browse9.5m$disthus)] <- 3000



nrow(browse9.5m)
nrow(na.omit(browse9.5m))

summary(browse9.5m)

sum(is.na(browse9.5m$beitetrykkr))


browse9.5m <- as.data.frame(browse9.5m)

browse9.5m <- browse9.5m[,c(1:24,29)]

## Creating new variable proportion large trees "proplt"
browse9.5m$proplt <- ifelse(browse9.5m$treartgruppe9 == "FURU", 
                            (browse9.5m$Furu_0_25_PRHA + browse9.5m$Furu_25_49_PRHA)/(browse9.5m$Furu_0_PRHA + browse9.5m$Furu_0_25_PRHA + browse9.5m$Furu_25_49_PRHA), 
                            (browse9.5m$Lauv_0_25_PRHA + browse9.5m$Lauv_25_49_PRHA)/ (browse9.5m$Lauv_0_PRHA + browse9.5m$Lauv_0_25_PRHA + browse9.5m$Lauv_25_49_PRHA))


nrow(browse9.5m)
nrow(na.omit(browse9.5m))

browse9.5m <- na.omit(browse9.5m)


browse9.5msc <- browse9.5msc[,c(1,3,7,15:17,19:20,25:26,33,40)]
save(browse9.5msc, file = "browse9.5msc.rda")


# Subset on different cuts for disthus
browse9.6 <- browse9.5m[ which(browse9.5m$disthus > 400), ]
save(browse9.6, file = "browse9.6.rda")

# browse9.7pos <- browse9.6pos[ which(browse9.6pos$disthus > 200), ]
# save(browse9.7pos, file = "browse9.7pos.rda")



save(browse9.5, file = "browse9.5.rda")

save(browse9.5m, file = "browse9.5m.rda")



```

## Test for difference between regions
```{r}
##### House
load("processdata/browse9.5msc.rda")
browse9.5m_region <- browse9.5m
browse9.5m_region <- browse9.5m_region[,c(1:3,7,14:20,25:26,33,40)]

pvars <- c("disthus_400", "HOH")
browse9.5m_regionsc <- browse9.5m_region
browse9.5m_regionsc[pvars] <- lapply(browse9.5m_regionsc[pvars],scale)

save(browse9.5m_regionsc, file = "browse9.5m_regionsc.rda")

## Subset municipalities ----
hedmark <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Hedmark", ]
agder <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Agder", ]
buskerud <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Buskerud", ]
hordaland <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Hordaland", ]
more_og_romsdal <- browse9.5m_regionsc[browse9.5m_regionsc$region == "More og Romsdal", ]
nord_trondelag <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Nord-Trondelag", ]
nordland <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Nordland", ]
oppland <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Oppland", ]
ostfold_akershus <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Ostfold-Akershus", ]
rogaland <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Rogaland", ]
sogn_og_fjordane <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Sogn og Fjordane", ]
sor_trondelag <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Sor-Trondelag", ]
troms <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Troms", ]
vestfold_telemark <- browse9.5m_regionsc[browse9.5m_regionsc$region == "Vestfold-Telemark", ]



##### Road
load("processdata/browse9.6.rda")
browse9.6_region <- browse9.6
browse9.6_region <- browse9.6_region[,c(1:3,7,14:20,25:26,33,40)]

pvars <- c("HOH","ac" ) #  "helling", "browspres_merged", "ac" 
browse9.6_regionsc <- browse9.6_region
browse9.6_regionsc[pvars] <- lapply(browse9.6_regionsc[pvars],scale)

save(browse9.6_regionsc, file = "browse9.6_regionsc.rda")
```


## browse9.5m

##### threshold cuts disthus
```{r}
browse9.5m$disthus_1000 <- ifelse(browse9.5m$disthus >= 1000, 1000, browse9.5m$disthus)
# browse9.5m$disthus_cut950 <- ifelse(browse9.5m$disthus >= 950, 950, browse9.5m$disthus)
browse9.5m$disthus_900 <- ifelse(browse9.5m$disthus >= 900, 900, browse9.5m$disthus)
# browse9.5m$disthus_cut850 <- ifelse(browse9.5m$disthus >= 850, 850, browse9.5m$disthus)
browse9.5m$disthus_800 <- ifelse(browse9.5m$disthus >= 800, 800, browse9.5m$disthus)
# browse9.5m$disthus_cut750 <- ifelse(browse9.5m$disthus >= 750, 750, browse9.5m$disthus)
browse9.5m$disthus_700 <- ifelse(browse9.5m$disthus >= 700, 700, browse9.5m$disthus)
# browse9.5m$disthus_cut650 <- ifelse(browse9.5m$disthus >= 650, 650, browse9.5m$disthus)
browse9.5m$disthus_600 <- ifelse(browse9.5m$disthus >= 600, 600, browse9.5m$disthus)
browse9.5m$disthus_500 <- ifelse(browse9.5m$disthus >= 500, 500, browse9.5m$disthus)
browse9.5m$disthus_400 <- ifelse(browse9.5m$disthus >= 400, 400, browse9.5m$disthus)
browse9.5m$disthus_350 <- ifelse(browse9.5m$disthus >= 350, 350, browse9.5m$disthus)
browse9.5m$disthus_300 <- ifelse(browse9.5m$disthus >= 300, 300, browse9.5m$disthus)
browse9.5m$disthus_200 <- ifelse(browse9.5m$disthus >= 200, 200, browse9.5m$disthus)
# browse9.5m$disthus_150 <- ifelse(browse9.5m$disthus >= 150, 150, browse9.5m$disthus)
browse9.5m$disthus_175 <- ifelse(browse9.5m$disthus >= 175, 175, browse9.5m$disthus)
browse9.5m$disthus_160 <- ifelse(browse9.5m$disthus >= 160, 160, browse9.5m$disthus)
# browse9.5m$disthus_190 <- ifelse(browse9.5m$disthus >= 190, 190, browse9.5m$disthus)
# browse9.2$disthus_trunc_cut1000 <- ifelse(browse9.2$disthus_trunc >= 1000, 1000, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut900 <- ifelse(browse9.2$disthus_trunc >= 900, 900, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut800 <- ifelse(browse9.2$disthus_trunc >= 800, 800, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut700 <- ifelse(browse9.2$disthus_trunc >= 700, 700, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut600 <- ifelse(browse9.2$disthus_trunc >= 600, 600, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut500 <- ifelse(browse9.2$disthus_trunc >= 500, 500, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut400 <- ifelse(browse9.2$disthus_trunc >= 400, 400, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut300 <- ifelse(browse9.2$disthus_trunc >= 300, 300, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut200 <- ifelse(browse9.2$disthus_trunc >= 200, 200, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut150 <- ifelse(browse9.2$disthus_trunc >= 150, 150, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut175 <- ifelse(browse9.2$disthus_trunc >= 175, 175, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut160 <- ifelse(browse9.2$disthus_trunc >= 160, 160, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut190 <- ifelse(browse9.2$disthus_trunc >= 190, 190, browse9.2$disthus_trunc)
# browse9.2$disthus_trunc_cut100 <- ifelse(browse9.2$disthus_trunc >= 100, 100, browse9.2$disthus_trunc)

save(browse9.5m, file = "browse9.5m.rda")
```





## Add moose density (mean over 5 year period to datasets and models)
```{r}
library(readr)
moose_mean_density <- read_csv("processdata_copy/five_y_mean_density.csv", na = "NA")
str(moose_mean_density)
library(tidyverse)
moose_mean_density <- moose_mean_density %>%
  rename(KOMNR = knr2017) %>%
  rename(moose_density = five_y_mean_density)
str(moose_mean_density)

```

#### Recruits house
```{r}
## HOUSE
load("processdata_copy/browse9.5msc.rda")
browse9.5msc_test <- left_join(browse9.5msc, moose_mean_density, by = "KOMNR")
str(browse9.5msc_test$moose_density)
sum(is.na(browse9.5msc_test$moose_density))
which(is.na(browse9.5msc_test$moose_density))
browse9.5msc_test[2254, ]
browse9.5msc_test[2257, ]
browse9.5msc_test$moose_density <- ifelse(is.na(browse9.5msc_test$moose_density), 0, browse9.5msc_test$moose_density)
sum(is.na(browse9.5msc_test$moose_density))

# browse9.5msc_test_pos <- browse9.5msc_test[browse9.5msc_test$bt100 > 0, ]
# sum(is.na(browse9.5msc_test_pos))

load("processdata_copy/browse9.5mpossc.rda")
browse9.5msc_test_pos <- left_join(browse9.5mpossc, moose_mean_density, by = "KOMNR")
str(browse9.5msc_test_pos$moose_density)
sum(is.na(browse9.5msc_test_pos$moose_density))
browse9.5msc_test_pos$moose_density <- ifelse(is.na(browse9.5msc_test_pos$moose_density), 0, browse9.5msc_test_pos$moose_density)
sum(is.na(browse9.5msc_test_pos$moose_density))

```


#### Recruits roads
```{r}
## ROADS
load("processdata_copy/browse9.6sc.rda")
browse9.6sc_test <- left_join(browse9.6sc, moose_mean_density, by = "KOMNR")
str(browse9.6sc_test$moose_density)
sum(is.na(browse9.6sc_test$moose_density))
which(is.na(browse9.6sc_test$moose_density))
browse9.6sc_test[1497, ]
browse9.6sc_test[5522, ]
browse9.6sc_test$moose_density <- ifelse(is.na(browse9.6sc_test$moose_density), 0, browse9.6sc_test$moose_density)
sum(is.na(browse9.6sc_test$moose_density))

load("processdata_copy/browse9.6scpos.rda")
browse9.6sc_test_pos <- left_join(browse9.6scpos, moose_mean_density, by = "KOMNR")
str(browse9.6sc_test_pos$moose_density)
sum(is.na(browse9.6sc_test_pos$moose_density))
browse9.6sc_test_pos$moose_density <- ifelse(is.na(browse9.6sc_test_pos$moose_density), 0, browse9.6sc_test_pos$moose_density)



```







# ##############################################
# #         NEW data set                     ##
#############################################


## Look at changes in proportion of trees from 9th to 10th takst (plant performance)

```{r}
load("rawdata/browse0.1.rda")
names(browse0.1)

library(readr)
new_trees <- read_csv("rawdata/new_trees.csv", na = "NA")
str(new_trees)
new_trees_treatment <- read_csv("processdata/new_trees_forest_treatment.csv", na = "NA")
str(new_trees)

new_trees <- inner_join(new_trees, new_trees_treatment, by = "FLATEID")
str(new_trees)


nrow(new_trees)
nrow(na.omit(new_trees)) #9 difference
new_trees <- na.omit(new_trees)

head(new_trees)

# library(tidyverse)
# moose_mean_density <- moose_mean_density %>%
#   rename(KOMNR = knr2017) %>%
#   rename(moose_density = five_y_mean_density)
# str(moose_mean_density)


library(dplyr)


```

#### Prepare data set for plant performance test
```{r}
recruits_full <- browse0.1[, c(1,3,5,7:9,62:67,74,88,90,103,105,112:115)]
str(recruits_full)
names(recruits_full)


# #### TRY THIS BEFORE SPLITTING
# recruits_full$Lauv1 <- ifelse(recruits_full$takst == 10, recruits_full$Lauv_0_PRHA + recruits_full$Bjork_0_PRHA, NA)
# recruits_full$Lauv2 <- ifelse(recruits_full$takst == 10, recruits_full$Lauv_0_25_PRHA + recruits_full$Bjork_0_25_PRHA, NA)
# recruits_full$Lauv3 <- ifelse(recruits_full$takst == 10, recruits_full$Lauv_25_49_PRHA + recruits_full$Bjork_25_49_PRHA, NA)
# 
# recruits <- subset(recruits_full, treartgruppe10!="BJORK")


recruits_910 <- split(recruits_full, recruits_full$takst)

recruits_9 <- recruits_910[['9']]
# summary(recruits_9)
names(recruits_9)

# recruits_9 <- recruits_9[ ,-c(7:12,16:18,22:24,32:39,48:52)]
# names(recruits_9)



# recruits_10 <- recruits_910[['10']]
# # summary(recruits_10)
# names(recruits_10)
# 
# recruits_10 <- recruits_10[ ,-c(2,4:11, 13:14,16:51)]
# names(recruits_10)





## inner_join works, when I do not have the dublicated rows
recruits9 <- distinct(recruits_9, FLATEID, .keep_all = TRUE)
# recruits10 <- distinct(recruits_10, FLATEID, .keep_all = TRUE)
recruits <- inner_join(recruits9, new_trees, by = "FLATEID", suffix =
c("_9", "_new"))


names(recruits)


```

#### --> OLD! With trees below 5 cm DBH (same as in paper) OLD!
```{r}
# ### Number of trees 2(10)/ number of trees 1(9)
# 
# # 9th and 10th
# recruits_test$proplt3 <- (recruits_test$Furu_0_25_PRHA.y + recruits_test$Lauv2.y)/(recruits_test$Furu_0_PRHA.x + recruits_test$Lauv_0_PRHA.x)
# str(recruits_test$proplt2)
# 
# # 9th
# recruits_test$proplt1 <- (recruits_test$Furu_0_25_PRHA.x + recruits_test$Lauv_0_25_PRHA.x)/(recruits_test$Furu_0_PRHA.x + recruits_test$Lauv_0_PRHA.x)
# str(recruits_test$proplt1)
# 
# recruits_test$proplt2 <- (recruits_test$Furu_25_49_PRHA.x + recruits_test$Lauv_25_49_PRHA.x)/(recruits_test$Furu_0_25_PRHA.x + recruits_test$Lauv_0_25_PRHA.x)
# str(recruits_test$proplt2)
# 
# 
# 
# ### Number of trees 3(10)/ number of trees 1(9)
# 
# # 9th and 10th
# recruits_test$proplt4 <- (recruits_test$Furu_25_49_PRHA.y + recruits_test$Lauv3.y)/(recruits_test$Furu_0_PRHA.x + recruits_test$Lauv_0_25_PRHA.x)
# str(recruits_test$proplt3)
# 
# names(recruits_test)
# recruits_cor <- recruits_test[, c(98,100)]
# recruits_cor1 <- recruits_test[, c(99,101)]
# 
# 
# recruits_cor <- recruits_cor[Reduce(`&`, lapply(recruits_cor, is.finite)),]
# recruits_cor1 <- recruits_cor1[Reduce(`&`, lapply(recruits_cor1, is.finite)),]
# 
# 
# cor(recruits_cor)
# #           proplt2    proplt   proplt3
# # proplt2 1.0000000 0.5951219 0.3916736
# # proplt  0.5951219 1.0000000 0.5284639
# # proplt3 0.3916736 0.5284639 1.0000000
# 
# cor(recruits_cor1)



```







## Make a new recruits dataset

#### Duplicate rows and make column treartgruppe9
```{r}
# recruits is made in: #### Prepare data set for plant performance test

str(recruits)
table(as.factor(recruits$treartgruppe9))

recruits <- recruits %>% slice(rep(1:n(), each = 3))
head(recruits)

x <- factor(c("ROS","FURU","LAUV"))
recruits$treartgruppe9 <- rep(x, 11596) #nrows recruits/3 = 11596, also just nrows recruits before replication
head(recruits$treartgruppe9)
head(recruits)

```


#### Make column tretetthet9
```{r}
recruits$tretetthet9 <- ifelse(recruits$treartgruppe9 == "ROS", recruits$elgROSha9, NA)
recruits$tretetthet9 <- ifelse(recruits$treartgruppe9 == "FURU", recruits$elgFURUha9, recruits$tretetthet9)
recruits$tretetthet9 <- ifelse(recruits$treartgruppe9 == "LAUV", recruits$elgLAUVha9, recruits$tretetthet9)

head(recruits$tretetthet9, n = 12)
sum(is.na(recruits$tretetthet9)) #5, meaning that there were no trees to count? But then it should have been 0, no?


```


#### Make column beitetrykk9
```{r}
recruits$beitetrykk9 <- ifelse(recruits$treartgruppe9 == "ROS", recruits$elgbROSprs9, NA)
recruits$beitetrykk9 <- ifelse(recruits$treartgruppe9 == "FURU", recruits$elgbFURUprs9, recruits$beitetrykk9)
recruits$beitetrykk9 <- ifelse(recruits$treartgruppe9 == "LAUV", recruits$elgbLAUVprs9, recruits$beitetrykk9)

head(recruits$beitetrykk9, n = 12)
sum(is.na(recruits$beitetrykk9)) #13724, meaning that there could not have been browsing pressure, because there were no trees to browse on 


## NA the ones where tretetthet is 0, but we have beitetrykk values
recruits$beitetrykk9 <- ifelse(recruits$tretetthet9 != 0, recruits$beitetrykk9, NA)
sum(is.na(recruits$beitetrykk9)) #13827, meaning that there could not have been browsing pressure, because there were no trees to browse on  


```


#### Make column proplt (which will be height class 4 (>5cm BHD) from the 10th and 11th cycle) 
```{r}
recruits$recruitment <- ifelse(recruits$treartgruppe9 == "ROS", recruits$nyrosprha1011, NA)
recruits$recruitment <- ifelse(recruits$treartgruppe9 == "FURU", recruits$nyfuruprha1011, recruits$recruitment)
recruits$recruitment <- ifelse(recruits$treartgruppe9 == "LAUV", recruits$nylauvprha1011, recruits$recruitment)

head(recruits$recruitment)
sum(is.na(recruits$recruitment)) #0, because there were no trees to count in that height class? Or where there no trees of the species? If there were trees of the species but not in height class, should in not be 0?


recruits$recruitment_ratio <- ifelse(recruits$treartgruppe9 == "ROS", recruits$recros1011, NA)
recruits$recruitment_ratio <- ifelse(recruits$treartgruppe9 == "FURU", recruits$recfuru1011, recruits$recruitment_ratio)
recruits$recruitment_ratio <- ifelse(recruits$treartgruppe9 == "LAUV", recruits$reclauv1011, recruits$recruitment_ratio)

head(recruits$recruitment_ratio)
sum(is.na(recruits$recruitment_ratio)) #0


```


#### Make new variable moose density (mean over 5 year period to datasets and models)
```{r}
library(readr)
moose_mean_density <- read_csv("processdata/five_y_mean_density_2021_02_01.csv", na = "NA")
str(moose_mean_density)
library(tidyverse)
moose_mean_density <- moose_mean_density %>%
  rename(KOMNR = knr2017) %>%
  rename(moose_density = five_y_mean_density)
str(moose_mean_density)

recruits <- left_join(recruits, moose_mean_density, by = "KOMNR")
str(recruits$moose_density)
sum(is.na(recruits$moose_density)) # 4560 --> Either "elgareal" was 0 or NA
which(is.na(recruits$moose_density))
recruits$moose_density <- ifelse(is.na(recruits$moose_density), 0, recruits$moose_density)
sum(is.na(recruits$moose_density))

```

#### Reduce data set to variables we are going to use further
```{r}
names(recruits)
recruits <- recruits[,-c(7:12,22:35)] 
names(recruits)
```



#### Stuff like... NAs, level order, etc.
```{r}

nrow(recruits)
nrow(na.omit(recruits))

sum(is.na(recruits$distvei2))
sum(is.na(recruits$disthus)) 
sum(is.na(recruits$HOH))
sum(is.na(recruits$region))
sum(is.na(recruits$skogkategori))
sum(is.na(recruits$kant))
sum(is.na(recruits$helling)) #417, what does it mean when there is NA in helling? helling from 1-100, so I guess NA is when there is no helling, so 0?
sum(is.na(recruits$treartgruppe9))
sum(is.na(recruits$beitetrykk9)) # 13827, because there are 0 trees (in the 9th cycle), hence no BP --> still, there are recruits in the 10th/11th... 
sum(is.na(recruits$tretetthet9)) #5, meaning no trees to count? Should these not just be 0 then? Or delete?
sum(is.na(recruits$recruitment))
sum(is.na(recruits$recruitment_ratio))


recruits$distvei2[is.na(recruits$distvei2)] <- 2218 # max distance to roads
recruits$disthus[is.na(recruits$disthus)] <- 3000# max distance to house
recruits$helling[is.na(recruits$helling)] <- 0 # no helling

nrow(recruits) 
nrow(na.omit(recruits)) # only NAs left are the ones in beitetrykk and tretetthet
recruits <- na.omit(recruits)


library(ggplot2)
ggplot(data = recruits, aes(skogkategori)) +
  geom_bar()+
  theme_classic()

ggplot(data = recruits, aes(kant)) +
  geom_bar()+
  theme_classic()

recruits$skogkategori <- factor(recruits$skogkategori, levels = c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog",  "Lavproduktiv_ungskog", "Uproduktiv_skog"))
recruits$kant <- factor(recruits$kant, levels = c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant")) 



```


#### How many plots have been logged in cycle 10/11?
```{r}
sum(recruits$hogd10 == 1) + sum(recruits$hogd11 == 1) # 619

recruits <- recruits[which(recruits$hogd10 == 0 & recruits$hogd11 == 0), ]
sum(recruits$hogd10 == 1) + sum(recruits$hogd11 == 1)

```


## Modelling spatial autocorrelation (https://onlinelibrary.wiley.com/doi/full/10.1111/j.2007.0906-7590.05171.x)
Change the projection and make coordinates and a spatial point data frame to make a shapefile

```{r}
library(tidyr)
recruits$UTMmerge <- paste(recruits$UTM_OV_33, recruits$UTM_SN_33, sep="_")
sum(duplicated(recruits$UTMmerge))


```


```{r}

library(sp)
library(rgdal)

crs <- CRS("+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs")
coords <- cbind(recruits$UTM_OV_33, recruits$UTM_SN_33)
data_spdf <- SpatialPointsDataFrame(coords, recruits, proj4string = crs)
# plot(data_spdf)
writeOGR(data_spdf, dsn = "output", layer = "spac", driver = "ESRI Shapefile")

```


```{r}

library(spdep)
library(ape)


# spdep package:

# prepare neighbour lists for spatial autocorrelation analysis
nb.list <- dnearneigh(as.matrix(recruits[recruits$beitetrykk9, c("UTM_OV_33", "UTM_SN_33")]), 0, 5000)
nb.weights <- nb2listw(nb.list, zero.policy = TRUE)
# moran.test(recruits$beitetrykk9, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)


# Make a matrix of coordinates - use when chunk above was not run
coords <- as.matrix(cbind(recruits$UTM_OV_33, recruits$UTM_SN_33))

# compute the autocovariate based on the above distance and weight
ac <- autocov_dist(recruits$beitetrykk9, coords, nbs = 5000, type="inverse", zero.policy = TRUE)
recruits$ac <- ac



```


#### Threshold distance to house
```{r}
recruits$disthus_600 <- ifelse(recruits$disthus >= 600, 600, recruits$disthus)
recruits$disthus_500 <- ifelse(recruits$disthus >= 500, 500, recruits$disthus)
recruits$disthus_400 <- ifelse(recruits$disthus >= 400, 400, recruits$disthus)
recruits$disthus_350 <- ifelse(recruits$disthus >= 350, 350, recruits$disthus)
recruits$disthus_300 <- ifelse(recruits$disthus >= 300, 300, recruits$disthus)
recruits$disthus_200 <- ifelse(recruits$disthus >= 200, 200, recruits$disthus)
recruits$disthus_150 <- ifelse(recruits$disthus >= 150, 150, recruits$disthus)

```






#### Reduce data set to variables used further
```{r}
# names(recruits)
# recruits <- recruits[,-c(3:5,16:17,21)]
# names(recruits)
# 
# save(recruits, file = "recruits_2021_02_01.rda")


```


#### Threshold distance to road
```{r}
# load("processdata/recruits_2021_02_01.rda")
# names(recruits)

recruits$distvei_600 <- ifelse(recruits$distvei2 >= 600, 600, recruits$distvei2)
recruits$distvei_500 <- ifelse(recruits$distvei2 >= 500, 500, recruits$distvei2)
recruits$distvei_400 <- ifelse(recruits$distvei2 >= 400, 400, recruits$distvei2)
recruits$distvei_350 <- ifelse(recruits$distvei2 >= 350, 350, recruits$distvei2)
recruits$distvei_300 <- ifelse(recruits$distvei2 >= 300, 300, recruits$distvei2)
recruits$distvei_200 <- ifelse(recruits$distvei2 >= 200, 200, recruits$distvei2)
recruits$distvei_150 <- ifelse(recruits$distvei2 >= 150, 150, recruits$distvei2)
recruits$distvei_100 <- ifelse(recruits$distvei2 >= 100, 100, recruits$distvei2)
recruits$distvei_50 <- ifelse(recruits$distvei2 >= 50, 50, recruits$distvei2)


sum(is.na(recruits)) # 0


names(recruits)
recruits <- recruits[,-c(3:5,16:17,22)]
names(recruits)

save(recruits, file = "processdata/recruits_2021_04_19.rda")

```



<!-- ## Roads data set -->
<!-- ```{r} -->
<!-- load("processdata/recruits_2021_02_01.rda") -->
<!-- names(recruits) -->
<!-- recruits_roads <- recruits[,-c(4,18:23)] -->
<!-- names(recruits_roads) -->
<!-- recruits_roads <- recruits_roads[ which(recruits_roads$disthus > 600), ] # Threshold for the bin model was 500, but for the beta model 600 so we use 600  -->
<!-- summary(recruits_roads$disthus) -->

<!-- str(recruits_roads) -->


<!-- save(recruits_roads, file = "recruits_roads_2021_02_03.rda") -->

<!-- ``` -->
