==> 02_browsing_pressure.Rmd <==
---
title: "browsing_pressure"
author: "anne"
date: "13 november 2018"
output: word_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```





```{r}

# load("processdata/browse_sub.rda")
load("processdata/browse_9.rda")
# load("processdata/browse9bro.rda")
# load("processdata/browse9bro_dist.rda")
# load("processdata/browse_10.rda")
# load("processdata/broadhouse9.rda")
# load("processdata/broadhouse10.rda")
load("processdata/browse_9_dist.rda")

```



# gam

```{r}
ggplot(browse_9, aes(x = log(disthus), y = beitetrykk)) +
    geom_point() +
    stat_smooth(method = "glm", color = "green") +
  stat_smooth(color = "blue")

# +
#   scale_x_continuous(breaks=c(100,150, 200, 250, 300, 400, 500, 1000, 1500, 2000, 3000))

```


```{r}
browse_9$disthus_cut500 <- ifelse(browse_9$disthus >= 500, 500, browse_9$disthus)
browse_9$disthus_cut550 <- ifelse(browse_9$disthus >= 550, 550, browse_9$disthus)
browse_9$disthus_cut450 <- ifelse(browse_9$disthus >= 450, 450, browse_9$disthus)
browse_9$disthus_cut400 <- ifelse(browse_9$disthus >= 400, 400, browse_9$disthus)
browse_9$disthus_cut350 <- ifelse(browse_9$disthus >= 350, 350, browse_9$disthus)
browse_9$disthus_cut300 <- ifelse(browse_9$disthus >= 300, 300, browse_9$disthus)
browse_9$disthus_cut250 <- ifelse(browse_9$disthus >= 250, 250, browse_9$disthus)
browse_9$disthus_cut200 <- ifelse(browse_9$disthus >= 200, 200, browse_9$disthus)
browse_9$disthus_cut150 <- ifelse(browse_9$disthus >= 150, 150, browse_9$disthus)
browse_9$disthus_cut180 <- ifelse(browse_9$disthus >= 180, 180, browse_9$disthus)
browse_9$disthus_cut210 <- ifelse(browse_9$disthus >= 210, 210, browse_9$disthus)
browse_9$disthus_cut220 <- ifelse(browse_9$disthus >= 220, 220, browse_9$disthus)

# 210 is the thing! (the threshold)




```

```{r}
library(mgcv)
gam1 <- gam(beitetrykk ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet,
             data = browse_9,
             method = "REML")

gam1a <- gam(beitetrykk ~ log(distvei1 + 1) + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet,
             data = browse_9,
             method = "REML")

gam2 <- gam(beitetrykk ~ s(distvei1) + disthus_cut550 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML") 

gam3 <- gam(beitetrykk ~ s(distvei1) + disthus_cut500 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam4 <- gam(beitetrykk ~ s(distvei1) + disthus_cut450 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam5 <- gam(beitetrykk ~ s(distvei1) + disthus_cut400 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam6 <- gam(beitetrykk ~ s(distvei1) + disthus_cut350 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam7 <- gam(beitetrykk ~ s(distvei1) + disthus_cut300 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam8 <- gam(beitetrykk ~ s(distvei1) + disthus_cut250 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam9 <- gam(beitetrykk ~ s(distvei1) + disthus_cut200 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam10 <- gam(beitetrykk ~ s(distvei1) + disthus_cut150 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam11 <- gam(beitetrykk ~ s(distvei1) + disthus_cut180 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam12 <- gam(beitetrykk ~ s(distvei1) + disthus_cut210 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")

gam13 <- gam(beitetrykk ~ s(distvei1) + disthus_cut220 + skogkategori + treartgruppe + region + kant + s(helling) + s(HOH) + s(tretetthet),
             data = browse_9,
             method = "REML")


```





Testing/ selection
```{r}
anova.gam()


AIC(gam2, gam3, gam4, gam5, gam6, gam7, gam8, gam9, gam10, gam11, gam12, gam13)

```



```{r}
summary(gam1)
plot.gam(gam1)
plot.gam(gam1, all.terms = TRUE)
plot.gam(gam1a, all.terms = TRUE)

par(mfrow = c(2,2))
gam.check(gam2)
gam.check(gam3)

```










# (Generalized) Linear (mixed) models

---- !All models in the different sections (with different datasets) have the same names! ----

### with browse_9
```{r}

m1lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse_9)
## m1lm fits data better than m2lm and m3lm.
# m2lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + treartgruppe, data = browse_9)
# 
# m3lm <- lm(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + treartgruppe, data = browse_9)

m4lm <- lm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse_9)

m5lm <- lm(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse_9)

m5lmN <- lm(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)

m6lm <- lm(beitetrykkr ~ distvei1 + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse_9)

m7lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1, data = browse_9)

m7alm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1, data = browse_9)

m8lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + treartgruppe*distvei1, data = browse_9)

m9lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + skogkategori*distvei1 + treartgruppe*distvei1, data = browse_9)

m10lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse_9)

m11lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + tretetthet*distvei1, data = browse_9)

m12lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, data = browse_9)

m13lm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1, data = browse_9)



m1glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "poisson", 
            data = browse_9)

m2glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasipoisson", 
            data = browse_9)

m3glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasi", 
            data = browse_9)

# m2glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "Gamma", 
#             data = browse_9)

m4glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse_9)

# m5glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "inverse.gaussian", 
#             data = browse_9)

m5glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse_9)

## weights on tretetthet makes the model very bad.
# m6glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "gaussian", 
#             weights = tretetthet,
#             data = browse_9)
# 
# m7glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "poisson", 
#             weights = tretetthet,
#             data = browse_9)

m8glm <- glm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "quasipoisson", 
            data = browse_9)




m1mix <- glmer(beitetrykkr ~ scale(distvei3) + scale(disthus) + skogkategori + kant + scale(helling) + hohkatb + treartgruppe + (1|region), 
            data = browse_9,
            family = poisson())

m2mix <- lmer(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + hohkatb + treartgruppe + (1|region), data = browse_9)

m3mix <- lmer(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + hohkatb + treartgruppe + (1|landsdel), data = browse_9)




```

Testing fit and effect plots
```{r}


AIC(m4lm, m5lm, m6lm, m7lm, m8lm, m9lm, m10lm, m11lm, m12lm, m8glm)
AIC(m1glm, m2glm, m3glm, m4glm, m5glm, m6glm, m7glm, m8glm)
AIC(m7lm, m7alm)
AIC(m10lm, m11lm)

plot(m13lm)

plot(allEffects(m1lm))
plot(allEffects(m4lm))
plot(allEffects(m5lm))
plot(allEffects(m7lm))
plot(allEffects(m10lm))
plot(allEffects(m12lm))


plot(allEffects(m1glm))
plot(allEffects(m2glm))
plot(allEffects(m3glm))
plot(allEffects(m4glm))
plot(allEffects(m7glm))



plot(predictorEffect("distvei3", m4lm))
plot(predictorEffect("distvei1", m5lm))
plot(Effect(c("distvei1","skogkategori"), m7lm))
plot(Effect(c("distvei1","skogkategori"), m12lm))
plot(Effect(c("distvei1","treartgruppe"), m12lm))


plot(allEffects(m2mix))
plot(allEffects(m3mix))



plot(m1lm)
plot(m10lm)

plot(m2glm) # quasipoisson (accounts for overdispersion)
plot(m4glm) # gaussian


summary(m1lm)
summary(m4lm)
summary(m5lm)
summary(m5lmN)
summary(m7lm)
summary(m10lm)

summary(m1glm)
summary(m2glm)
summary(m4glm)
summary(m5glm)
summary(m6glm)

summary(m1mix)
summary(m2mix)
summary(m3mix)

anova(m3mix)


hist(resid(m1lm))
hist(resid(m1glm))
hist(resid(m4glm))

hist(resid(m1mix))
hist(resid(m2mix))
hist(resid(m3mix))

```







### With browse9bro
```{r}

m1lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)
## m1lm fits data better than m2lm and m3lm.
# m2lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + treartgruppe, data = browse9bro)
# 
# m3lm <- lm(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + treartgruppe, data = browse9bro)

m4lm <- lm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)

m5lm <- lm(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)
m5lmN <- lm(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)

m6lm <- lm(beitetrykkr ~ distvei1 + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro)

m7lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + skogkategori*distvei1, data = browse9bro)

m7alm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + + kant + helling + hohkatb + skogkategori*distvei1, data = browse9bro)

m8lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + treartgruppe*distvei1, data = browse9bro)

m9lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + skogkategori*distvei1 + treartgruppe*distvei1, data = browse9bro)


m10lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro)


m11lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + tretetthet*distvei1, data = browse9bro)


m12lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, data = browse9bro)

m13lm <- lm(beitetrykkr ~ veitype1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro)

m14lm <- lm(beitetrykkr ~ distvei1 + vegetasjon + treartgruppe + region + + kant + helling + hohkatb + tretetthet + vegetasjon*distvei1, data = browse9bro)





m1glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "poisson", 
            data = browse9bro)

m2glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasipoisson", 
            data = browse9bro)

m3glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasi", 
            data = browse9bro)

# m2glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "Gamma", 
#             data = browse9bro)

m4glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse9bro)

# m5glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "inverse.gaussian", 
#             data = browse9bro)

m5glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse9bro)

## weights on tretetthet makes the model very bad.
# m6glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "gaussian", 
#             weights = tretetthet,
#             data = browse9bro)
# 
# m7glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "poisson", 
#             weights = tretetthet,
#             data = browse9bro)

m8glm <- glm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "quasipoisson", 
            data = browse9bro)

m9glm <- glm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "gaussian", 
            data = browse9bro)






m1mix <- glmer(beitetrykkr ~ scale(distvei3) + scale(disthus) + skogkategori + kant + scale(helling) + hohkatb + treartgruppe + (1|region), 
            data = browse9bro,
            family = poisson())

m2mix <- lmer(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + hohkatb + treartgruppe + (1|region), data = browse9bro)

m3mix <- lmer(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + hohkatb + treartgruppe + (1|landsdel), data = browse9bro)

m4mix <- lmer(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + (1|region), data = browse9bro)


```

Testing fit and effect plots
```{r}


AIC(m1lm, m4lm, m5lm, m6lm, m7lm, m8lm, m9lm, m10lm, m11lm, m12lm)
AIC(m1glm, m2glm, m3glm, m4glm, m5glm, m8glm, m9glm)

AIC(m7lm, m7alm)
AIC(m10lm, m14lm)
AIC(m10lm, m9glm)

AIC(m1mix, m2mix, m3mix, m4mix)


plot(allEffects(m1lm))
plot(allEffects(m4lm))
plot(allEffects(m5lm))
plot(allEffects(m7lm))
plot(allEffects(m10lm))
plot(allEffects(m12lm))
plot(allEffects(m13lm))

plot(allEffects(m1glm))
plot(allEffects(m2glm))
plot(allEffects(m3glm))
plot(allEffects(m4glm))
plot(allEffects(m7glm))



plot(predictorEffect("distvei3", m4lm))
plot(predictorEffect("distvei1", m5lm))
plot(Effect(c("distvei1","skogkategori"), m7lm))
plot(Effect(c("distvei1","skogkategori"), m10lm))
plot(Effect(c("distvei1","skogkategori"), m12lm))
plot(Effect(c("distvei1","treartgruppe"), m12lm))


plot(allEffects(m2mix))
plot(allEffects(m3mix))
plot(allEffects(m4mix))


plot(m1lm)
plot(m10lm)

plot(m2glm) # quasipoisson (accounts for overdispersion)
plot(m4glm) # gaussian


summary(m1lm)
summary(m4lm)
summary(m5lm)
summary(m5lmN)
summary(m7lm)
summary(m10lm) # best
summary(m12lm) # (2nd best with 3 points difference)
summary(m13lm) # best, but logical?
summary(m14lm)

summary(m1glm)
summary(m2glm)
summary(m4glm)
summary(m5glm)
summary(m6glm)
summary(m9glm) # best


summary(m1mix)
summary(m2mix)
summary(m3mix)
summary(m4mix) # best

anova(m3mix)


hist(resid(m10lm))
hist(resid(m9glm))

hist(resid(m1mix))
hist(resid(m2mix))
hist(resid(m3mix))
hist(resid(m4mix))

```







### With browse9bro_dist

```{r}


m1lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro_dist)
## m1lm fits data better than m2lm and m3lm.
# m2lm <- lm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + treartgruppe, data = browse9bro_dist)
# 
# m3lm <- lm(beitetrykkr ~ distvei3 + disthus + skogkategori + kant + helling + treartgruppe, data = browse9bro_dist)

m4lm <- lm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro_dist)

m5lm <- lm(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro_dist)

m6lm <- lm(beitetrykkr ~ distvei1 + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse9bro_dist)

m7lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1, data = browse9bro_dist)

m7alm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1, data = browse9bro_dist)

m8lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + treartgruppe*distvei1, data = browse9bro_dist)

m9lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + skogkategori*distvei1 + treartgruppe*distvei1, data = browse9bro_dist)


m10lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)
m10alm <- lm(beitetrykkr ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)
m10blm <- lm(log(beitetrykkr+1) ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)
m10clm <- lm(sqrt(beitetrykkr) ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)
m10dlm <- lm(bt100 ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)

m11lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + tretetthet*distvei1, data = browse9bro_dist)


m12lm <- lm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, data = browse9bro_dist)
m12alm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, data = browse9bro_dist)


# m13lm <- lm(beitetrykkr ~ veitype1 + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)
# m13alm <- lm(beitetrykkr ~ veitype1 + disthus + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)

m14lm <- lm(beitetrykkr ~ distvei1 + vegetasjon + treartgruppe + region + + kant + helling + hohkatb + tretetthet + vegetasjon*distvei1, data = browse9bro_dist)

m15lm <- lm(beitetrykkr ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + tretetthet*distvei1, data = browse9bro_dist)

m16lm <- lm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + tretetthet*distvei1, data = browse9bro_dist)

m17lm <- lm(beitetrykkr ~ distvei1 + disthus+ skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, data = browse9bro_dist)



# ------------------- glm ----------------



m1glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "poisson", 
            data = browse9bro_dist)

m2glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasipoisson", 
            data = browse9bro_dist)

m3glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "quasi", 
            data = browse9bro_dist)

# m2glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "Gamma", 
#             data = browse9bro_dist)

m4glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse9bro_dist)

# m5glm <- glm(beitetrykk ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "inverse.gaussian", 
#             data = browse9bro_dist)

m5glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
            family = "gaussian", 
            data = browse9bro_dist)

## weights on tretetthet makes the model very bad.
# m6glm <- glm(beitetrykkr ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "gaussian", 
#             weights = tretetthet,
#             data = browse9bro_dist)
# 
# m7glm <- glm(beitetrykkr ~ distvei3 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, 
#             family = "poisson", 
#             weights = tretetthet,
#             data = browse9bro_dist)

m8glm <- glm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "quasipoisson", 
            data = browse9bro_dist)

m9glm <- glm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "gaussian", 
            data = browse9bro_dist)
m9aglm <- glm(log(beitetrykkr+1) ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "gaussian", 
            data = browse9bro_dist)

m9dglm <- glm(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "gaussian", 
            data = browse9bro_dist)

m10aglm <- glm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1, 
            family = "poisson", 
            data = browse_9_dist)


m11glm <- glm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1, 
            family = "poisson", 
            data = browse_9_dist)

m12glm <- glm(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1, 
            family = "negbin", 
            data = browse_9_dist)

plot(m11glm)



library(MASS)
m9bglm <- glm.nb(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1, 
            data = browse_9_dist)

plot(m9bglm)

m9cglm <- glm.nb(log(beitetrykkr+1) ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            data = browse9bro_dist)

m11glm <- glm.nb(log(beitetrykkr+1) ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, 
             data = browse9bro_dist)

m11aglm <- glm.nb(log(beitetrykkr+1) ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1 + treartgruppe*distvei1, 
             data = browse9bro_dist)

m12glm <- glm(btbinary ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "binomial", 
            data = browse9bro_dist)


# Binomial logistic regression with two-column integer matrix as response
m11aglm <- glm(cbind(beitetrykkr, nobro) ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "binomial", 
            weights = tretetthet,
            data = browse9bro_dist)

m11bglm <- glm(cbind(beitetrykkr, nobro) ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "binomial", 
            data = browse9bro_dist)

m11cglm <- glm(cbind(beitetrykkr, nobro) ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            family = "quasibinomial", 
            data = browse9bro_dist) # was supposed to help with overdispersion (residual deviance is much higer than residual d.f.), but no change!

plot(allEffects(m11aglm))
plot(allEffects(m11bglm))

```


#### Beta regression
```{r}
# ------------------- beta regression ----------------

library(betareg)
# for the normal betareg to run the response has to be altered to have no 0.0, but e.g. 0.01 instead, which of course makes it wrong
m1br <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1, 
            data = browse9bro_dist)



# zero-or-one inflated beta regression
# zoib
starttime <- Sys.time()

m5zibr <- zoib(bt100 ~ distvei1|1|distvei1 + disthus + skogkategori|1,
            data = browse9bro_dist,
            zero.inflation = c(T), one.inflation = c(F))
endtime <- Sys.time()
endtime-starttime # Time difference of 1.118052 days (mod4)
save(m5zibr, file = "m5zibr.rda")

load("output/m2zibr.rda")
load("output/m3zibr.rda")
load("output/m4zibr.rda")

summary(m2zibr)
summary(m3zibr)
summary(m4zibr)
coeff2 <- m2zibr$coeff
coeff3 <- m3zibr$coeff
coeff4 <- m4zibr$coeff
summary(coeff2)
summary(coeff3)
summary(coeff4)
xb <- m2zibr$Xb
x0 <- m2zibr$X0
x1 <- m2zibr$X1
yobs <- m2zibr$yobs
ypred <- m2zibr$ypred
plot(ypred[[2]])
plot(yobs)

traceplot(coeff)
autocorr.plot(coeff)
check.psrf(coeff)

summ1 <- summary(m2zibr)
summ1 <- cbind(summ1$stat[,1],summ1$quant[,c(1,5)]) # does not work... why?
paraplot()

dic.samples(m2zibr)


n <- nrow(browse9bro_dist)
ypred1 <- rbind(m2zibr$ypred[[1]][,1:n], m2zibr$ypred[[2]][,1:n])
ypred2 <- rbind(m2zibr$ypred[[1]][,(n+1):(2*n)], m2zibr$ypred[[2]][,(n+1):(2*n)])
post.mean1<- apply(ypred1,2,mean)
post.mean2 <- apply(ypred2,2,mean)
plot(browse9bro_dist$y1, post.mean1, xlim=c(0,1),ylim=c(0,1), col='green2', pch=2,xlab='Observed y', ylab='Predicted y', main="")
points(browse9bro_dist$y2,post.mean2,col='purple')
abline(0,1,col='red')
legend(0.1,0.9,col=c('green2','purple'),c("y1","y2"),pch=c(2,1))



# -----------  gamlss -----------
# m1gamlss <- gamlss(bt100 ~ distvei1,
#                    family = BEZI, 
#                    data = browse9bro_dist)
# summary(m1gamlss)
# plot(m1gamlss)
# 
# fitted(m1gamlss)[1]
# summary(m1gamlss)
# fitted(m1gamlss,"mu")[1]         #fitted mu
# fitted(m1gamlss,"sigma")[1]      #fitted sigma
# fitted(m1gamlss,"nu")[1]         #fitted nu
# meanBEZI(m1gamlss)[1] # expected value of the response
# plotBEZI(mu = 0.36, sigma = 1.5, nu = 0.3, from = 0, to = 0.99, n = 12680)

```



#### Testing fit and effect plots
```{r}

AIC(m1lm, m4lm, m5lm, m6lm, m7lm, m7alm, m8lm, m9lm, m10lm, m10alm, m11lm, m12lm, m12alm, m14lm, m15lm, m16lm)
AIC(m1glm, m2glm, m3glm, m4glm, m5glm, m8glm, m9glm, m9aglm)
AIC(m11aglm, m11bglm, m11cglm)
AIC(m9bglm, m9cglm, m11glm)

# candidates
AIC(m9aglm, m9bglm, m9cglm, z3a, m1br, m12glm)
AIC(m9dglm, gam14, m1gamlss)


AIC(m7lm, m7alm)
AIC(m13lm, m13alm)
AIC(m10lm, m9glm)
AICc(m2glm, m8glm)
AIC(m10lm, m10alm)
AIC(m10alm, m10blm, m10clm)
AIC(m11aglm, m11bglm)

AIC(m9aglm, m11bglm, h2b, z3a, gam14)
BIC(m9aglm, m11bglm, h2b, z3a)



plot(allEffects(m1lm))
plot(allEffects(m4lm))
plot(allEffects(m5lm))
plot(allEffects(m7lm))
plot(allEffects(m10lm))
plot(allEffects(m10alm))
plot(allEffects(m12lm))
plot(allEffects(m12alm))
# plot(allEffects(m13lm))
# plot(allEffects(m13alm))
plot(allEffects(m15lm))
plot(allEffects(m16lm))

plot(allEffects(m1glm))
plot(allEffects(m2glm))
plot(allEffects(m3glm))
plot(allEffects(m4glm))
plot(allEffects(m7glm))
plot(allEffects(m9glm))
plot(allEffects(m9aglm))
plot(allEffects(m9dglm))
plot(allEffects(m10aglm))
plot(allEffects(m11bglm))
plot(allEffects(m12glm))

plot(allEffects(m1br))



plot(predictorEffect("distvei3", m4lm))
plot(predictorEffect("distvei1", m5lm))
plot(Effect(c("distvei1","skogkategori"), m7lm))
plot(Effect(c("distvei1","skogkategori"), m10lm))
plot(Effect(c("distvei1","skogkategori"), m12lm))
plot(Effect(c("distvei1","treartgruppe"), m12lm))


plot(allEffects(m2mix))
plot(allEffects(m3mix))
plot(allEffects(m4mix))


par(mfrow = c(2, 2))
plot(m10alm)
plot(m10blm)
plot(m12alm)

plot(m9aglm)
plot(m9bglm)
plot(m9cglm)
plot(m9dglm)
plot(m10aglm)

plot(m1br)
plot(m1gamlss)
gam.check(gam14)

drop1(m9bglm, test = "Chi")
drop1(m9cglm, test = "Chi")


plot(m8glm) # quasipoisson (accounts for overdispersion)
plot(m10aglm) # poisson


# ------- linear models ------
summary(m1lm)
summary(m4lm)
summary(m5lm)
summary(m5lmN)
summary(m7lm)
summary(m10lm) # best
summary(m12lm) # 2nd best with 3 points difference
# summary(m13lm) # best, but logical?
summary(m14lm)

# ------- generalized linear models --------
summary(m1glm)
summary(m2glm)
summary(m4glm)
summary(m5glm)
summary(m6glm)
summary(m9aglm) 
summary(m9bglm) # best
summary(m9cglm) # best (without house)

summary(m1br) # best (beta reg)


# summary(m1mix)
# summary(m2mix)
# summary(m3mix)
# summary(m4mix) # best



hist(resid(m10lm))
hist(resid(m10alm))
hist(resid(m12lm))
hist(resid(m12alm))

plot(resid(m10alm))
plot(resid(m12alm))

plot(resid(m9bglm))
plot(resid(m9cglm))



# test for overdispersion

# m10aglm - poisson
E1 <- resid(m10aglm, type = "pearson")
Dispersion <- sum(E1^2) / m10aglm$df.residual

dispersiontest(m10aglm, trafo = NULL, alternative = c("greater", "two.sided", "less"))

# m9bglm - glm.nb
E2 <- resid(m9bglm, type = "pearson")
Dispersion2 <- sum(E2^2) / m9bglm$df.residual

# m9cglm - glm.nb
E2a <- resid(m9cglm, type = "pearson")
Dispersion2a <- sum(E2a^2) / m9cglm$df.residual

# z3a - zeroinfl
N <- nrow(browse9bro_dist)
E3 <- resid(z3a, type = "pearson")
Dispersion3 <- sum(E2^2) / (N - 17) # best

# m11glm - glm.nb
E4 <- resid(m11glm, type = "pearson")
Dispersion4 <- sum(E4^2) / m11glm$df.residual

# m11aglm - glm.nb
E5 <- resid(m11aglm, type = "pearson")
Dispersion5 <- sum(E5^2) / m11aglm$df.residual


# m1br - beta regression
E6 <- resid(m1br, type = "pearson")
Dispersion6 <- sum(E6^2) / m1br$df.residual

# m9dglm - glm with bt100 (0,0-1,0)
E7 <- resid(m9dglm, type = "pearson")
Dispersion7 <- sum(E7^2) / m9dglm$df.residual



```
