==> 03_zero_inflated_browsing_pressure.Rmd <==
---
title: "hurdle"
author: "anne"
date: "3 desember 2018"
output: word_document
editor_options: 
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}

load("processdata/browse_9.rda")
# load("processdata/browse9pos.rda")
load("processdata/browse9bro_dist.rda")
```




## Hurdle model (Also: two-way model) and two-way model "old school" (Zuur 2009, chapter: Zero-truncated and zero-inflated models for count data)
For count data, helps with excess zeros and overdispersion

##### Hurdle model, using hurdle() from pscl package
```{r}

f1 <- formula(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe)

f2 <- formula(beitetrykkr ~ distvei1 + disthus + region + skogkategori + kant + helling + HOH + treartgruppe + skogkategori*distvei1)

f3 <- formula(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1)
f3a <- formula(cbind(beitetrykkr, nobro) ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1)

f4 <- formula(beitetrykkr ~ distvei1 + disthus + skogkategori + treartgruppe + region + + kant + helling + hohkatb + tretetthet + skogkategori*distvei1)


# hurdle
h1a <- hurdle(f2, dist = "poisson",  
              data = browse_9_dist)  

h1b <- hurdle(f2, dist = "negbin", 
              data = browse_9_dist)  


h2a <- hurdle(f2, dist = "negbin", 
              data = browse9bro_dist)

h2b <- hurdle(f3, dist = "negbin", 
              data = browse9bro_dist)

h2b2 <- hurdle(f3a, dist = "negbin", 
              data = browse9bro_dist)

h2c <- hurdle(f3, dist = "poisson", 
              data = browse9bro_dist)



hglm0 <- glm(factor(beitetrykkr > 0) ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse9bro_dist)

# USE BETAREG here! or use the binomial with cbind
hglm1 <- glm(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + hohkatb + tretetthet + skogkategori*distvei1,
             data = browse9bro_dist,
            family = ztpoisson,
            subset = beitetrykkr > 0)


# zero inflated

f3 <- formula(beitetrykkr ~ distvei1 + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet)

f5 <- formula(proplt ~ log(distvei2 + 1) + skogkategori + treartgruppe9 + HOH +  
    I(HOH^2) + browspres_merged)

zinbdata <- browse9.6sc

zinbdata$proplt <- zinbdata$proplt*100
zinbdata$proplt <- as.integer(zinbdata$proplt)

zinbTest <- zeroinfl(f5, dist = "negbin", 
              data = zinbdata)

z3a <- zeroinfl(f3, dist = "negbin", 
              data = browse9bro_dist)

z3b <- zeroinfl(f3, dist = "geometric", 
              data = browse9bro_dist)

z3c <- zeroinfl(f4, dist = "negbin", 
              data = browse9bro_dist)

z3d <- zeroinfl(f3, dist = "poisson", 
              data = browse9bro_dist)

zinbTest2 <- glmmTMB(f5, data = zinbdata, family = nbinom1(), ziformula = ~0,
  dispformula = ~1, weights = NULL, offset = NULL,
  contrasts = NULL, na.action = na.fail, se = TRUE,
  verbose = FALSE, doFit = TRUE, control = glmmTMBControl(),
  REML = FALSE)

summary(zinbTest)
predictedZINB <- predict(zinbTest)
plot(predictedZINB)

plot_model(zinbTest2, type = c("eff"))
plot(allEffects(zinbTest2))

## These are glms that can be used instead of hurdle, so its possible to use effect plots. But it only works for poisson....
# H1btest <- glm(factor(beitetrykkr > 0) ~ distvei3 + disthus + region + skogkategori + kant + helling + hohkatb + treartgruppe, data = browse_91,
#                family = binomial(link = "cloglog"))
# dv0c <- glm(visits ~ gender + health + income + poly(age, 2), data = DoctorVisits, family = ztpoisson, subset = visits > 0)


# hurdletest(H2c) ---- I don't get why this is not working

plot(residuals(h1a) ~ fitted(h1a))
plot(residuals(h1b) ~ fitted(h1b))
hist(resid(z3a))
hist(resid(h1a))

summary(h1a)
summary(h1b)
summary(h2b) # best
summary(h2c)

summary(z3a) # best
summary(z3b)

f <- fitted(z3a)
p <- predict(z3a)
resid <- residuals(z3a, type = "pearson")
plot(resid ~ f)
plot(resid ~ browse9bro_dist$distvei1)
plot(resid ~ browse9bro_dist$disthus)
plot(resid ~ browse9bro_dist$skogkategori)
plot(resid ~ browse9bro_dist$treartgruppe)
plot(resid ~ browse9bro_dist$region)
plot(resid ~ browse9bro_dist$kant)
plot(resid(z3a))


plot(allEffects(hglm0))
plot(allEffects(hglm1))

lrtest(h1a, h1b, h2a, h2b, h2c)
lrtest(z3a, z3c)

AIC(h2b)
AIC(z3a)
AICc(h2b)
AIC(z3a, z3c)


# For zeroinfl - shows real and not real zeros
plot(factor(beitetrykkr == 0) ~ distvei1, data = browse9bro_dist, main = "Zero")
plot(log(beitetrykkr) ~ distvei1, data = browse9bro_dist, subset = beitetrykkr > 0,
  main = "Count")

```





##### Two-way model "old school" with first running a binomial glm with a binomial respose (browsing yes/no), then running a poisson glm 
```{r}

B1 <- glm(bt.binom ~ distvei1 + disthus + landsdel + skogkategori + treartgruppe + kant + helling + hohkatb, data = browse9,
            family = "binomial")
P1 <- glm(beitetrykk ~ distvei1 + disthus + landsdel + skogkategori + treartgruppe + kant + helling + hohkatb, data = browse9pos,
            family = "poisson") 
P2 <- glm(beitetrykk ~ distvei1 + landsdel + skogkategori + treartgruppe + kant + helling + hohkatb, data = browse9pos,
            family = "poisson") # Here I took out the distance to house, because it was not significant. However, when comparing both AIC, the models barely differ.



summary(B1)
summary(P1)
summary(P2)
plot(P1)
plot(P2)
AIC(P1, P2)

```




```{r}
mod1 <- glm(beitetrykk ~ distvei1 + disthus + landsdel + skogkategori + kant + helling + hohkatb + treartgruppe, family = "poisson", data = browse9)

mu <- predict(mod1, type = "response") # predict expected mean count
exp <- sum(dpois(x = 0, lambda = mu)) # sum the probabilities of a 0 count for each mean
exp # predicted number of 0's
sum(browse9$beitetrykk < 1, na.rm = TRUE)  # observed number of 0's

```

```{r}

cor((as.matrix(browse9bro_dist[browse9bro_dist$beitetrykkr, c("distvei1", "disthus", "helling", "HOH")])),
      use = "complete")
```
