==> 04_new_approach_browsing_pressure.Rmd <==
---
title: "new_approach"
author: "anne"
date: "9 januar 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
load("processdata/browse9bro_dist.rda")
browse9bro_dist$beitetrykkr <- as.integer(browse9bro_dist$beitetrykkr)
browse9bro_dist$bt100 <- browse9bro_dist$beitetrykkr/100
```

```{r}
load("processdata/browse_9.rda")
# browse_9$beitetrykkr <- as.integer(round(browse_9$beitetrykk))
# browse_9$bt100 <- browse_9$beitetrykkr/100
```


```{r}
load("processdata/browse_9_dist.rda")
# browse_9_dist$beitetrykkr <- as.integer(round(browse_9_dist$beitetrykk))
# browse_9_dist$bt100 <- browse_9_dist$beitetrykkr/100
load("processdata/browse_9_distbin.rda")
load("processdata/browse_9_distbeta.rda")
```



#### Modelling spatial autocorrelation (https://onlinelibrary.wiley.com/doi/full/10.1111/j.2007.0906-7590.05171.x)
Change the projection and make coordinates and a spatial point data frame to make a shapefile

```{r}
unite(browse9bro_dist, UTMmerge, c(UTM_OV_33, UTM_SN_33), remove = FALSE)
duplicated(browse9bro_dist$UTMmerge)
```


```{r}

library(sp)
library(rgdal)

crs <- CRS("+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs")
coords <- cbind(browse9bro_dist$UTM_OV_33, browse9bro_dist$UTM_SN_33)
data_spdf <- SpatialPointsDataFrame(coords, browse9bro_dist, proj4string = crs)
plot(data_spdf)
writeOGR(data_spdf, dsn = "output", layer = "spdf", driver = "ESRI Shapefile")

```


```{r}

library(spdep)
# ?autocov_dist

# prepare neighbour lists for spatial autocorrelation analysis
# nb.list <- dnearneigh(as.matrix(browse9bro_dist[browse9bro_dist$bt100 != 0, c("UTM_OV_33", "UTM_SN_33")]), 0, 5000)
# nb.weights <- nb2listw(nb.list, zero.policy = TRUE)

# Make a matrix of coordinates - use when chunk above was not run
coords <- as.matrix(cbind(browse_9_dist$UTM_OV_33, browse_9_dist$UTM_SN_33))

# compute the autocovariate based on the above distance and weight

ac <- autocov_dist(browse_9_dist$bt100, coords, nbs = 5000, type="inverse", zero.policy = TRUE)
browse_9_dist$ac <- ac

# moran.plot(resid_m1abeta, nb.weights, zero.policy = TRUE, spChk = NULL, labels = TRUE, xlab = NULL, ylab = NULL, quiet = NULL)
# 
# moran.test(resid_m1beta, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)
# 
# moran.test(resid_m1abeta, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)

save(browse_9_dist, file = "browse_9_dist.rda")


```




#### First part of a hurdle/ zeroinlf model to account for zero inflation
##### glm
```{r}
m1bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m2bin <- glm(factor(beitetrykkr > 0) ~ distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m3bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m4bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m5bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m6bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m7bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m8bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m9bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling + HOH,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m10bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling + HOH + tretetthet,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m11bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + skogkategori + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m12bin <- glm(factor(beitetrykkr > 0) ~ skogkategori + treartgruppe + ac + region + kant + helling + HOH +  + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m13bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m14bin <- glm(factor(beitetrykkr > 0) ~ distvei1 + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m15bin <- glm(factor(beitetrykkr > 0) ~ disthus + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m15abin <- glm(factor(beitetrykkr > 0) ~ disthus + treartgruppe + ac + KOMNR + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m16bin <- glm(factor(beitetrykkr > 0) ~ disthus + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m17bin <- glm(factor(beitetrykkr > 0) ~ disthus + distvei1 + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m18bin <- glm(factor(beitetrykkr > 0) ~ disthus + skogkategori + treartgruppe + ac + region + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
            family = binomial(link = "cloglog"),
            data = browse_9_dist)

m19bin <- glm(factor(beitetrykkr > 0) ~ disthus_cut210_scale + treartgruppe + ac_scale + KOMNR + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale,
            family = binomial(),
            data = browse_9_dist)

```

```{r}
AIC(m1bin, m2bin, m3bin, m4bin, m5bin, m6bin, m7bin, m8bin, m9bin, m10bin, m11bin, m12bin, m13bin, m14bin, m15bin)
# lowest AIC m1bin and 15, 15 is the more parsimonious 
AIC(m15bin, m16bin, m17bin, m18bin) # still 15 then
AIC(m15bin, m15abin)
AIC(m15bin, m19bin)
```


```{r}
summary(m15bin)
summary(m15abin)

plot(allEffects(m15abin))
plot(Effect("region", m1bbin))
plot(Effect("disthus", m1bbin))
plot(Effect("distvei1", m1bbin))
browse_9_dist$KOMNR <- as.factor(browse_9_dist$KOMNR)
plot(Effect("KOMNR", m15abin))

coef19 <- coef(m19bin)
coef4 <- fixef(m4glmm)
coef4a <- fixef(m4aglmm)

plot(coef4a[-1], coef19[!grepl("KOMNR", names(coef19))][-1])
plot(coef4, coef4a)
abline(0,1)
cbind(coef4a, coef19[!grepl("KOMNR", names(coef19))])

```



#### glmm bin

##### Checking for the cut for disthus
```{r}
m1aglmm <- glmer(factor(beitetrykkr > 0) ~ disthus + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1bglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut500 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1cglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut450 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1dglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut400 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1eglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut350 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1fglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut300 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m1gglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut250 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9) #7

m1hglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9) #8

m1iglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut150 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9) #9

m1jglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut220 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9) #10

m1kglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut210 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)#11
```

```{r}
AIC(m1aglmm, m1bglmm, m1cglmm, m1dglmm, m1eglmm, m1fglmm, m1gglmm, m1hglmm, m1iglmm)



library(AICcmodavg)

Candmods_bin <- list(m1aglmm, m1bglmm, m1cglmm, m1dglmm, m1eglmm, m1fglmm, m1gglmm, m1hglmm, m1iglmm, m1jglmm, m1kglmm)


glmmbin_tab <- aictab(Candmods_bin, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)

# disthus_cut200 is the thing!

```

```{r}
pt1 <- plot_model(m1aglmm, type = c("eff"), transform = NULL, terms = c("disthus_scale [all]"),
           grid = TRUE)
pt2 <- plot_model(m1bglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut500 [all]"),
           grid = TRUE)
pt3 <- plot_model(m1cglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut450 [all]"),
           grid = TRUE)
pt4 <- plot_model(m1dglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut400 [all]"),
           grid = TRUE)
pt5 <- plot_model(m1eglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut350 [all]"),
           grid = TRUE)
pt6 <- plot_model(m1fglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut300 [all]"),
           grid = TRUE)
pt7 <- plot_model(m1gglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut250 [all]"),
           grid = TRUE)
pt8 <- plot_model(m1hglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut200 [all]"),
           grid = TRUE)
pt9 <- plot_model(m1iglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut150 [all]"),
           grid = TRUE)
pt10 <- plot_model(m1jglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut220 [all]"),
           grid = TRUE)
pt11 <- plot_model(m1kglmm, type = c("eff"), transform = NULL, terms = c("disthus_cut210 [all]"),
           grid = TRUE)


library(gridExtra)
plot_grid(list(pt1, pt2, pt3, pt4, pt5, pt6, pt7, pt8, pt9, pt10, pt11))

```


```{r}
browse_9_dist$HOH_scale = scale(browse_9_dist$HOH)
browse_9_dist$tretetthet_scale = scale(browse_9_dist$tretetthet)
browse_9_dist$helling_scale = scale(browse_9_dist$helling)
browse_9_dist$disthus_cut200_scale = scale(browse_9_dist$disthus_cut200)
browse_9_dist$distvei1_scale = scale(browse_9_dist$distvei1)
browse_9_dist$ac_scale = scale(browse_9_dist$ac)

browse_9_distbin$HOH_scale = scale(browse_9_distbin$HOH)
browse_9_distbin$tretetthet_scale = scale(browse_9_distbin$tretetthet)
browse_9_distbin$helling_scale = scale(browse_9_distbin$helling)
browse_9_distbin$disthus_cut200_scale = scale(browse_9_distbin$disthus_cut200)
browse_9_distbin$distvei1_scale = scale(browse_9_distbin$distvei1)
browse_9_distbin$ac_scale = scale(browse_9_distbin$ac)

browse_9_distbeta$HOH_scale = scale(browse_9_distbeta$HOH)
browse_9_distbeta$tretetthet_scale = scale(browse_9_distbeta$tretetthet)
browse_9_distbeta$helling_scale = scale(browse_9_distbeta$helling)
browse_9_distbeta$disthus_cut210_scale = scale(browse_9_distbeta$disthus_cut210)
browse_9_distbeta$distvei1_scale = scale(browse_9_distbeta$distvei1)
browse_9_distbeta$ac_scale = scale(browse_9_distbeta$ac)

browse_9$HOH_scale = scale(browse_9$HOH)
browse_9$tretetthet_scale = scale(browse_9$tretetthet)
browse_9$helling_scale = scale(browse_9$helling)
browse_9$disthus__scale = scale(browse_9$disthus)
browse_9$distvei1_scale = scale(browse_9$distvei1)
browse_9$ac_scale = scale(browse_9$ac)
browse_9$disthus_cut200_scale = scale(browse_9$disthus_cut200)
browse_9$disthus_cut210_scale = scale(browse_9$disthus_cut210)

save(browse_9, file = "browse_9.rda")
save(browse_9_distbin, file = "browse_9_distbin.rda")
save(browse_9_distbeta, file = "browse_9_distbeta.rda")

```

##### glmer model
```{r}

load("processdata/browse_9_distbin.rda")


# m1glmm <- glmer(bt_bin ~ disthus_cut210_scale + treartgruppe + ac + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|region),
#             family = binomial(link = "cloglog"),
#             data = browse_9_dist)

# m2glmm <- glmer(bt_bin ~ disthus_cut210_scale + treartgruppe + ac + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|region),
#             family = binomial(),
#             data = browse_9_dist)
# 
# m3glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut210_scale + treartgruppe + ac + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|region),
#             family = binomial(),
#             data = browse_9_dist)
# m2 and m3 bring same results in the summary ("different" response)

library(lme4)

m1glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9_dist)

m1glmmtest <- glmer(factor(beitetrykkr > 0) ~ disthus_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9_distbin)

m1glmmtest2 <- glmer(factor(beitetrykkr > 0) ~ disthus_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m2glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m4glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + distvei1_scale*skogkategori + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m4aglmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|region),
            family = binomial(),
            data = browse_9)

m5glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            family = binomial(link = "cloglog"),
            data = browse_9)

m6glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m7glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + distvei1_scale +  treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m8glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + treartgruppe*tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m9glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + distvei1_scale + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + treartgruppe*tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m10glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + distvei1_scale + skogkategori + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + treartgruppe*tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)

m11glmm <- glmer(factor(beitetrykkr > 0) ~ disthus_cut200_scale + treartgruppe + skogkategori + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + distvei1_scale*tretetthet_scale + (1|KOMNR),
            family = binomial(),
            data = browse_9)


```

```{r}

AIC(m1glmm, m2glmm, m4glmm, m4aglmm, m5glmm, m6glmm, m7glmm, m8glmm, m9glmm, m10glmm, m11glmm)
# m4glmm (or m10glmm)

```



Checking which optimizers do not converge (if that is the case for a model)
```{r}
# Check singularity (in case of model failed to converge)
tt <- getME(m4glmm,"theta")
ll <- getME(m4glmm,"lower")
min(tt[ll==0])
# Check which optimizer doesn't converge
allFit(m4glmm)
```


```{r}
library(AICcmodavg)

Candmods_bin <- list(m1glmm, m2glmm, m4glmm, m4aglmm, m5glmm, m6glmm, m7glmm, m8glmm, m9glmm, m10glmm, m11glmm)


glmmbin_tab <- aictab(Candmods_bin, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)


```

Testing for normality of errors (Although I don't know if this works for binomial)
```{r}
summary(m4glmm)

plot(m4glmm)
hist(resid(m4glmm))

resid.m4glmm <- resid(m4glmm)
plot(resid.m4glmm)

par(mfrow = c(2,3))
plot(resid.m4glmm ~ disthus_cut210_scale, data = browse_9_dist)
plot(resid.m4glmm ~ treartgruppe, data = browse_9_dist)
plot(resid.m4glmm ~ kant, data = browse_9_dist)
plot(resid.m4glmm ~ helling_scale, data = browse_9_dist)
plot(resid.m4glmm ~ HOH_scale, data = browse_9_dist)
plot(resid.m4glmm ~ tretetthet_scale, data = browse_9_dist)

plot_model(m4glmm, type = c("diag"), transform, terms = NULL)
plot_model(m4glmm, type = c("resid"), transform, terms = NULL)
#  treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + distvei1_scale*skogkategori 

```


```{r}
library(sjPlot)
tab_model(m4glmm, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE,
          show.std = NULL, show.p = TRUE, 
          show.stat = FALSE, show.df = FALSE, show.zeroinf = TRUE, show.r2 = TRUE, 
          show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, 
          show.dev = FALSE, show.obs = TRUE)


```

```{r}
p1 <- plot_model(m4glmm, type = c("est"), transform = NULL, show.values = TRUE, value.offset = 0.5, 
                 vline.color = "grey", colors = c("#004949", "#009292"),
                 axis.title = "Estimates")
p1 + theme_sjplot() + font_size(axis_title.x = 15)
# whiskers are ci, possible to show se instead; Colourbling friendly colours "#004949","#009292",
```

```{r}

p2 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("disthus_cut210_scale [all]"),
           grid = TRUE)
# p2 + theme_sjplot()

p3 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("treartgruppe"),
           grid = TRUE)
# p3 + theme_sjplot()

p4 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("kant"))
# p4 + theme_sjplot()

p5 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("helling_scale [all]"))
# p5 + theme_sjplot()

p6 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("HOH_scale [all]"))
# p6 + theme_sjplot()

# p6a <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("HOH_scale^2 [all]"))
# p6a + theme_sjplot()

p7 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("tretetthet_scale [all]"))
p7 + theme_sjplot()

p8 <- plot_model(m4glmm, type = c("eff"), transform = NULL, terms = c("distvei1_scale [all]", "skogkategori"))
p8 + theme_sjplot()


library(gridExtra)
plot_grid(list(p2, p3, p4, p5, p6, p7, p8))


# library(effects)
# plot(allEffects(m4glmm))
```


Code for plot:
plot_model(model, type = c("est", "re", "eff", "pred", "int", "std",
  "std2", "slope", "resid", "diag"), transform, terms = NULL,
  sort.est = NULL, rm.terms = NULL, group.terms = NULL,
  order.terms = NULL, pred.type = c("fe", "re"),
  mdrt.values = c("minmax", "meansd", "zeromax", "quart", "all"),
  ri.nr = NULL, title = NULL, axis.title = NULL,
  axis.labels = NULL, legend.title = NULL, wrap.title = 50,
  wrap.labels = 25, axis.lim = NULL, grid.breaks = NULL,
  ci.lvl = NULL, se = NULL, colors = "Set1",
  show.intercept = FALSE, show.values = TRUE, show.p = TRUE,
  show.data = FALSE, show.legend = TRUE, show.zeroinf = TRUE,
  value.offset = NULL, value.size, jitter = NULL, digits = 2,
  dot.size = NULL, line.size = NULL, vline.color = NULL,
  p.threshold = c(0.05, 0.01, 0.001), grid, case, auto.label = TRUE,
  prefix.labels = c("none", "varname", "label"), bpe = "median",
  bpe.style = "line", bpe.color = "white", ...)






#### Instead of having the second part of a hurdle or zeroinfl model, use a beta regression, because we have percentage data (use a subset of the dataset with only response values above 0) - need to account for spatial autocorrelation
```{r}

m1beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m1abeta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0) # nbs (in ac) = 5000

m1bbeta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m2beta <- betareg(bt100 ~ distvei1 + ac,
            data = browse_9_dist,
            subset = beitetrykkr > 0) 

m3beta <- betareg(bt100 ~ distvei1 + disthus+ ac,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m3abeta <- betareg(bt100 ~ disthus+ ac,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m4beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + ac,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m5beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m6beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m7beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m8beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling + HOH,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m9beta <- betareg(bt100 ~ distvei1 + disthus + skogkategori + treartgruppe + ac + kant + helling + HOH + tretetthet,
            data = browse_9_dist,
            subset = beitetrykkr > 0)


m10beta <- betareg(bt100 ~ distvei1 + skogkategori + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m11beta <- betareg(bt100 ~ disthus + skogkategori + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m12beta <- betareg(bt100 ~ skogkategori + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m13beta <- betareg(bt100 ~ distvei1 + disthus + treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m14beta <- betareg(bt100 ~ treartgruppe + ac + region + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m15beta <- betareg(bt100 ~ distvei1 + disthus + treartgruppe + ac + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

m16beta <- betareg(bt100 ~ treartgruppe + ac + KOMNR + kant + helling + HOH + tretetthet + skogkategori*distvei1,
            data = browse_9_dist,
            subset = beitetrykkr > 0)

```


Checking for the cut for disthus
```{r}
library(glmmTMB)
browse_9_sub <- browse_9[browse_9$bt100 > 0, ]

m1aglmmbeta <- glmmTMB(bt100 ~ disthus_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1bglmmbeta <- glmmTMB(bt100 ~ disthus_cut500 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1cglmmbeta <- glmmTMB(bt100 ~ disthus_cut450 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1dglmmbeta <- glmmTMB(bt100 ~ disthus_cut400 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1eglmmbeta <- glmmTMB(bt100 ~ disthus_cut350 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1fglmmbeta <- glmmTMB(bt100 ~ disthus_cut300 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())

m1gglmmbeta <- glmmTMB(bt100 ~ disthus_cut250 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family()) #7

m1hglmmbeta <- glmmTMB(bt100 ~ disthus_cut200 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family()) #8

m1iglmmbeta <- glmmTMB(bt100 ~ disthus_cut150 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family()) #9

m1jglmmbeta <- glmmTMB(bt100 ~ disthus_cut220 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family()) #10

m1kglmmbeta <- glmmTMB(bt100 ~ disthus_cut210 + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())#11

m1lglmmbeta <- glmmTMB(bt100 ~ log(disthus) + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
            data = browse_9_sub,
            family = beta_family())
```

```{r}
AIC(m1aglmmbeta, m1bglmmbeta, m1cglmmbeta, m1dglmmbeta, m1eglmmbeta, m1fglmmbeta, m1gglmmbeta, m1hglmmbeta, m1iglmmbeta, m1jglmmbeta, m1kglmmbeta)


# disthus_cut210 is the thing!

```


```{r}
pt12 <- plot_model(m1aglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_scale [all]"),
           grid = TRUE)
pt13 <- plot_model(m1bglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut500 [all]"),
           grid = TRUE)
pt14 <- plot_model(m1cglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut450 [all]"),
           grid = TRUE)
pt15 <- plot_model(m1dglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut400 [all]"),
           grid = TRUE)
pt16 <- plot_model(m1eglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut350 [all]"),
           grid = TRUE)
pt17 <- plot_model(m1fglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut300 [all]"),
           grid = TRUE)
pt18 <- plot_model(m1gglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut250 [all]"),
           grid = TRUE)
pt19 <- plot_model(m1hglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut200 [all]"),
           grid = TRUE)
pt20 <- plot_model(m1iglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut150 [all]"),
           grid = TRUE)
pt21 <- plot_model(m1jglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut220 [all]"),
           grid = TRUE)
pt22 <- plot_model(m1kglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut210 [all]"),
           grid = TRUE)

pt23 <- plot_model(m1lglmmbeta, type = c("eff"), transform = NULL, terms = c("disthus [all]"),
           grid = TRUE)


library(gridExtra)
plot_grid(list(pt12, pt23, pt13, pt14, pt15, pt16, pt17, pt18, pt19, pt20, pt21, pt22))

```

```{r}
browse_9_distsub <- browse_9_dist[browse_9_dist$bt100 > 0, ]
load("processdata/browse_9_distbeta.rda")
browse_9_distbeta_sub <- browse_9_distbeta[browse_9_distbeta$bt100 > 0, ]
browse_9_sub <- browse_9[browse_9$bt100 > 0, ]

library(glmmTMB)

m1glmmbeta <- glmmTMB(bt100 ~ disthus_cut210_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
                    data = browse_9_distsub,
                    family = beta_family())

m1glmmbetatest <- glmmTMB(bt100 ~ disthus_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
                    data = browse_9_distbeta_sub,
                    family = beta_family())

m1glmmbetatest2 <- glmmTMB(bt100 ~ disthus_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + skogkategori*distvei1_scale + (1|KOMNR),
                    data = browse_9_sub,
                    family = beta_family())

m2glmmbeta <- glmmTMB(bt100 ~ disthus_cut210_scale + distvei1_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + (1|KOMNR),
                    data = browse_9_distsub,
                    family = beta_family())

m3glmmbeta <- glmmTMB(bt100 ~ disthus_cut210_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + distvei1_scale*skogkategori + (1|KOMNR),
                    data = browse_9_distsub,
                    family = beta_family())


m4glmmbeta <- glmmTMB(bt100 ~ disthus_cut210_scale + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet_scale + distvei1_scale*skogkategori + (1|KOMNR),
                    data = browse_9_distsub,
                    family = beta_family())




```



```{r}

AIC(m1beta, m1abeta, m1bbeta, m2beta, m3beta, m4beta, m5beta, m6beta, m7beta, m8beta, m9beta)
# m1bbeta and m13beta lowest AIC, but also m14 is really the same...
AIC(m1bbeta, m13beta, m14beta, m15beta)

AIC(m1glmmbeta, m2glmmbeta, m3glmmbeta, m4glmmbeta) # m4

```

```{r}
resid_m4glmmbeta <- resid(m4glmmbeta, type = "pearson")
plot(resid_m4glmmbeta)

overdisp_fun <- function(model) {
    rdf <- df.residual(model)
    rp <- residuals(model,type="pearson")
    Pearson.chisq <- sum(rp^2)
    prat <- Pearson.chisq/rdf
    pval <- pchisq(Pearson.chisq, df=rdf, lower.tail=FALSE)
    c(chisq=Pearson.chisq,ratio=prat,rdf=rdf,p=pval)
}

overdisp_fun(m4glmmbeta)
# Have overdispersion, but he (Ben Bolker) says: "Remember that (1) overdispersion is irrelevant for models that estimate a scale parameter (i.e. almost anything but Poisson or binomial: Gaussian, Gamma, negative binomial …) and (2) overdispersion is not estimable (and hence practically irrelevant) for Bernoulli models (= binary data = binomial with N=1)."
```


```{r}
library(sjPlot)

tab_model(m4glmmbeta, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE)


```

```{r}
p1_est <- plot_model(m4glmmbeta, type = c("est"), transform = NULL, show.values = TRUE, value.offset = 0.5, 
                 vline.color = "grey", colors = c("#004949", "#009292"),
                 axis.title = "Estimates")
p1_est + theme_sjplot() + font_size(axis_title.x = 15)
# whiskers are ci, possible to show se instead; Colourbling friendly colours "#004949","#009292",
```



```{r}
pb1 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("disthus_cut210_scale [all]"))
# pb1 + theme_sjplot()

# pb2 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("skogkategori"))
# pb2 + theme_sjplot()

pb3 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("treartgruppe"),
                  show.p = TRUE)
pb3 + theme_sjplot()

pb4 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("kant"))
# pb4 + theme_sjplot()

pb5 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("helling_scale [all]"))
# pb5 + theme_sjplot()

pb6 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("HOH_scale [all]"))
# pb6 + theme_sjplot()

pb7 <- plot_model(m4glmmbeta, type = c("eff"), transform = NULL, terms = c("tretetthet_scale [all]"))
# pb7 + theme_sjplot()

pb8 <- plot_model(m4glmmbeta, type = c("int"), transform = NULL)
pb8 + theme_sjplot()

plot_grid(list(pb1, pb3, pb4, pb5, pb6, pb7, pb8))


# type = "pred" = Predicted values (marginal effects) for specific model terms. See ggpredict for details.
# type = "eff" = Similar to type = "pred", however, discrete predictors are held constant at their proportions (not reference level). See ggeffect for details. ggeffect() computes marginal effects by internally calling Effect. Effect constructs an "eff" object for a term (usually a high-order term), absorbing the lower-order terms marginal to the term in question, and averaging over other terms in the model.

```


```{r}
# devtools::install_github("glmmTMB/glmmTMB/glmmTMB",ref="effects")
# library(glmmTMB)
# source(system.file("other_methods","effectsglmmTMB.R",package="glmmTMB"))
# plot(allEffects(m3glmmbeta))
```


```{r}
# summary(m1beta)
# summary(m1bbeta)
# vcov(m1abeta)
# 
# lrtest(m1beta, m1abeta)
# 
# plot(m1beta)
# plot(m1abeta)
# 
# visreg(m1abeta, type = "conditional", scale = "response")
# visreg(m1abeta, type = "contrast", scale = "response")
# 
# resid_m1beta <- resid(m1beta, type = "pearson")
# resid_m1abeta <- resid(m1abeta, type = "pearson")
# 
# predbetareg <- cbind(
# predict(m1abeta, type = "response"), # fitted means of response
# predict(m1abeta, type = "link"), # corresponding linear predictor
# predict(m1abeta, type = "variance"), # fitted variances of response
# predict(m1abeta, type = "quantile", at = c(0.25, 0.5, 0.75)) # r fitted quantile(s) of the response distribution
# )
# colnames(predbetareg) <- c("response", "link", "variance", "qu0.25", "qu0.5", "qu0.75")
# 
# head(predbetareg)
# 
# broom::tidy(m1bbeta)
# broom::augment(m1bbeta)
# 
# browse9bro_dist_sub <- browse9bro_dist[browse9bro_dist$bt100 != 0, ]
# 
# browse9bro_dist_sub$model <- predict(m1bbeta)
# ggplot(data = browse9bro_dist_sub, aes(x = model, y = bt100)) + 
#    geom_point() + geom_abline(slope = 1)
# 
# plot_model(model = m1bbeta, type = "eff", terms = c("distvei1", "disthus"))
# get_model_data(m1bbeta)
# ggpredict(m1bbeta, terms = "distvei1")

```


## New package glmmTMB, for glmm and extentions (zero-inlfated with random effects, hurdle...)
- hurdle model does not work with beta family... But he says in his FAQ doc: "If you think that zero values are generated by a separate process, the simplest solution is to fit a Bernoulli model to the zero/non-zero data, then a conditional continuous model for the non-zero values; this is effectively a hurdle model."
```{r}
m1ziglmm <- glmmTMB(bt100 ~ disthus + (1|KOMNR),
                    data = browse_9_dist,
                    subset(), # make a datasubset...
                    ziformula = ~ 0,
                    family = beta_family())



# Hurdle Poisson model
glmmTMBhurdle1 <- glmmTMB(beitetrykkr ~ treartgruppe + ac + kant + helling + HOH + tretetthet + skogkategori*distvei1 + (1|KOMNR),
                          zi = ~ disthus + treartgruppe + ac + kant + helling + HOH + I(HOH^2) + tretetthet + skogkategori*distvei1,
                          family = truncated_poisson, 
                          data = browse_9_dist)
# NA/NaN function evaluationModel convergence problem; non-positive-definite Hessian matrix. See vignette('troubleshooting')Model convergence problem; function evaluation limit reached without convergence (9). See vignette('troubleshooting')

glmmTMBhurdle2 <- glmmTMB(beitetrykkr ~ distvei1 + disthus_cut210 + (1|KOMNR),
                          zi = ~ distvei1 + disthus_cut210,
                          family = truncated_poisson, 
                          data = browse_9_dist) # works


glmmTMBhurdle2a <- glmmTMB(beitetrykkr ~ distvei1 + disthus_cut210 + (1|KOMNR),
                          zi = ~ distvei1 + disthus_cut210,
                          family = beta_family(), 
                          data = browse_9_dist) # doesn't work: Error in eval(family$initialize) : y values must be 0 < y < 1

```
