==> 05_browsing_hurdle.Rmd <==
---
title: "05_browsing_pressure_new"
author: "anne"
date: "1 februar 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



########################################################################
## Generell test for zero inflation for browsing response variable ##
#######################################################################
```{r}
load("processdata/browse.rda")

library(ggplot2)
library(ggpubr)

ggplot(browse, aes(beitetrykk9)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

ggplot(browse, aes(bt100)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

str(browse$beitetrykk9)
summary(browse$beitetrykk9)



#### Test for zero-inflation

# Use glmmTMB package to run glm
library("glmmTMB")
m_test_zi <- glm(beitetrykk9 ~ disthus + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac, family = poisson, data = browse)

library("performance")
check_zeroinflation(m_test_zi)

# # Check for zero-inflation
# 
#    Observed zeros: 5681
#   Predicted zeros: 1
#             Ratio: 0.00
# 
# Model is underfitting zeros (probable zero-inflation).



```





###################################################################################
##     HOUSE                HOUSE                 HOUSE                       #####
###################################################################################

##       House binary - First part of hurdle model to account for zero inflation  ##


## Bin house threshold distance to house

#### Load data house models
```{r}
load("processdata/browse_2021_02_01.rda")
names(browse)

browse_house <- browse

names(browse_house)

```


#### Scale variables if necessary
```{r}

browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_600 <- scale(browse_house$disthus_600)
browse_house$disthus_500 <- scale(browse_house$disthus_500)
browse_house$disthus_400 <- scale(browse_house$disthus_400)
browse_house$disthus_350 <- scale(browse_house$disthus_350)
browse_house$disthus_300 <- scale(browse_house$disthus_300)

browse_house$disthus_200 <- scale(browse_house$disthus_200)

browse_house$disthus_150 <- scale(browse_house$disthus_150)

summary(browse_house$disthus_200)
```


#### Checking for the cut for disthus
```{r}
library(lme4)


bin_cuthouse1 <- glmer(factor(beitetrykk9 > 0) ~ disthus + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse2 <- glmer(factor(beitetrykk9 > 0) ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse3 <- glmer(factor(beitetrykk9 > 0) ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cuthouse4 <- glmer(factor(beitetrykk9 > 0) ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse5 <- glmer(factor(beitetrykk9 > 0) ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse6 <- glmer(factor(beitetrykk9 > 0) ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse7 <- glmer(factor(beitetrykk9 > 0) ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # BEST

bin_cuthouse8 <- glmer(factor(beitetrykk9 > 0) ~ disthus_150 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cuthouse1, bin_cuthouse2, bin_cuthouse3, bin_cuthouse4, bin_cuthouse5, bin_cuthouse6, bin_cuthouse7, bin_cuthouse8)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
#
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod7 19 21508.86      0.00  0.66   0.66 -10735.43
# Mod8 19 21510.71      1.84  0.26   0.92 -10736.35
# Mod6 19 21513.22      4.35  0.07   1.00 -10737.61
# Mod5 19 21521.55     12.69  0.00   1.00 -10741.77
# Mod4 19 21528.51     19.65  0.00   1.00 -10745.26
# Mod3 19 21539.56     30.70  0.00   1.00 -10750.78
# Mod2 19 21550.45     41.59  0.00   1.00 -10756.23
# Mod1 19 21581.44     72.57  0.00   1.00 -10771.72

```

#### Save browse_house dataset
```{r}
names(browse_house)
browse_house <- browse_house[, -c(16:20,22)]
names(browse_house)
browse_house$beite_bin <- ifelse(browse_house$beitetrykk9 == 0, 0, 1)

str(browse_house)
save(browse_house, file = "browse_house_2021_02_01.rda")



```





## Bin house models

#### Load data
```{r}
load("processdata/browse_house_2021_02_01.rda")
str(browse_house)

```


#### Scale variables if necessary
```{r}

# browse_house$disthus <- scale(browse_house$disthus)
# browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_200 <- scale(browse_house$disthus_200)

summary(browse_house$disthus_200)

```



#### Bin house models
```{r}
library(lme4)
browse_h_bin1 <- glmer(beite_bin ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # BEST
summary(browse_h_bin1)

browse_h_bin2 <- glmer(beite_bin ~ disthus_200 + log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 






```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_browse_house_bin <- list(browse_h_bin1, browse_h_bin2)
aictab(candmods_browse_house_bin, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 19 21483.80      0.00  0.98   0.98 -10722.90
# Mod2 15 21491.83      8.03  0.02   1.00 -10730.92


```

#### Summary bin house
```{r}
summary(browse_h_bin1)
# vcov(browse_h_bin5)
```



#### Plots bin house
```{r}
# library(effects)
# 
# plot(allEffects(rbrowse_h_bin)5)
# plot(effect("skogkategori", browse_h_bin5))


library(sjPlot)
plot_model(browse_h_bin1, type = "diag")
plot_model(browse_h_bin1, type = "eff")
plot_model(browse_h_bin1, type = "est")

# library(performance)
# # check_model(browse_h_bin1)
# model_performance(browse_h_bin1)

 
tab_model(browse_h_bin1, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

# ggstatsplot::ggcoefstats(
#   x = browse_h_bin1,
#   title = "generalized linear mixed-effects model"
# )

```

#### Save bin house
```{r}
save(browse_h_bin1, file = "browse_h_bin1.rda")

```






##       House beta - Second part of the hurdle, Beta  ##


## Load data house models
```{r}
load("processdata/browse_house.rda")


str(browse_house)
names(browse_house)
```


#### Prepare data set for second part of the hurdle
```{r}
browse_house <- browse_house[browse_house$bt100 > 0, ]
str(browse_house)
summary(browse_house$bt100)

```


#### Scale variables if necessary
```{r}
browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_600 <- scale(browse_house$disthus_600)
browse_house$disthus_500 <- scale(browse_house$disthus_500)
browse_house$disthus_400 <- scale(browse_house$disthus_400)
browse_house$disthus_350 <- scale(browse_house$disthus_350)
browse_house$disthus_300 <- scale(browse_house$disthus_300)
browse_house$disthus_200 <- scale(browse_house$disthus_200)
browse_house$disthus_150 <- scale(browse_house$disthus_150)

summary(browse_house$disthus_600)
```

#### Checking for the cut for disthus
```{r}
library(glmmTMB)

beta_cuthouse1 <- glmmTMB(bt100 ~ disthus + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse2 <- glmmTMB(bt100 ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse3 <- glmmTMB(bt100 ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse4 <- glmmTMB(bt100 ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse5 <- glmmTMB(bt100 ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse6 <- glmmTMB(bt100 ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse7 <- glmmTMB(bt100 ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST

beta_cuthouse8 <- glmmTMB(bt100 ~ disthus_150 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3, beta_cuthouse4, beta_cuthouse5, beta_cuthouse6, beta_cuthouse7, beta_cuthouse8)
# candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod7 20 -9735.49      0.00  0.92   0.92 4887.75
# Mod8 20 -9729.17      6.32  0.04   0.95 4884.59
# Mod5 20 -9729.17      6.33  0.04   0.99 4884.58
# Mod6 20 -9724.93     10.57  0.00   1.00 4882.46
# Mod4 20 -9722.46     13.03  0.00   1.00 4881.23
# Mod3 20 -9719.30     16.19  0.00   1.00 4879.65
# Mod2 20 -9714.78     20.72  0.00   1.00 4877.39
# Mod1 20 -9702.39     33.10  0.00   1.00 4871.20

```


#### Save browse_house dataset
```{r}
## !! Use browse_house - all variables needed are in there !! ##


# names(browse_house)
# browse_house <- browse_house[, -c(4,18:23)]
# str(browse_house)
# save(browse_house, file = "browse_house.rda")

```







## Beta house models

Prepare data set for second part of the hurdle
```{r}
load("processdata/browse_house.rda")
browse_house <- browse_house[browse_house$bt100 > 0, ]
str(browse_house)
summary(browse_house$bt100)

```

#### Scale variables if necessary
```{r}

# browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_200 <- scale(browse_house$disthus_200)


summary(browse_house$disthus_200)
```


#### Beta house models
```{r}
library(glmmTMB)

browse_h_beta1 <- glmmTMB(bt100 ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())
summary(browse_h_beta1)

browse_h_beta2 <- glmmTMB(bt100 ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST



```


#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_browse_house_beta <- list(browse_h_beta1,browse_h_beta2)
aictab(candmods_browse_house_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod2 19 -9764.46      0.00  0.59   0.59 4901.23
# Mod1 20 -9763.74      0.72  0.41   1.00 4901.87

```


#### Summary beta house
```{r}
summary(browse_h_beta2)
# vcov(rec_h_bin5)
```



#### Plots beta house
```{r}
# library(effects)
# plot(allEffects(rec_h_beta7))
# plot(effect("skogkategori", rec_h_beta7))

library(sjPlot)
plot_model(browse_h_beta2, type = "diag")
plot_model(browse_h_beta2, type = "eff")
plot_model(browse_h_beta2, type = "est")

# library(performance)
# # check_model(browse_h_beta2)
# model_performance(browse_h_beta2)



tab_model(browse_h_beta2, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))




# ggstatsplot::ggcoefstats(
#   x = browse_h_beta2,
#   title = "generalized linear mixed-effects model"
# ) # Not sure what exactly this shows.

```


#### Save house beta model
```{r}
save(browse_h_beta2, file = "browse_h_beta2.rda")
```





###################################################################################
##     ROADS                ROADS                 ROADS                       #####
###################################################################################

##       Roads Bin - First part of a hurdle model to account for zero inflation  ##


## Bin roads 

#### Load data roads models
```{r}
load("processdata/browse_roads_2021_02_02.rda")
str(browse_roads)
names(browse_roads)


```



#### Scale variables if necessary
```{r}
browse_roads$beite_bin <- ifelse(browse_roads$beitetrykk9 == 0, 0, 1)


# browse_roads$distvei2 <- scale(browse_roads$distvei2)

browse_roads$tretetthet9 <- scale(browse_roads$tretetthet9)
browse_roads$helling <- scale(browse_roads$helling)
browse_roads$HOH <- scale(browse_roads$HOH)
# browse_roads$moose_density <- scale(browse_roads$moose_density)

str(browse_roads)
summary(browse_roads$tretetthet9)
```



## Bin roads models
```{r}
library(lme4)
browse_r_bin1 <- glmer(beite_bin ~ log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_roads) # BEST
summary(browse_r_bin1) 

browse_r_bin1_log <- glmer(beite_bin ~ distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_roads)

browse_r_bin2 <- glmer(beite_bin ~ log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_roads) 

browse_r_bin3 <- glmer(beite_bin ~ log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_roads)

browse_r_bin4 <- glmer(beite_bin ~ log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling  + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_roads)


```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_browse_roads_bin <- list(browse_r_bin1, browse_r_bin1_log, browse_r_bin2, browse_r_bin3, browse_r_bin4)
aictab(candmods_browse_roads_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt       LL
# Mod1 18 19051.56      0.00  0.95   0.95 -9507.78
# Mod3 14 19057.74      6.19  0.04   1.00 -9514.87
# Mod2 18 19062.89     11.34  0.00   1.00 -9513.45
# Mod4 16 19079.85     28.29  0.00   1.00 -9523.93
# Mod5 12 19091.59     40.04  0.00   1.00 -9533.80



```

#### Summary bin roads
```{r}
summary(browse_r_bin1)
# vcov(rec_h_bin5)
```



#### Plots bin roads
```{r}
# library(effects)
# 
# plot(allEffects(rec_r_bin7))
# plot(effect("skogkategori", rec_r_bin7))

# library(performance)
# # check_model(browse_r_bin1)
# model_performance(browse_r_bin1)

library(sjPlot)
plot_model(browse_r_bin1, type = "diag")
plot_model(browse_r_bin1, type = "eff")
plot_model(browse_r_bin1, type = "est")

tab_model(browse_r_bin1, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

# ggstatsplot::ggcoefstats(
#   x = browse_r_bin1,
#   title = "generalized linear mixed-effects model"
# )

```

#### Save bin roads
```{r}
save(browse_r_bin1, file = "output/browsing_2021_02_02/browse_r_bin1.rda")

```





##       Roads beta - Second part of the hurdle, beta regression  ##


## Load data road models
```{r}
load("processdata/browse_roads.rda")
str(browse_roads)
names(browse_roads)
```


#### Prepare data set for second part of the hurdle
```{r}
browse_roads <- browse_roads[browse_roads$bt100 > 0, ]
str(browse_roads)
summary(browse_roads$bt100)



```


#### Scale variables if necessary
```{r}

# recruits_roads$distvei2 <- scale(recruits_roads$distvei2)

browse_roads$tretetthet9 <- scale(browse_roads$tretetthet9)
browse_roads$helling <- scale(browse_roads$helling)
browse_roads$HOH <- scale(browse_roads$HOH)
browse_roads$moose_density <- scale(browse_roads$moose_density)


```


## Beta roads models
```{r}
library(glmmTMB)
browse_r_beta1 <- glmmTMB(bt100 ~ log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_roads,
            family = beta_family())
summary(browse_r_beta1)

browse_r_beta2 <- glmmTMB(bt100 ~ log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
            data = browse_roads,
            family = beta_family()) # BEST

browse_r_beta3 <- glmmTMB(bt100 ~ distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + ac + (1|KOMNR),
            data = browse_roads,
            family = beta_family()) 


```




#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_browse_roads_beta <- list(browse_r_beta1, browse_r_beta2, browse_r_beta3)
aictab(candmods_browse_roads_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod2 18 -8794.57      0.00  0.63   0.63 4415.28
# Mod1 19 -8793.49      1.08  0.37   1.00 4415.75
# Mod3 18 -8774.53     20.04  0.00   1.00 4405.26

```


#### Summary beta roads
```{r}
summary(browse_r_beta2)
# vcov(rec_h_bin5)
```



#### Plots beta roads
```{r}
# library(effects)
# 
# plot(allEffects(rec_r_beta4))
# plot(effect("skogkategori", rec_r_beta4))

# library(performance)
# # check_model(browse_r_beta3)
# model_performance(browse_r_beta3)

library(sjPlot)
plot_model(browse_r_beta2, type = "diag")
plot_model(browse_r_beta2, type = "eff") 
plot_model(browse_r_beta2, type = "est")

tab_model(browse_r_beta2, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

# ggstatsplot::ggcoefstats(
#   x = browse_r_beta3,
#   title = "generalized linear mixed-effects model"
# )

```


#### Save roads beta model
```{r}
save(browse_r_beta2, file = "browse_r_beta2.rda")
```





## Figures bin
```{r}
library(sjPlot)
library(ggplot2)

## House
load("processdata/browse_house_2021_02_01.rda")
str(browse_house) # Check if variables are scales - if not scale
load("output/browsing_2021_02_02/browse_h_bin1.rda")
browse_house$beite_bin <- ifelse(browse_house$beitetrykk9 == 0, 0, 1)
summary(browse_h_bin1)

# # Scale
browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
# browse_house$moose_density <- scale(browse_house$moose_density)
browse_house$disthus_200 <- scale(browse_house$disthus_200)




p_browse_bin1_house <- plot_model(browse_h_bin1, type = c("eff"), transform = NULL, terms = c("disthus_200"))
p_browse_bin1_house <- p_browse_bin1_house + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Browsing pressure\n")



## Roads
load("processdata/browse_roads_2021_02_02.rda")
str(browse_roads) # Check if variables are scales - if not scale
load("output/browsing_2021_02_02/browse_r_bin1.rda")
summary(browse_r_bin1)
browse_roads$beite_bin <- ifelse(browse_roads$beitetrykk9 == 0, 0, 1)

# Scale
browse_roads$tretetthet9 <- scale(browse_roads$tretetthet9)
browse_roads$helling <- scale(browse_roads$helling)
browse_roads$HOH <- scale(browse_roads$HOH)
# browse_roads$moose_density <- scale(browse_roads$moose_density)



p_browse_bin1 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("distvei2  [0:500]"))
p_browse_bin1 <- p_browse_bin1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to road\n") +
  ylab("Browsing pressure\n")


p_browse_bin2 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("skogkategori"))
p_browse_bin2 <- p_browse_bin2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30)) +
  xlab("Forest category\n") +
  ylab("Browsing pressure\n") 


p_browse_bin3 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p_browse_bin3 <- p_browse_bin3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Browsing pressure\n") 


p_browse_bin4 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("kant"))
p_browse_bin4 <- p_browse_bin4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30)) +
  xlab("Edge\n") +
  ylab("Browsing pressure\n")


p_browse_bin5 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("helling [all]"))
p_browse_bin5 <- p_browse_bin5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Browsing pressure\n")


p_browse_bin6 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
p_browse_bin6 <- p_browse_bin6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Browsing pressure\n")


p_browse_bin7 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("moose_density [all]"))
p_browse_bin7 <- p_browse_bin7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Moose density\n") +
  ylab("Browsing pressure\n")


p_browse_bin8 <- plot_model(browse_r_bin1, type = c("eff"), transform = NULL, terms = c("tretetthet9  [all]"))
p_browse_bin8 <- p_browse_bin8 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree density\n") +
  ylab("Browsing pressure\n")



library(gridExtra)
browse_eff_plot_1_bin <- grid.arrange(p_browse_bin1_house, p_browse_bin1, p_browse_bin2, p_browse_bin3,
          nrow = 2, ncol = 2)

browse_eff_plot_2_bin <- grid.arrange(p_browse_bin8, p_browse_bin4, p_browse_bin5, p_browse_bin6, p_browse_bin7,
          nrow = 3, ncol = 2)

# type = "pred" = Predicted values (marginal effects) for specific model terms. See ggpredict for details.
# type = "eff" = Similar to type = "pred", however, discrete predictors are held constant at their proportions (not reference level). See ggeffect for details. ggeffect() computes marginal effects by internally calling Effect. Effect constructs an "eff" object for a term (usually a high-order term), absorbing the lower-order terms marginal to the term in question, and averaging over other terms in the model.

ggsave(file = "browse_eff_plot_1_bin.tiff", browse_eff_plot_1_bin, width = 30, height = 30, units = "cm", dpi = 150) #, dpi = 150
ggsave(file = "browse_eff_plot_2_bin.tiff", browse_eff_plot_2_bin, width = 30, height = 30, units = "cm", dpi = 150) #, dpi = 150

```



## Figures beta
```{r}
library(sjPlot)
library(ggplot2)

## House
load("processdata/browse_house.rda")
str(browse_house) # Check if variables are scales - if not scale
load("output/browsing_2021_02_02/browse_h_beta2.rda")
browse_house <- browse_house[browse_house$bt100 > 0, ]
summary(browse_h_beta2)

# Scale
browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)
browse_house$disthus_200 <- scale(browse_house$disthus_200)



p_browse_beta1_house <- plot_model(browse_h_beta2, type = c("eff"), transform = NULL, terms = c("disthus_200"))
p_browse_beta1_house <- p_browse_beta1_house + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Browsing pressure\n")



## Roads
load("processdata/browse_roads.rda")
str(browse_roads) # Check if variables are scales - if not scale
load("output/browsing_2021_02_02/browse_r_beta2.rda")
summary(browse_r_beta2)
browse_roads <- browse_roads[browse_roads$bt100 > 0, ]

# Scale
browse_roads$tretetthet9 <- scale(browse_roads$tretetthet9)
browse_roads$helling <- scale(browse_roads$helling)
browse_roads$HOH <- scale(browse_roads$HOH)
browse_roads$moose_density <- scale(browse_roads$moose_density)




p_browse_beta1 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("distvei2  [0:500]"))
p_browse_beta1 <- p_browse_beta1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to road\n") +
  ylab("Browsing pressure\n")


p_browse_beta2 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("skogkategori"))
p_browse_beta2 <- p_browse_beta2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30)) +
  xlab("Forest category\n") +
  ylab("Browsing pressure\n") 


p_browse_beta3 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p_browse_beta3 <- p_browse_beta3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Browsing pressure\n") 


p_browse_beta4 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("kant"))
p_browse_beta4 <- p_browse_beta4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30)) +
  xlab("Edge\n") +
  ylab("Browsing pressure\n")


p_browse_beta5 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("helling [all]"))
p_browse_beta5 <- p_browse_beta5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Browsing pressure\n")


p_browse_beta6 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
p_browse_beta6 <- p_browse_beta6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Browsing pressure\n")


p_browse_beta7 <- plot_model(browse_r_beta2, type = c("eff"), transform = NULL, terms = c("tretetthet9 [all]"))
p_browse_beta7 <- p_browse_beta7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree density\n") +
  ylab("Browsing pressure\n")



library(gridExtra)
grid.arrange(p_browse_beta1_house, p_browse_beta1, p_browse_beta2, p_browse_beta3, p_browse_beta7, p_browse_beta4, p_browse_beta5, p_browse_beta6,
          nrow = 4, ncol = 2)

# type = "pred" = Predicted values (marginal effects) for specific model terms. See ggpredict for details.
# type = "eff" = Similar to type = "pred", however, discrete predictors are held constant at their proportions (not reference level). See ggeffect for details. ggeffect() computes marginal effects by internally calling Effect. Effect constructs an "eff" object for a term (usually a high-order term), absorbing the lower-order terms marginal to the term in question, and averaging over other terms in the model.

```
