==> 05_browsing_hurdle_remake.Rmd <==
---
title: "05_browsing_hurdle_remake"
author: "Anne"
date: "12/04/2021"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



########################################################################
## Generell test for zero inflation for browsing response variable ##
#######################################################################
```{r}
load("processdata/browse.rda")

library(ggplot2)
library(ggpubr)

ggplot(browse, aes(beitetrykk9)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

ggplot(browse, aes(bt100)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

str(browse$beitetrykk9)
summary(browse$beitetrykk9)



#### Test for zero-inflation

# Use glmmTMB package to run glm
library("glmmTMB")
m_test_zi <- glm(beitetrykk9 ~ disthus + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac, family = poisson, data = browse)

library("performance")
check_zeroinflation(m_test_zi)

# # Check for zero-inflation
# 
#    Observed zeros: 5681
#   Predicted zeros: 1
#             Ratio: 0.00
# 
# Model is underfitting zeros (probable zero-inflation).



```



##       BINARY - First part of hurdle model to account for zero inflation  ##

## Bin threshold distance to house

#### Load data house threshold
```{r}
load("processdata/browse_2021_04_15.rda")
names(browse)

browse_house <- browse

names(browse_house)

```


###### Scale variables if necessary
```{r}

browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_600 <- scale(browse_house$disthus_600)
browse_house$disthus_500 <- scale(browse_house$disthus_500)
browse_house$disthus_400 <- scale(browse_house$disthus_400)
browse_house$disthus_350 <- scale(browse_house$disthus_350)
browse_house$disthus_300 <- scale(browse_house$disthus_300)

browse_house$disthus_200 <- scale(browse_house$disthus_200)

browse_house$disthus_150 <- scale(browse_house$disthus_150)

summary(browse_house$disthus_200)
```


###### Checking for the threshold for disthus
```{r}
library(lme4)


bin_cuthouse1 <- glmer(factor(beitetrykk9 > 0) ~ disthus + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse2 <- glmer(factor(beitetrykk9 > 0) ~ disthus_600 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse3 <- glmer(factor(beitetrykk9 > 0) ~ disthus_500 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cuthouse4 <- glmer(factor(beitetrykk9 > 0) ~ disthus_400 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse5 <- glmer(factor(beitetrykk9 > 0) ~ disthus_350 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cuthouse6 <- glmer(factor(beitetrykk9 > 0) ~ disthus_300 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # BEST

bin_cuthouse7 <- glmer(factor(beitetrykk9 > 0) ~ disthus_200 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cuthouse8 <- glmer(factor(beitetrykk9 > 0) ~ disthus_150 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

```

###### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cuthouse1, bin_cuthouse2, bin_cuthouse3, bin_cuthouse4, bin_cuthouse5, bin_cuthouse6, bin_cuthouse7, bin_cuthouse8)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
# Model selection based on AICc:
#
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# 200 19 21508.86      0.00  0.66   0.66 -10735.43
# 150 19 21510.71      1.84  0.26   0.92 -10736.35
# 300 19 21513.22      4.35  0.07   1.00 -10737.61
# 350 19 21521.55     12.69  0.00   1.00 -10741.77
# 400 19 21528.51     19.65  0.00   1.00 -10745.26
# 500 19 21539.56     30.70  0.00   1.00 -10750.78
# 600 19 21550.45     41.59  0.00   1.00 -10756.23
# full 19 21581.44     72.57  0.00   1.00 -10771.72

```

###### Save browse_house dataset
```{r}
names(browse_house)
browse_house <- browse_house[, -c(16,18:20,22)]
names(browse_house)
browse_house$beite_bin <- ifelse(browse_house$beitetrykk9 == 0, 0, 1)

str(browse_house)
save(browse_house, file = "processdata/browse_house_2021_04_12.rda")



```





## Bin threshold distance to roads

#### Load data roads threshold
```{r}
load("processdata/browse_2021_04_15.rda")
names(browse)

browse_house <- browse

names(browse_house)

```


###### Scale variables if necessary
```{r}

browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$distvei_600 <- scale(browse_house$distvei_600)
browse_house$distvei_500 <- scale(browse_house$distvei_500)
browse_house$distvei_400 <- scale(browse_house$distvei_400)
browse_house$distvei_350 <- scale(browse_house$distvei_350)
browse_house$distvei_300 <- scale(browse_house$distvei_300)
browse_house$distvei_200 <- scale(browse_house$distvei_200)
browse_house$distvei_150 <- scale(browse_house$distvei_150)
browse_house$distvei_100 <- scale(browse_house$distvei_100)
browse_house$distvei_50 <- scale(browse_house$distvei_50)


summary(browse_house$distvei_200)

```


###### Checking for the threshold for distvei
```{r}
library(lme4)


bin_cutroad1 <- glmer(factor(beitetrykk9 > 0) ~ disthus + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cutroad2 <- glmer(factor(beitetrykk9 > 0) ~ distvei_600 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cutroad3 <- glmer(factor(beitetrykk9 > 0) ~ distvei_500 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cutroad4 <- glmer(factor(beitetrykk9 > 0) ~ distvei_400 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cutroad5 <- glmer(factor(beitetrykk9 > 0) ~ distvei_350 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cutroad6 <- glmer(factor(beitetrykk9 > 0) ~ distvei_300 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

bin_cutroad7 <- glmer(factor(beitetrykk9 > 0) ~ distvei_200 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cutroad8 <- glmer(factor(beitetrykk9 > 0) ~ distvei_150 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # BEST

bin_cutroad9 <- glmer(factor(beitetrykk9 > 0) ~ distvei_100 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

bin_cutroad10 <- glmer(factor(beitetrykk9 > 0) ~ distvei_50 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

```

###### AIC selection threshold road
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cutroad1, bin_cutroad2, bin_cutroad3, bin_cutroad4, bin_cutroad5, bin_cutroad6, bin_cutroad7, bin_cutroad8, bin_cutroad9, bin_cutroad10)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
#
#        K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod8  27 21409.97      0.00  0.98   0.98 -10677.99
# Mod7  27 21417.86      7.89  0.02   1.00 -10681.93
# Mod9  27 21427.20     17.23  0.00   1.00 -10686.60
# Mod5  27 21453.18     43.21  0.00   1.00 -10699.59
# Mod10 27 21456.79     46.82  0.00   1.00 -10701.39
# Mod4  27 21462.39     52.41  0.00   1.00 -10704.19
# Mod6  19 21471.36     61.39  0.00   1.00 -10716.68
# Mod3  27 21476.46     66.49  0.00   1.00 -10711.23
# Mod2  27 21490.41     80.44  0.00   1.00 -10718.21
# Mod1  27 21543.19    133.22  0.00   1.00 -10744.60

```


```{r}
browse_house$disthus_200 <- scale(browse_house$disthus_200)


distveilog <- glmer(factor(beitetrykk9 > 0) ~ disthus_200 + skogkategori + log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # Best

distveithresh <- glmer(factor(beitetrykk9 > 0) ~ disthus_200 + skogkategori + distvei_150 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)


candmods_bin_thresh <- list(distveilog, distveithresh)
aictab(candmods_bin_thresh, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 19 18511.50      0.00     1      1  -9236.75
# Mod2 19 21421.47   2909.98     0      1 -10691.74

```





## Bin models

#### Load data
```{r}
load("processdata/browse_house_2021_04_12.rda")
# load("processdata/browse_2021_04_15.rda")
str(browse_house)
names(browse_house)


```


#### Scale variables if necessary
```{r}

# browse_house$disthus <- scale(browse_house$disthus)
# browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

# browse_house$disthus_200 <- scale(browse_house$disthus_200)

# summary(browse_house$disthus_200)

```


#### Bin house models
```{r}
library(lme4)

browse_h_bin1 <- glmer(beite_bin ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 
summary(browse_h_bin1)


## test for interactions ----


browse_h_bin2 <- glmer(beite_bin ~ disthus_200*skogkategori + log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) 

browse_h_bin3 <- glmer(beite_bin ~ disthus_200*treartgruppe9 + skogkategori + log(distvei2+1) + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

browse_h_bin4 <- glmer(beite_bin ~ disthus_200*skogkategori + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)

browse_h_bin5 <- glmer(beite_bin ~ disthus_200 + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house) # BEST


browse_h_bin6 <- glmer(beite_bin ~ disthus_200 + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = browse_house)



```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_browse_house_bin <- list(browse_h_bin1, browse_h_bin2, browse_h_bin3, browse_h_bin4, browse_h_bin5, browse_h_bin6)
candmods_browse_house_bin <- list(browse_h_bin5, browse_h_bin6)
aictab(candmods_browse_house_bin, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 23 21451.17      0.00     1      1 -10702.59   # mod5
# Mod2 21 21483.14     31.96     0      1 -10720.57

# Interactions
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod5 23 21451.17      0.00  0.62   0.62 -10702.59   # vei*skog
# Mod4 27 21452.16      0.99  0.38   1.00 -10699.08
# Mod2 23 21470.01     18.84  0.00   1.00 -10712.01
# Mod1 19 21483.80     32.63  0.00   1.00 -10722.90
# Mod3 21 21486.88     35.71  0.00   1.00 -10722.44



save(browse_h_bin1, file = "output/2021_04_28/browse_h_bin1.rda")
save(browse_h_bin2, file = "output/2021_04_28/browse_h_bin2.rda")
save(browse_h_bin3, file = "output/2021_04_28/browse_h_bin3.rda")
save(browse_h_bin4, file = "output/2021_04_28/browse_h_bin4.rda")
save(browse_h_bin5, file = "output/2021_04_28/browse_h_bin5.rda")

```

#### Summary bin house
```{r}
summary(browse_h_bin5)

```



#### Plots bin house
```{r}
# library(effects)
# 
# plot(allEffects(rbrowse_h_bin)5)
# plot(effect("skogkategori", browse_h_bin5))


library(sjPlot)
plot_model(browse_h_bin5, type = "diag")
plot_model(browse_h_bin5, type = "eff")
plot_model(browse_h_bin5, type = "est")

tab_model(browse_h_bin5, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

plot_model(browse_h_bin5, type = "eff", terms = c("distvei2", "skogkategori"))
# plot_model(browse_h_bin5, type = "int")

pred_browse_h_bin5 <- get_model_data(browse_h_bin5, type = "eff")

```

#### Save bin house
```{r}
save(browse_h_bin5, file = "output/2021_04_28/browse_h_bin5.rda")

```








##       ZERO-TRUNCATED - Second part of the hurdle, Beta  ##

## Load data house threshold
```{r}
# load("processdata/browse_house.rda")
load("processdata/browse_house_2021_04_12.rda")
names(browse)

browse_house <- browse

names(browse_house)

```


#### Prepare data set for second part of the hurdle threshold
```{r}
browse_house <- browse_house[browse_house$bt100 > 0, ]
str(browse_house)
summary(browse_house$bt100)

```


#### Scale variables if necessary
```{r}
browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$disthus_600 <- scale(browse_house$disthus_600)
browse_house$disthus_500 <- scale(browse_house$disthus_500)
browse_house$disthus_400 <- scale(browse_house$disthus_400)
browse_house$disthus_350 <- scale(browse_house$disthus_350)
browse_house$disthus_300 <- scale(browse_house$disthus_300)
browse_house$disthus_200 <- scale(browse_house$disthus_200)
browse_house$disthus_150 <- scale(browse_house$disthus_150)

summary(browse_house$disthus_600)
```

#### Checking for the threshold for disthus
```{r}
library(glmmTMB)

beta_cuthouse1 <- glmmTMB(bt100 ~ disthus + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse2 <- glmmTMB(bt100 ~ disthus_600 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) 

beta_cuthouse3 <- glmmTMB(bt100 ~ disthus_500 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse4 <- glmmTMB(bt100 ~ disthus_400 + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse5 <- glmmTMB(bt100 ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse6 <- glmmTMB(bt100 ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cuthouse7 <- glmmTMB(bt100 ~ disthus_200 + skogkategori + log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST

beta_cuthouse8 <- glmmTMB(bt100 ~ disthus_150 + distvei2 + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3, beta_cuthouse4, beta_cuthouse5, beta_cuthouse6, beta_cuthouse7, beta_cuthouse8)
# candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# 200  20 -9735.49      0.00  0.92   0.92 4887.75
# 150  20 -9729.17      6.32  0.04   0.95 4884.59
# 300  20 -9729.17      6.33  0.04   0.99 4884.58
# 350  20 -9724.93     10.57  0.00   1.00 4882.46
# 400  20 -9722.46     13.03  0.00   1.00 4881.23
# 500  20 -9719.30     16.19  0.00   1.00 4879.65
# 600  20 -9714.78     20.72  0.00   1.00 4877.39
# full 20 -9702.39     33.10  0.00   1.00 4871.20

```





## Load data road threshold
```{r}
load("processdata/browse_2021_04_15.rda")
browse_house <- browse
str(browse_house)
names(browse_house)

```


#### Prepare data set for second part of the hurdle threshold
```{r}
browse_house <- browse_house[browse_house$bt100 > 0, ]
str(browse_house)
summary(browse_house$bt100)

```


#### Scale variables if necessary
```{r}
browse_house$disthus <- scale(browse_house$disthus)
browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

browse_house$distvei_600 <- scale(browse_house$distvei_600)
browse_house$distvei_500 <- scale(browse_house$distvei_500)
browse_house$distvei_400 <- scale(browse_house$distvei_400)
browse_house$distvei_350 <- scale(browse_house$distvei_350)
browse_house$distvei_300 <- scale(browse_house$distvei_300)
browse_house$distvei_200 <- scale(browse_house$distvei_200)
browse_house$distvei_150 <- scale(browse_house$distvei_150)
browse_house$distvei_100 <- scale(browse_house$distvei_100)
browse_house$distvei_50 <- scale(browse_house$distvei_50)

summary(browse_house$disthus_600)
```

#### Checking for the threshold for distvei
```{r}
library(glmmTMB)

beta_cutroad1 <- glmmTMB(bt100 ~ disthus + skogkategori + distvei2 + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) 

beta_cutroad2 <- glmmTMB(bt100 ~ distvei_600 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cutroad3 <- glmmTMB(bt100 ~ distvei_500 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cutroad4 <- glmmTMB(bt100 ~ distvei_400 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cutroad5 <- glmmTMB(bt100 ~ distvei_300 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cutroad6 <- glmmTMB(bt100 ~ distvei_350 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

beta_cutroad7 <- glmmTMB(bt100 ~ distvei_200 + skogkategori + disthus*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST

beta_cutroad8 <- glmmTMB(bt100 ~ distvei_150 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

```

#### AIC selection threshold road
```{r}
library(AICcmodavg)

candmods_beta_cut_road <- list(beta_cutroad1, beta_cutroad2, beta_cutroad3, beta_cutroad4, beta_cutroad5, beta_cutroad6, beta_cutroad7, beta_cutroad8)
# candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3)
aictab(candmods_beta_cut_road, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod7 20 -9780.70      0.00  0.82   0.82 4910.35
# Mod8 20 -9777.44      3.26  0.16   0.98 4908.72
# Mod5 20 -9773.24      7.46  0.02   1.00 4906.62
# Mod6 20 -9769.91     10.79  0.00   1.00 4904.96
# Mod4 20 -9764.46     16.24  0.00   1.00 4902.23
# Mod3 20 -9752.98     27.72  0.00   1.00 4896.49
# Mod2 20 -9743.97     36.73  0.00   1.00 4891.98
# Mod1 20 -9701.83     78.87  0.00   1.00 4870.92

```



```{r}
browse_house$disthus_200 <- scale(browse_house$disthus_200)


distveilog_beta <-glmmTMB(bt100 ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) 

distveithresh_beta <- glmmTMB(bt100 ~ distvei_200 + disthus + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # Best


candmods_beta_thresh <- list(distveilog_beta, distveithresh_beta)
aictab(candmods_beta_thresh, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod2 20 -9780.70      0.00     1      1 4910.35
# Mod1 20 -9762.64     18.06     0      1 4901.32

save(distveilog_beta, file = "output/log_or_thresh/distveilog_beta.rda")
save(distveithresh_beta, file = "output/log_or_thresh/distveithresh_beta.rda")

library(sjPlot)
plot_model(distveithresh_beta, type = "eff", terms = "distvei_200")

```





## Beta house models

Prepare data set for second part of the hurdle
```{r}
load("processdata/browse_house_2021_04_12.rda")
browse_house <- browse_house[browse_house$bt100 > 0, ]
str(browse_house)
summary(browse_house$bt100)

```

#### Scale variables if necessary
```{r}
# browse_house$disthus <- scale(browse_house$disthus)
# browse_house$distvei2 <- scale(browse_house$distvei2)

browse_house$tretetthet9 <- scale(browse_house$tretetthet9)
browse_house$helling <- scale(browse_house$helling)
browse_house$HOH <- scale(browse_house$HOH)
browse_house$moose_density <- scale(browse_house$moose_density)

# browse_house$disthus_200 <- scale(browse_house$disthus_200)


# summary(browse_house$disthus_200)
```


#### Beta house models
```{r}
library(glmmTMB)

browse_h_beta1 <- glmmTMB(bt100 ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())


## Test for interactions:

browse_h_beta2 <- glmmTMB(bt100 ~ disthus_200*skogkategori + log(distvei2+1) + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

browse_h_beta3 <- glmmTMB(bt100 ~ disthus_200*skogkategori + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family())

browse_h_beta4 <- glmmTMB(bt100 ~ disthus_200 + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST


browse_h_beta5 <- glmmTMB(bt100 ~ disthus_200 + log(distvei2+1)*skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
            data = browse_house,
            family = beta_family()) # BEST

```


#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_browse_house_beta <- list(browse_h_beta1, browse_h_beta2, browse_h_beta3, browse_h_beta4)
candmods_browse_house_beta <- list(browse_h_beta4, browse_h_beta5)
aictab(candmods_browse_house_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod2 23 -9778.47      0.00  0.71   0.71 4912.23
# Mod1 24 -9776.69      1.78  0.29   1.00 4912.34


# Interactions
#       K      AIC Delta_AIC AICWt Cum.Wt      LL
# Mod4 24 -9776.69      0.00  0.95   0.95 4912.34
# Mod3 28 -9770.53      6.16  0.04   1.00 4913.26
# Mod1 20 -9762.64     14.04  0.00   1.00 4901.32
# Mod2 24 -9761.77     14.92  0.00   1.00 4904.88


save(browse_h_beta1, file = "output/2021_04_28/browse_h_beta1.rda")
save(browse_h_beta2, file = "output/2021_04_28/browse_h_beta2.rda")
save(browse_h_beta3, file = "output/2021_04_28/browse_h_beta3.rda")
save(browse_h_beta4, file = "output/2021_04_28/browse_h_beta4.rda")
save(browse_h_beta5, file = "output/2021_04_28/browse_h_beta5.rda")

```


#### Summary beta house
```{r}
summary(browse_h_beta5)

```



#### Plots beta house
```{r}
# library(effects)
# plot(allEffects(rec_h_beta7))
# plot(effect("skogkategori", rec_h_beta7))

library(sjPlot)
plot_model(browse_h_beta5, type = "diag")
plot_model(browse_h_beta5, type = "eff")
plot_model(browse_h_beta5, type = "est")

tab_model(browse_h_beta5, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

plot_model(browse_h_beta5, type = "eff", terms = c("distvei2","skogkategori"))
plot_model(browse_h_beta5, type = "int")

get_model_data(browse_h_beta5, type = "eff")

```


#### Save house beta model
```{r}
save(browse_h_beta5, file = "output/2021_04_28/browse_h_beta5.rda")
```
