==> 05_browsing_pressure_new.Rmd <==
---
title: "05_browsing_pressure_new"
author: "anne"
date: "1 februar 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


# Load data
```{r}
load("processdata/browse9.2.rda")
load("processdata/browse9.2_dist.rda")
load("processdata/browse9.2pos_dist.rda")
```




#### Modelling spatial autocorrelation (https://onlinelibrary.wiley.com/doi/full/10.1111/j.2007.0906-7590.05171.x)
Change the projection and make coordinates and a spatial point data frame to make a shapefile

```{r}
library(tidyr)
browse9.5$UTMmerge <- paste(browse9.5$UTM_OV_33, browse9.5$UTM_SN_33, sep="_")
sum(duplicated(browse9.5$UTMmerge))
```


```{r}

library(sp)
library(rgdal)

crs <- CRS("+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs")
coords <- cbind(browse9.5$UTM_OV_33, browse9.5$UTM_SN_33)
data_spdf <- SpatialPointsDataFrame(coords, browse9.5, proj4string = crs)
# plot(data_spdf)
writeOGR(data_spdf, dsn = "output", layer = "spac", driver = "ESRI Shapefile")

```


```{r}

library(spdep)
library(ape)

# spdep package:

# prepare neighbour lists for spatial autocorrelation analysis
# nb.list <- dnearneigh(as.matrix(browse9.3[browse9.3$beitetrykkr, c("UTM_OV_33", "UTM_SN_33")]), 0, 5000)
# nb.weights <- nb2listw(nb.list, zero.policy = TRUE)
# 
# moran.plot(resid_m1abin, nb.weights, zero.policy = TRUE, spChk = NULL, labels = TRUE, xlab = NULL, ylab = NULL, quiet = NULL)
# 
# resid_m1abin <- resid(m1abin, type = "pearson")
# resid_m1bbin <- resid(m1bbin, type = "pearson")
# 
# moran.test(browse9.3$beitetrykkr, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)
# 
# moran.test(resid_m1bbin, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)


# Make a matrix of coordinates - use when chunk above was not run
coords <- as.matrix(cbind(browse9.5$UTM_OV_33, browse9.5$UTM_SN_33))

# compute the autocovariate based on the above distance and weight
ac <- autocov_dist(browse9.5$beitetrykkr, coords, nbs = 5000, type="inverse", zero.policy = TRUE)
browse9.5$ac <- ac


save(browse9.3, file = "browse9.3.rda")


```




#### gam

```{r}
library(ggplot2)
ggplot(browse9.5, aes(x = disthus, y = beitetrykkr)) +
    geom_point() +
    stat_smooth(method = "glm", color = "green") +
  stat_smooth(color = "blue")

# +
#   scale_x_continuous(breaks=c(100,150, 200, 250, 300, 400, 500, 1000, 1500, 2000, 3000))

```

```{r}
library(mgcv)
gam1 <- gam(beitetrykkr ~ distvei2 + disthus + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet9,
             data = browse9.2,
             method = "REML")

gam2 <- gam(beitetrykkr ~ s(distvei2) + s(disthus) + skogkategori + treartgruppe + region + kant + helling + HOH + tretetthet9,
             data = browse9.2,
             method = "REML")



gam1 <- gam(beitetrykkr ~ distvei2 + disthus + skogkategori + treartgruppe9 + region + kant + helling + HOH + andtre,
             data = browse9.5,
             method = "REML")

gam2 <- gam(beitetrykkr ~ s(distvei2) + s(disthus) + skogkategori + treartgruppe9 + region + kant + helling + HOH + andtre,
             data = browse9.5,
             method = "REML")

```

```{r}
plot.gam(gam1, all.terms = TRUE)
plot.gam(gam2, all.terms = TRUE)
# threshold: disthus ~700/900m, distvei2 ~200/300m
```






######################################################################
## Bin - First part of a hurdle model to account for zero inflation###
######################################################################

##### Checking for the cut for disthus and log-selection
```{r}
library(lme4)
m1abin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1bbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut1000_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1cbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut900_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1dbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut800_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1ebin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut700_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1fbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut600_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1gbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut500_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1hbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut400_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1ibin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut300_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1jbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) # second best

m1kbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut150_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) 

m1lbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut175_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) # best

m1lbin2 <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut175_scale + log(distvei2+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1sbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut100_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) 


# ------------ log selection -----------
m1nbin <- glmer(factor(beitetrykkr > 0) ~ log(disthus+1) + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale +  (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1oin <- glmer(factor(beitetrykkr > 0) ~ log(disthus+1) + distvei2_trunc_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale +  (1|KOMNR),
            family = binomial(),
            data = browse9.2)

m1pbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) 

m1qbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut175_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) # best

m1rbin <- glmer(factor(beitetrykkr > 0) ~ log(disthus+1) + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2) 

```


```{r}

AIC(m1abin, m1gbin, m1hbin, m1ibin, m1jbin, m1kbin, m1lbin)
AIC(m1lbin, m1lbin2)
AIC(m1jbin, m1jbin2)

AIC(m1nbin, m1oin, m1pbin, m1qbin, m1rbin)


library(AICcmodavg)

candmods_BP_hbin <- list(m1abin, m1gbin, m1hbin, m1ibin, m1jbin, m1kbin, m1lbin)


glmmbinh_tab <- aictab(candmods_BP_hbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)

# cut150 is the thing!


```


```{r}
library(sjPlot)
p1_below <- plot_model(m1kbin, type = c("eff"), transform = NULL, terms = c("disthus_trunc_cut150_scale [all]"),
           grid = TRUE)
p1_below + theme_sjplot()



tab_model(m1lbin, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))

```

##### Model selection BIN (roads)
```{r}
library(lme4)

m2bin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR), 
               family = binomial(), 
               data = browse9.2_dist) # Best

m3bin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR), 
               family = binomial(), 
               data = browse9.2_dist)

m4bin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + tretetthet9_trunc_scale + (1|KOMNR), 
               family = binomial(), 
               data = browse9.2_dist)

m4abin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            family = binomial(),
            data = browse9.2_dist) # BEST, just as m2bin but with truncated distvei2

# m2bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*skogkategori + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)

# m2abin <- m1kbin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*skogkategori + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)

# m3bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*veitype2 + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)
# 
# m3abin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*veitype2 + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)

# m4bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)



# m5bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*veitype2 + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)
# 
# m6bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*veitype2 + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)

# m7bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2_trunc+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)
# 
# m8bin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_scale + log(distvei2_trunc+1) + veitype2 + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)

```



```{r}

AIC(m2bin, m3bin, m4bin, m4abin2)
AIC(m4bin, m7bin, m8bin)


library(AICcmodavg)
candmods_BP_rbin <- list(m4abin2, m3bin, m4bin)

glmmbinr_tab <- aictab(candmods_BP_rbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)

# m4(a)bin is the thing!

modList <- list(l7, l6, l5, l4, b7, b6, b5, b4)
aics <- sapply(modList, function(x) AIC(pcrfit(reps, 1, 2, x))) 
akaike.weights(aics)$weights

```


```{r}
# Check which optimizer doesn't converge
allFit(m4abin)
# says: one optimizer failed, but doesn't tell me which...
```

Testing for normality of errors (Although I don't know if this works for binomial)
```{r}
summary(m2abin)

plot(m4bin)
hist(resid(m4bin))

resid.m4bin <- resid(m4bin)
plot(resid.m4bin)

par(mfrow = c(3,3))
plot(resid.m2abin ~ skogkategori, data = browse9.2_dist)
plot(resid.m2abin ~ treartgruppe, data = browse9.2_dist)
plot(resid.m2abin ~ kant, data = browse9.2_dist)
plot(resid.m2abin ~ helling_scale, data = browse9.2_dist)
plot(resid.m2abin ~ HOH_scale, data = browse9.2_dist)
plot(resid.m2abin ~ tretetthet9_trunc_scale, data = browse9.2_dist)
# plot(resid.m2abin ~ disthus_trunc_scale, data = browse9.2_dist)
plot(resid.m2abin ~ log(distvei2+1), data = browse9.2_dist)

library(sjPlot)
plot_model(m2abin, type = c("diag"), transform, terms = NULL)
# plot_model(m10bin, type = c("resid"), transform, terms = NULL)

vcov.merMod(m1qbin)
coef(m4abin)
fixef(m4abin)

cor((as.matrix(browse9.2_dist[browse9.2_dist$beitetrykkr, c("distvei2", "helling", "HOH", "veitype2")])))


```

Table of estimates:
```{r}
library(sjPlot)
tab_model(m4abin2, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE,
          show.std = NULL, show.p = TRUE, 
          show.stat = FALSE, show.df = FALSE, show.r2 = TRUE, 
          show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, 
          show.dev = FALSE, show.obs = TRUE)


```

Forest plot:
Whiskers are ci, possible to show se instead; Colourblind friendly colours "#004949","#009292"
```{r}
p1 <- plot_model(m4abin, type = c("est"), transform = NULL, show.values = TRUE, value.offset = 0.5, 
                 vline.color = "grey", colors = c("#004949", "#009292"),
                 axis.title = "Estimates")
p1 + theme_sjplot() + font_size(axis_title.x = 15)

```
## figures bin
```{r}
library(sjPlot)
library(ggplot2)
p1 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = c("treartgruppe9"),
           grid = TRUE)
p1 <- p1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Browsing pressure (%)\n")


p2 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = c("kant"))
p2 <- p2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Edge effects\n") +
  ylab("Browsing pressure (%)\n") +
  label_angle(angle.x = 45, base.theme = theme_sjplot())

p3 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = c("helling [all]"))
p3 <- p3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Browsing pressure (%)\n")

p4 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = c("HOH_scale [all]"))
p4 <- p4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Browsing pressure (%)\n")

p5 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = c("tretetthet9_trunc [all]"))
p5 <- p5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree density\n") +
  ylab("Browsing pressure (%)\n")

p6 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = "distvei2_trunc [0:500]")
p6 <- p6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to roads\n") +
  ylab("Browsing pressure (%)\n")

p7 <- plot_model(m4abin2, type = c("eff"), transform = NULL, terms = "skogkategori")
p7 <- p7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Forest category\n") +
  ylab("Browsing pressure (%)\n") +
  label_angle(angle.x = 45, base.theme = theme_sjplot())

# p8 <- plot_model(m4abin, type = c("eff"), transform = NULL, terms = c("disthus_trunc_scale"))
# p7 + theme_sjplot()

# p9 <- plot_model(m4abin, type = c("eff"), transform = NULL, terms = c("distvei2 [0:3000]", "veitype2"))
# p9 + theme_sjplot()

# p10 <- plot_model(m8bin, type = c("eff"), transform = NULL, terms = c("veitype2"))
# p10 + theme_sjplot()

p1_house <- plot_model(m1kbin2, type = c("eff"), transform = NULL, terms = c("disthus_trunc_cut200_scale [all]"),
           grid = TRUE)
p1_house <- p1_house + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Browsing pressure (%)\n")


library(gridExtra)
grid.arrange(p1_house, p6, p1, p5, p7, p2, p3, p4,
          nrow = 4, ncol = 2)

# ggpredict(m1kbin2, terms="skogkategori [all]", ci.lvl = 0.95, type = c("re"), typical_value(x, fun = c(numeric = "median", factor = "mean")))


# library(effects)
# plot(allEffects(m4glmm))
```




## Beta - Second part of a hurdle model

##### Checking for the cut for disthus and best beta-house-model
```{r}
library(glmmTMB)
browse9.2pos <- browse9.2[browse9.2$bt100 > 0, ]

m1beta <- glmmTMB(bt100 ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale +  (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1bbeta <- glmmTMB(bt100 ~ disthus_trunc_cut1000_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1cbeta <- glmmTMB(bt100 ~ disthus_trunc_cut900_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1dbeta <- glmmTMB(bt100 ~ disthus_trunc_cut800_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1ebeta <- glmmTMB(bt100 ~ disthus_trunc_cut700_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1fbeta <- glmmTMB(bt100 ~ disthus_trunc_cut600_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1gbeta <- glmmTMB(bt100 ~ disthus_trunc_cut500_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1hbeta <- glmmTMB(bt100 ~ disthus_trunc_cut400_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1ibeta <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family()) # best

m1jbeta <- glmmTMB(bt100 ~ disthus_trunc_cut175_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1kbeta <- glmmTMB(bt100 ~ disthus_trunc_cut150_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1lbeta <- glmmTMB(bt100 ~ disthus_trunc_cut190_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family()) 

# ------ log selection ------
m1mbbeta <- glmmTMB(bt100 ~ log(disthus+1) + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1nbeta <- glmmTMB(bt100 ~ log(disthus+1) + distvei2_trunc_scale + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family())

m1obeta <- glmmTMB(bt100 ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family()) 

m1pbeta <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos,
            family = beta_family()) # Best


```


```{r}
AIC(m1beta, m1gbeta, m1hbeta, m1ibeta, m1jbeta)
# cut200 is the thing!
AIC(m1mbbeta, m1nbeta, m1obeta, m1pbeta)
# m1pbeta is the thing!

# library(AICcmodavg)
library(bblme)

# candmods_BP_hbeta <- list(m1beta, m1gbeta, m1hbeta, m1ibeta, m1jbeta)
# 
# glmmbin_tab <- aictab(candmods_BP_hbeta, modnames = NULL,
#                        second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

ICtab(m1beta, m1gbeta, m1hbeta, m1ibeta, m1jbeta, type = "AIC", weights = TRUE, delta = TRUE)


save(m1gbeta, file = "m1gbeta.rda")
save(m1hbeta, file = "m1hbeta.rda")
save(m1ibeta, file = "m1ibeta.rda")
save(m1jbeta, file = "m1jbeta.rda")
save(m1kbeta, file = "m1kbeta.rda")
save(m1lbeta, file = "m1lbeta.rda")

save(m1mbbeta, file = "m1mbbeta.rda")
save(m1nbeta, file = "m1nbeta.rda")
save(m1obeta, file = "m1obeta.rda")
save(m1pbeta, file = "m1pbeta.rda")

load("output/recentBP_selection/m1gbeta.rda")
load("output/recentBP_selection/m1hbeta.rda")
load("output/recentBP_selection/m1ibeta.rda")
load("output/recentBP_selection/m1jbeta.rda")
load("output/recentBP_selection/m1kbeta.rda")
load("output/recentBP_selection/m1lbeta.rda")

load("output/recentBP_selection/m1mbbeta.rda")
load("output/recentBP_selection/m1nbeta.rda")
load("output/recentBP_selection/m1obeta.rda")
load("output/recentBP_selection/m1pbeta.rda")
```


```{r}
library(sjPlot)
p2_below <- plot_model(m1ibeta, type = c("eff"), transform = NULL, terms = c("disthus_cut200_scale [all]"), grid = TRUE)

p2_below + theme_sjplot()


tab_model(m1ibeta, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))

```

##### Model selection beta (roads)
```{r}
library(glmmTMB)
browse9.2pos_dist <- browse9.2_dist[browse9.2_dist$bt100 > 0, ]

m2beta <- glmmTMB(bt100 ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos_dist,
            family = beta_family())

m3beta <- glmmTMB(bt100 ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos_dist,
            family = beta_family())

m4beta <- glmmTMB(bt100 ~ log(distvei2+1) + skogkategori + treartgruppe + ac_scale + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos_dist,
            family = beta_family())

m4abeta <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos_dist,
            family = beta_family()) # still BEST (full model which is m2beta, but with distvei truncated)

m5beta <- glmmTMB(bt100 ~ log(distvei2_trunc+1) +  treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
            data = browse9.2pos_dist,
            family = beta_family())

# m3beta <- glmmTMB(bt100 ~ disthus_trunc_scale + log(distvei2+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*skogkategori + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())

# m3abeta <- glmmTMB(bt100 ~ log(distvei2+1) + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*skogkategori + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())

# m4beta <- glmmTMB(bt100 ~ disthus_trunc_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())



# m5beta <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale +  tretetthet9_trunc_scale + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())

# m6beta <- glmmTMB(bt100 ~ disthus_trunc_scale + log(distvei2_trunc+1) + veitype2 + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())


```

```{r}
AIC(m4beta, m4abeta, m5beta)
# m4(a)beta is the thing!
AIC(m2beta, m3beta, m4beta, m4abeta, testBeta, m5beta)
AIC(m2beta, m4abeta)

save(m2beta, file = "m2beta.rda")
save(m3beta, file = "m3beta.rda")
save(m4beta, file = "m4beta.rda")
save(m4abeta, file = "output/recentBP_selection/m4abeta.rda")
save(m5beta, file = "output/recentBP_selection/m5beta.rda")
save(testBeta, file = "output/recentBP_selection/testBeta.rda")

load("output/recentBP_selection/m2beta.rda")
load("output/recentBP_selection/m3beta.rda")
load("output/recentBP_selection/m4beta.rda")

summary(m4abeta)

# library(AICcmodavg)
library(bblme)

# candmods_BP_hbeta <- list(m1beta, m1gbeta, m1hbeta, m1ibeta, m1jbeta)
# 
# glmmbin_tab <- aictab(candmods_BP_hbeta, modnames = NULL,
#                        second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

ICtab(m4abeta, m2beta, m3beta, m4beta, type = "AIC", weights = TRUE, delta = TRUE)
AIC(m4abeta, m2beta, m3beta, m4beta)

```



```{r}
resid_m4abeta <- resid(m4abeta, type = "pearson")
plot(resid_m4abeta)

overdisp_fun <- function(model) {
    rdf <- df.residual(model)
    rp <- residuals(model,type="pearson")
    Pearson.chisq <- sum(rp^2)
    prat <- Pearson.chisq/rdf
    pval <- pchisq(Pearson.chisq, df=rdf, lower.tail=FALSE)
    c(chisq=Pearson.chisq,ratio=prat,rdf=rdf,p=pval)
}

overdisp_fun(m3beta)
# Have overdispersion, but he (Ben Bolker) says: "Remember that (1) overdispersion is irrelevant for models that estimate a scale parameter (i.e. almost anything but Poisson or binomial: Gaussian, Gamma, negative binomial …) and (2) overdispersion is not estimable (and hence practically irrelevant) for Bernoulli models (= binary data = binomial with N=1)."
```

Table of estimates:
```{r}
library(sjPlot)

tab_model(m4abeta2, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))

tab_model(m4abeta2, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))

r2(m4abeta2)

```

Forest plot:
Whiskers are ci, possible to show se instead; Colourblind friendly colours "#004949","#009292"
```{r}
p1_est <- plot_model(m4abeta2, type = c("est"), transform = NULL, show.values = TRUE, value.offset = 0.5, 
                 vline.color = "grey", colors = c("#004949", "#009292"),
                 axis.title = "Estimates")
p1_est + theme_sjplot() + font_size(axis_title.x = 15)


```


## figures beta
```{r}
library(sjPlot)
library(ggplot2)

load("processdata/browse9.4pos.rda")

pb2 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("distvei2_trunc [0:500]"))
pb2 <- pb2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to roads\n") +
  ylab("Browsing pressure\n")


pb3 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("skogkategori"))
pb3 <- pb3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Forest category\n") +
  ylab("Browsing pressure\n") + 
  label_angle(angle.x = 45, base.theme = theme_sjplot())


pb4 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("treartgruppe9"),
                  show.p = TRUE)
pb4 <- pb4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Browsing pressure\n")


pb5 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("kant"))
pb5 <- pb5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Edge effects\n") +
  ylab("Browsing pressure\n") + 
  label_angle(angle.x = 45, base.theme = theme_sjplot())


pb6 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("helling [all]"))
pb6 <- pb6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Browsing pressure\n")

pb7 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("HOH_scale [all]"))
pb7 <- pb7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Browsing pressure\n")

pb8 <- plot_model(m4abeta2, type = c("eff"), transform = NULL, terms = c("tretetthet9_trunc [all]"))
pb8 <- pb8 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree density\n") +
  ylab("Browsing pressure\n")

# pb9 <- plot_model(m5beta, type = c("eff"), transform = NULL, terms = c("distvei2 [0:3000]", "veitype2"))
# pb9 + theme_sjplot()

browse9.4_house_pos <- browse9.4_house[browse9.4_house$bt100 > 0, ]

p2_house <- plot_model(m1ibeta2, type = c("eff"), transform = NULL, terms = c("disthus_trunc_cut200_scale [all]"), grid = TRUE)
p2_house <- p2_house + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Browsing pressure\n")


library(gridExtra)
grid.arrange(p2_house, pb2, pb4, pb8, pb3, pb5, pb6, pb7,
          nrow = 4, ncol = 2)

# type = "pred" = Predicted values (marginal effects) for specific model terms. See ggpredict for details.
# type = "eff" = Similar to type = "pred", however, discrete predictors are held constant at their proportions (not reference level). See ggeffect for details. ggeffect() computes marginal effects by internally calling Effect. Effect constructs an "eff" object for a term (usually a high-order term), absorbing the lower-order terms marginal to the term in question, and averaging over other terms in the model.

```
