==> 06_treedensity.Rmd <==
---
title: "treedensity"
author: "anne"
date: "14 februar 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


## Load data
```{r}
# load("processdata/browse9.2.rda")
load("processdata/browse9.3.rda")
load("processdata/browse9.3_dist.rda")
```


```{r}
browse9.3_dist$log_distvei2_trunc <- log(browse9.3_dist$distvei2_trunc+1)
```


##### Model selection first step (log or not, cut or not)
```{r}
library(lme4)
# m1a <- glmer(tretetthet9_trunc ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             family = poisson(),
#             data = browse9.2)

m1 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

# m2 <- glmer(tretetthet9_trunc ~ disthus_trunc_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             family = gaussian(),
#             data = browse9.2)

m3 <- lmer(log(tretetthet9+1) ~ disthus_trunc_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

m4 <- lmer(tretetthet9_trunc ~ disthus_trunc_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

m5 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_scale + distvei2_trunc_scale + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

m6 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut200_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

m11 <- lmer(log(tretetthet9+1) ~ log(disthus_trunc+1) + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)
# 
# m7 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut100_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             data = browse9.3)
# 
# m8 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut300_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             data = browse9.3)
# 
# m9 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut250_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             data = browse9.3)
# 
# m10 <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut250_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
#             data = browse9.3_dist)

```

```{r}
AIC(m1, m3, m4, m5, m6, m11)
```


##### Model selection
```{r}
library(lme4)
mhus <- lmer(log(tretetthet9_trunc+1) ~ disthus_trunc_cut200_scale + log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3)

m11 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + (1|KOMNR),
            data = browse9.3_dist)

m12 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + vegetasjon + (1|KOMNR),
            data = browse9.3_dist)

m13 <- lmer(log(tretetthet9_trunc+1) ~ log_distvei2_trunc + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + helling + (1|KOMNR),
            data = browse9.3_dist)

m13a <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + helling + (1|KOMNR),
            data = browse9.3_dist)

m14 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + treartgruppe + ac_scale + kant + beitetrykkr + vegetasjon + (1|KOMNR),
            data = browse9.3_dist)

m15 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + skogkategori + ac_scale + kant + beitetrykkr + vegetasjon + (1|KOMNR),
            data = browse9.3_dist)

m16 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + ac_scale + kant + beitetrykkr + vegetasjon + (1|KOMNR),
            data = browse9.3_dist)

m17 <- lmer(log(tretetthet9_trunc+1) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + beitetrykkr + helling + HOH + (1|KOMNR),
            data = browse9.3_dist)

```


```{r}
AIC(m11, m12, m13, m14, m15, m16, m17)

#     df      AIC
# m11 15 60884.29
# m12 31 60675.62 <-- Lowest AIC, but I think we should mayeb not use vegetation as predictor
# m13 16 60870.28 <-- BEST
# m14 27 61097.84
# m15 29 63724.99
# m16 25 64001.84
# m17 17 60886.50

summary(m1)
summary(m9)
summary(m11)
summary(m13)

```


```{r}
plot(m13)
resid_m13 <- resid(m13,type = "pearson")
plot(resid_m13)
hist(resid_m13)
qqnorm(resid_m13)

vcov.merMod(m13)

```

##### Figs and tabs

###### Trying to get estimates
```{r}
# library(emmeans)
# library(lmerTest)
# emm_options(pbkrtest.limit = 22325, lmerTest.limit = 22325)
# 
# # ls_means(m13, test.effs=NULL, method.grad='simple')
# 
# coefdisthus <- as.data.frame(emmeans(mhus, specs = "disthus_trunc_cut200_scale"))
# coefdistvei <- emmeans(m13, specs = "distvei2_trunc")
# coefskogkat <- emmeans(m13, specs = "skogkategori")
# coeftreart <- emmeans(m13, specs = "treartgruppe")
# coefkant <- emmeans(m13, specs = "kant")
# coefbt <- emmeans(m13, specs = "beitetrykkr")
# coefhell <- emmeans(m13, specs = "helling")
# 
# newdata <- as.data.frame(matrix(ncol = 1, nrow = 19865))
# predicted_fe <- predict(m13, newdata = browse9.3_dist, type = c("response"), allow.new.levels = T)
# 
# predicted_ci <- confint(m13, method = c("boot"), nsim = 100)
# predicted_ci1 <- confint(m13, method = c("Wald"), intercept = F)
# predicted_ci2 <- confint(m13, method = c("Wald"))
# predicted_ci3 <- confint(m13, method = c("Wald"), boot.type = c("norm") )
# 
# library(effects)
# test <- as.data.frame(effect(mod = m13, se = T, term = "helling"))
# plot(allEffects(m13))
# mySumm2 <- function(.) {
#     c(beta=fixef(.),sigma=sigma(.), sig01=sqrt(unlist(VarCorr(.))))
# }
# 
# t1 <- mySumm2(m13)
# boo01 <- bootMer(m13, mySumm2, nsim = 100)

library(sjPlot)
resdisthus <- get_model_data(mhus, type = "eff",  terms = "disthus_trunc_cut200_scale",  pred.type = "re")
# resdistvei <- get_model_data(m13a, type = "eff",  terms = "log_distvei2_trunc",  pred.type = "re")
resdistvei <- get_model_data(m13a, type = "eff",  terms = "distvei2_trunc [0:2500]",  pred.type = "re")
resskogkat <- get_model_data(m13a, type = "eff",  terms = "skogkategori",  pred.type = "re")
restreart <- get_model_data(m13a, type = "eff",  terms = "treartgruppe",  pred.type = "re")
reskant <- get_model_data(m13a, type = "eff",  terms = "kant",  pred.type = "re")
reshelling <- get_model_data(m13a, type = "eff",  terms = "helling",  pred.type = "re")
resbeitetrykk <- get_model_data(m13a, type = "eff",  terms = "beitetrykkr",  pred.type = "re")

```


##### Figures with ggplot
```{r}
library(ggplot2)

p1 <- ggplot(data = resdisthus,aes(x, predicted)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nDistance to house") + 
  ylab("Tree density (log) \n") +
  scale_y_continuous(trans = "exp")

p2 <- ggplot(data = resdistveia,aes(x, predicted)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.title.y = element_blank(), axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nDistance to road") + 
  ylab("Tree density (log) \n") +
  scale_y_continuous(trans = "exp")

p3 <- ggplot(data = restreart,aes(x, predicted)) +
  geom_pointrange(aes(ymin = conf.low, ymax = conf.high), colour = "#009292", size = 0.6) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nTree species group") + 
  ylab("Tree density (log)\n") +
  scale_x_continuous(limits = c(1,3.5), breaks = c(1,2,3), labels = c("Pine", "Other deciduous \ntrees", "Rowan, aspen, \nwillow")) +
  scale_y_continuous(trans = "exp")

p4 <- ggplot(data = reskant,aes(x, predicted)) +
  geom_pointrange(aes(ymin = conf.low, ymax = conf.high), colour = "#009292", size = 0.6) +
  theme_light() +
  theme(axis.title.y = element_blank(), axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nEdge") + 
  ylab("Tree density (log)\n") +
  scale_x_continuous(labels = c("Edge <10m", "Edge >20m", "Edge 10-10m", "Edge \nmissing")) +
  scale_y_continuous(trans = "exp")

p5 <- ggplot(data = reshelling,aes(x, predicted)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.title.y = element_blank(),axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nSlope") + 
  ylab("Tree density (log)\n") +
  scale_y_continuous(trans = "exp")

p6 <- ggplot(data = resbeitetrykk,aes(x, predicted)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = conf.low, ymax = conf.high), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nBrowsing pressure (%)") + 
  ylab("Tree density (log) \n") +
  scale_y_continuous(trans = "exp")

p7 <- ggplot(data = resskogkat,aes(x, predicted)) +
  geom_pointrange(aes(ymin = conf.low, ymax = conf.high), colour = "#009292", size = 0.6) +
  theme_light() +
  theme(axis.title.y = element_blank(), axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nForest category") + 
  ylab("Tree density (log)\n") +
  scale_x_continuous(limits = c(1,5.4), breaks = c(1,2,3,4,5), labels = c("Highproductive \nold forest", "Highproductive \nyoung forest", "Lowproductive \nold forest", "Lowproductive \nyoung forest", "Unproductive \nforest")) +
  scale_y_continuous(trans = "exp")



library(gridExtra)
grid.arrange(p3, p4, p7, p1, p2, p5, p6,
          nrow = 3, ncol = 3)


```



##### Figures with sjPlot
```{r}
library(sjPlot)
library(ggplot2)
```

```{r}

tab_model(m13, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))

```

```{r}
p1_est <- plot_model(m13, type = c("est"), transform = NULL, show.values = TRUE, value.offset = 0.5, 
                 vline.color = "grey", colors = c("#004949", "#009292"),
                 axis.title = "Estimates")
p1_est + theme_sjplot()
```

log.y = F

```{r}
p1hus <- plot_model(mhusabin, type = c("eff"), terms = c("disthus_400"))
# p1hus + theme_light() +
#   theme(axis.text = element_text(family = "sans", size = 10, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
#   xlab("\nDistance to house") + 
#   ylab("Tree density (log)\n")

p1 <- plot_model(m16abeta, type = c("eff"), terms = c("distvei2 [0:500]"))
# p1 + theme_light() +
#   theme(axis.text = element_text(family = "sans", size = 10, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
#   xlab("\nDistance to road") + 
#   ylab("Tree density (log)\n")

p2 <- plot_model(m16abeta, type = c("eff"), transform = NULL, terms = c("skogkategori"))
# p2 + theme_light()

p3 <- plot_model(m16abeta, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
# pb3 + theme_sjplot()

# p4 <- plot_model(m16cbin, type = c("eff"), transform = NULL, terms = c("kant"))
# pb3 + theme_sjplot()

# p5 <- plot_model(m16cbin, type = c("eff"), transform = NULL, terms = c("helling [n=1]"))
# pb3 + theme_sjplot()

p6 <- plot_model(m16abeta, type = c("eff"), transform = NULL, terms = c("browspres_merged"))
# pb3 + theme_sjplot()

p7 <- plot_model(m16abeta, type = c("eff"), transform = NULL, terms = c("HOH"))
# pb3 + theme_sjplot()

library(gridExtra)
plot_grid(list(p1, p6, p2, p3, p7))
```
