==> 07_total_effect_hurdle.Rmd <==
---
title: "total hurdle"
author: "anne"
date: "15 februar 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r}
library(pscl)
library(lme4)
library(glmmTMB)
library(boot)
load("processdata/browse9.2_dist.rda")
load("processdata/browse9.2pos_dist.rda")
load("processdata/browse9.2.rda")
load("processdata/browse9.2pos.rda")
```

#### models for >200m from houses
```{r}
library(lme4)
# m4abin <- glmer(factor(beitetrykkr > 0) ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2_dist)
load("output/m4bin.rda") # is m4abin

# m4abeta <- glmmTMB(bt100 ~ log(distvei2_trunc+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             data = browse9.2pos_dist,
#             family = beta_family())
load("output/m4beta.rda") # is m4abeta

```

#### models for <200m from houses
```{r}
# library(lme4)
# m1kbin <- glmer(factor(beitetrykkr > 0) ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + (1|KOMNR),
#             family = binomial(),
#             data = browse9.2)
# save(m1kbin, file = "m1kbin.rda")
load("output/m1kbin.rda")

# library(glmmTMB)
# m1ibeta <- glmmTMB(bt100 ~ disthus_trunc_cut200_scale + log(distvei2+1) + skogkategori + treartgruppe + ac_scale + kant + helling_scale + HOH_scale + I(HOH_scale^2) + tretetthet9_trunc_scale + log(distvei2+1)*skogkategori + (1|KOMNR),
#             data = browse9.2pos,
#             family = beta_family()) 
# save(m1ibeta, file = "m1ibeta.rda")
load("output/m1ibeta.rda")

```

#### preparation for the function
```{r}
zeromod <- m1kbin
percmod <- m1ibeta

dat <- browse9.2
focal <- "disthus_trunc_cut200_scale"
```

#### function
```{r}
# focal: a numeric variable, doesn't work for factors, yet 
resp_curve <- function(focal, isfactor=FALSE, dat, zeromod, percmod, nvals=100, plot=T, nb_btstrp=999){
  if (!isfactor) rng <- quantile(dat[,focal], probs=c(0.01, 0.99))
  if (!isfactor) dat[c(1:nvals),focal] <- seq(rng[1], rng[2], length.out=nvals)
# model matrix for zeromod
mm1 <- model.matrix(terms(zeromod), data=dat)

# set to mean value, or set factors to intercept
colnms <- colnames(mm1)
fcts <- names(zeromod@frame)[sapply(zeromod@frame, class)=="factor"]
  if (!isfactor) mm1[c(1:nvals),-grep(focal, colnms)] <- rep(apply(mm1[,-grep(focal, colnms)], 2, median), each=nvals)
  if (!isfactor) mm1 <- mm1[c(1:nvals),] 
  if (isfactor) mm1 <- mm1[c(1:(1+length(grep(focal, colnms)))),] 
  for (i in c(fcts)){
    if (length(grep(i, colnms))>0) mm1[,grep(i, colnms)] <- 0
  }
  if (isfactor) {
    j=1
    for (i in grep(focal, colnms)){
      j=j+1
      mm1[j,i] <- 1
    }
  }

# model matrix for percmod
mm2 <- model.matrix(terms(percmod), data=dat)
  
# set to mean value, or set factors to intercept
colnms <- colnames(mm2)
fcts <- names(percmod$frame)[sapply(percmod$frame, class)=="factor"]
  if (!isfactor) mm2[c(1:nvals),-grep(focal, colnms)] <- rep(apply(mm2[,-grep(focal, colnms)], 2, median), each=nvals)
  if (!isfactor) mm2 <- mm2[c(1:nvals),] 
  if (isfactor) mm2 <- mm2[c(1:(1+length(grep(focal, colnms)))),] 
  for (i in c(fcts)){
    if (length(grep(i, colnms))>0) mm2[,grep(i, colnms)] <- 0
  }
  if (isfactor) {
    j=1
    for (i in grep(focal, colnms)){
      j=j+1
      mm2[j,i] <- 1
    }
  }
  
# bootstrap focal coefficent
zerocoef <- fixef(zeromod)
zeromat <- matrix(zerocoef, nrow=length(zerocoef), ncol=nb_btstrp+1, byrow=F)
ses <- sqrt(diag(as.matrix(vcov(zeromod))))
  if (!isfactor){
    ses <- ses[grep(focal, names(ses))]
    smpl <- c(0, rnorm(nb_btstrp, mean=0, sd=ses))
    zeromat[1,] <- zeromat[1,]+smpl
  } 
  if (isfactor){
    sess <- c(ses[1], ses[grep(focal, names(ses))])
    for (i in c(1:j)){
      if (i==1) smpl <- c(0, rnorm(nb_btstrp, mean=0, sd=sess[i]))
      if (i>1) smpl <- rbind(smpl, c(0, rnorm(nb_btstrp, mean=0, sd=sess[i])))
    }
    zeromat[c(1, grep(focal, names(ses))),] <- zeromat[c(1, grep(focal, names(ses))),]+smpl
  } 
  
require(boot)
predzero_smpl <- inv.logit(mm1 %*% zeromat) 
  
perccoef <- fixef(percmod)[[1]]
percmat <- matrix(perccoef, nrow=length(perccoef), ncol=nb_btstrp+1, byrow=F)
ses <- sqrt(diag(as.matrix(vcov(percmod))[[1]]))
  if (!isfactor){
    ses <- ses[grep(focal, names(ses))]
    smpl <- c(0, rnorm(nb_btstrp, mean=0, sd=ses))
    percmat[1,] <- percmat[1,]+smpl
  }
  if (isfactor){
    sess <- c(ses[1], ses[grep(focal, names(ses))])
    for (i in c(1:j)){
      if (i==1) smpl <- c(0, rnorm(nb_btstrp, mean=0, sd=sess[i]))
      if (i>1) smpl <- rbind(smpl, c(0, rnorm(nb_btstrp, mean=0, sd=sess[i])))
    }
    percmat[c(1, grep(focal, names(ses))),] <- percmat[c(1, grep(focal, names(ses))),]+smpl
  } 
  
predperc_smpl <- inv.logit(mm2 %*% percmat)*100 
  
pred_smpl <- predzero_smpl*predperc_smpl
  
  if (!isfactor){
    res <- data.frame(var=seq(rng[1], rng[2], length.out=nvals), point=pred_smpl[,1], mean=apply(pred_smpl, 1, mean), sd=apply(pred_smpl, 1, sd), 
                    CI025=apply(pred_smpl, 1, quantile, probs=0.025), CI975=apply(pred_smpl, 1, quantile, probs=0.975))
    if (plot) {
      plot(res[,"var"], res$point, type="l", lwd=2, xlab=focal, ylab="browsing pressure")
      lines(res[,"var"], res$CI975, col="dark grey")
      lines(res[,"var"], res$CI025, col="dark grey")
      lines(res[,"var"], res$point, col="black")
    }
  }
  
  if (isfactor){
    res <- data.frame(var=levels(dat[, grep(focal, names(dat))]), point=pred_smpl[,1], mean=apply(pred_smpl, 1, mean), sd=apply(pred_smpl, 1, sd), 
                      CI025=apply(pred_smpl, 1, quantile, probs=0.025), CI975=apply(pred_smpl, 1, quantile, probs=0.975))
    if (plot) {
      plot(c(1:nrow(res)), res$point, type="p", ylim=c(min(res$CI025), max(res$CI975)), xlab=focal, ylab="browsing pressure", xaxt = 'n')
      axis(1, at=c(1:nrow(res)), labels=res$var)
      segments(c(1:nrow(res)), res$CI975, c(1:nrow(res)), res$CI025)
    }
  }
  
  if (!plot) return(res)
}

```


```{r}
# load("browse9.2.rda")

browse9.2[,"treartgruppe9"] <- NULL

par(mfrow=c(3,3))
resdistvei <- resp_curve("distvei2_trunc", dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

resskogkat <- resp_curve("skogkategori", isfactor=T, dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

restreart <- resp_curve("treartgruppe", isfactor=T, dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

reskant <- resp_curve("kant", isfactor=T, dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

reshell <- resp_curve("helling_scale", dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

reshoh <- resp_curve("HOH_scale", dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

restretett <- resp_curve("tretetthet9_trunc_scale", dat=browse9.2, zeromod=m4abin, percmod=m4abeta, nvals=100, plot=F, nb_btstrp=999)

resdisthus <- resp_curve("disthus_trunc_cut200_scale", dat=browse9.2, zeromod=m1kbin, percmod=m1ibeta, nvals=100, plot=F, nb_btstrp=999)

```



```{r}
library(ggplot2)

p1 <- ggplot(data = resdisthus,aes(var, point)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = CI025, ymax = CI975), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nDistance to house") + 
  ylab("Browsing pressure (%)\n") +
  scale_y_continuous(limits = c(10, 30), breaks =  seq(from = 10, to = 30, by = 5))

p2 <- ggplot(data = resdistvei,aes(var, point)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = CI025, ymax = CI975), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.title.y = element_blank(), axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nDistance to road") + 
  ylab("Browsing pressure (%)\n") +
  scale_y_continuous(limits = c(10, 30), breaks =  seq(from = 10, to = 30, by = 5))


p3 <- ggplot(data = resskogkat,aes(var, point)) +
  geom_pointrange(aes(ymin = CI025, ymax = CI975), colour = "#009292", size = 1) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nForest category") + 
  ylab("Browsing pressure (%)\n") +
  scale_x_discrete(labels = c("Highproductive \nold forest", "Highproductive \nyoung forest", "Lowproductive \nold forest", "Lowproductive \nyoung forest", "Unproductive \nforest"))+
  scale_y_continuous(limits = c(5, 50), breaks =  seq(from = 5, to = 50, by = 10))


p4 <- ggplot(data = restreart,aes(var, point)) +
  geom_pointrange(aes(ymin = CI025, ymax = CI975), colour = "#009292", size = 1) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nTree species groups") + 
  ylab("Browsing pressure (%)\n") +
  scale_x_discrete(labels = c("Pine", "Other deciduous \ntrees", "Rowan, aspen, \nwillow"))+
  scale_y_continuous(limits = c(5, 50), breaks =  seq(from = 5, to = 50, by = 10))

p5 <- ggplot(data = reskant,aes(var, point)) +
  geom_pointrange(aes(ymin = CI025, ymax = CI975), colour = "#009292", size = 1) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nEdge") + 
  ylab("Browsing pressure (%)\n") +
  scale_x_discrete(labels = c("Edge <10m", "Edge >20m", "Edge 10-10m", "Edge missing"))+
  scale_y_continuous(limits = c(5, 50), breaks =  seq(from = 5, to = 50, by = 10))


p6 <- ggplot(data = reshell,aes(var, point)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = CI025, ymax = CI975), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.title.y = element_blank(), axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nSlope (%)") + 
  ylab("Browsing pressure (%)\n")+
  scale_y_continuous(limits = c(10, 30), breaks =  seq(from = 10, to = 30, by = 5))


p7 <- ggplot(data = reshoh,aes(var, point)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = CI025, ymax = CI975), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nMetres above sea level") + 
  ylab("Browsing pressure (%)\n")+
  scale_y_continuous(limits = c(10, 30), breaks =  seq(from = 10, to = 30, by = 5))


p8 <- ggplot(data = restretett,aes(var, point)) +
  geom_line(colour = "#009292", linetype = 1, size = 1)+
  geom_ribbon(aes(ymin = CI025, ymax = CI975), colour = "#004949", linetype = 2, alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 10),
        axis.title = element_text(family = "sans", size = 14, color = "grey30")) +
  xlab("\nTree density") + 
  ylab("Browsing pressure (%)\n")+
  scale_y_continuous(limits = c(10, 30), breaks =  seq(from = 10, to = 30, by = 5))

# library(cowplot)
# plot_grid(p1, p2, p3, p4, p5, p6, p7, p8,
#           nrow = 3, ncol = 3)

library(gridExtra)
grid.arrange(p4, p1, p2, p3, p7, p6, p5, p8,
          nrow = 3, ncol = 3)


```

"#004949", "#009292"
