==> 09_recruitment_hurdle.Rmd <==
---
title: "recruitment"
author: "Anne"
date: "27/11/2020"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```




########################################################################
## Generell test for zero inflation for recruitment response variable ##
#######################################################################
```{r}
load("processdata/recruits_2021_02_01.rda")

library(ggplot2)
library(ggpubr)

ggplot(recruits, aes(recruitment)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

str(recruits$recruitment)
summary(recruits$recruitment)



#### Test for zero-inflation

# Use glmmTMB package to run glm
library("glmmTMB")
m_test_zi <- glm(recruitment ~ disthus + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac, family = poisson, data = recruits)

library("performance")
check_zeroinflation(m_test_zi)

# # Check for zero-inflation
# 
#    Observed zeros: 13066
#   Predicted zeros: 4
#             Ratio: 0.00
# 
# Model is underfitting zeros (probable zero-inflation).



```




###################################################################################
##     HOUSE                HOUSE                 HOUSE                       #####
###################################################################################

##       House binary - First part of hurdle model to account for zero inflation  ##


## Bin house threshold distance to house

#### Load data house models
```{r}
load("processdata/recruits_2021_02_01.rda")
names(recruits)

recruits_house <- recruits

# load("processdata/recruits_house.rda")

names(recruits_house)
```


#### Scale variables if necessary
```{r}

recruits_house$disthus <- scale(recruits_house$disthus)

recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

recruits_house$disthus_400 <- scale(recruits_house$disthus_400)
recruits_house$disthus_350 <- scale(recruits_house$disthus_350)
recruits_house$disthus_300 <- scale(recruits_house$disthus_300)
recruits_house$disthus_200 <- scale(recruits_house$disthus_200)
recruits_house$disthus_150 <- scale(recruits_house$disthus_150)

summary(recruits_house$disthus)
```


#### Checking for the cut for disthus
```{r}
library(lme4)

bin_cuthouse1 <- glmer(factor(recruitment > 0) ~ disthus + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse2 <- glmer(factor(recruitment > 0) ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse3 <- glmer(factor(recruitment > 0) ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # BEST

bin_cuthouse4 <- glmer(factor(recruitment > 0) ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse5 <- glmer(factor(recruitment > 0) ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse6 <- glmer(factor(recruitment > 0) ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse7 <- glmer(factor(recruitment > 0) ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse8 <- glmer(factor(recruitment > 0) ~ disthus_150 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cuthouse1, bin_cuthouse2, bin_cuthouse3, bin_cuthouse4, bin_cuthouse5, bin_cuthouse6, bin_cuthouse7, bin_cuthouse8)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod3 19 22815.51      0.00  0.47   0.47 -11388.76
# Mod2 19 22815.58      0.07  0.46   0.93 -11388.79
# Mod4 19 22819.32      3.81  0.07   1.00 -11390.66
# Mod5 19 22825.61     10.10  0.00   1.00 -11393.81
# Mod6 19 22833.41     17.90  0.00   1.00 -11397.71
# Mod7 19 22847.68     32.17  0.00   1.00 -11404.84
# Mod8 19 22861.53     46.02  0.00   1.00 -11411.76
# Mod1 19 22886.53     71.02  0.00   1.00 -11424.27

```

#### Save recruits_house dataset
```{r}
str(recruits)
names(recruits)
recruits_house <- recruits[, -c(4,19:23)]
names(recruits_house)
save(recruits_house, file = "recruits_house_2021_02_02.rda")

```




## Bin house models

#### Load data
```{r}
load("processdata/recruits_house_2021_02_02.rda")
str(recruits_house)
recruits_house$recruitment_bin <- ifelse(recruits_house$recruitment == 0, 0, 1)

```

#### Scale variables if necessary
```{r}

recruits_house$disthus <- scale(recruits_house$disthus)

# recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

summary(recruits_house$helling)
```



```{r}
library(lme4)
rec_h_bin1 <- glmer(recruitment_bin ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)
summary(rec_h_bin1)

rec_h_bin2 <- glmer(recruitment_bin ~ disthus_500 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)


rec_h_bin2_600 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # Same as rec_h_bin2, only run with disthus_600, so that both recruits house models have the same disthus variables - easier in the function
summary(rec_h_bin2_600)

save(rec_h_bin2, file = "output/recruitment_2021_02_02/rec_h_bin2.rda")
save(rec_h_bin2_600, file = "output/recruitment_2021_02_02/rec_h_bin2_600.rda")



## Model with interaction for discussion
rec_h_bin2_600_int <- glmer(recruitment_bin ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # BEST
summary(rec_h_bin2_600_int)

rec_h_bin2_600_int2 <- glmer(recruitment_bin ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + tretetthet9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house) 
summary(rec_h_bin2_600_int2)

save(rec_h_bin2_600_int2, file = "output/recruitment_2021_02_02/rec_h_bin2_600_int2.rda")

```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_rec_house_bin <- list(rec_h_bin1, rec_h_bin2, rec_h_bin2_600, rec_h_bin2_600_int, rec_h_bin2_600_int2)
aictab(candmods_rec_house_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod5 21 21536.77      0.00     1      1 -10747.39
# Mod1 20 21557.53     20.76     0      1 -10758.77
# Mod4 20 22776.65   1239.88     0      1 -11368.33
# Mod2 19 22777.47   1240.70     0      1 -11369.73
# Mod3 19 22777.90   1241.13     0      1 -11369.95


candmods_rec_house_bin <- list(rec_h_bin2_600, rec_h_bin2_600_int)
aictab(candmods_rec_house_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod2 20 22776.65      0.00  0.65   0.65 -11368.33
# Mod1 19 22777.90      1.25  0.35   1.00 -11369.95


candmods_rec_house_bin <- list(rec_h_bin1, rec_h_bin2_600_int, rec_h_bin2_600_int2)
aictab(candmods_rec_house_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AIC:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod2 21 21536.77      0.00     1      1 -10747.39
# Mod1 20 22776.65   1239.88     0      1 -11368.33

```

#### Summary bin house
```{r}
summary(rec_h_bin2_600_int2)
# vcov(rec_h_bin5)
```



#### Plots bin house
```{r}
library(sjPlot)
plot_model(rec_h_bin2_600_int2, type = "diag")
plot_model(rec_h_bin2_600_int2, type = "eff")
plot_model(rec_h_bin2_600_int2, type = "est")

rec_h_bin2_600_int2 <- plot_model(rec_h_bin2_600_int2, type = "int", terms = "disthus_600 [all]:beitetrykk9", mdrt.values = "minmax")
# "minmax" -> (default) minimum and maximum values (lower and upper bounds) of the moderator are used to plot the interaction between independent variable and moderator(s).
# "meansd" -> uses the mean value of the moderator as well as one standard deviation below and above mean value to plot the effect of the moderator on the independent variable (following the convention suggested by Cohen and Cohen and popularized by Aiken and West (1991), i.e. using the mean, the value one standard deviation above, and the value one standard deviation below the mean as values of the moderator, see Grace-Martin K: 3 Tips to Make Interpreting Moderation Effects Easier).
# "quart" -> calculates and uses the quartiles (lower, median and upper) of the moderator value.



## TRY also performance::check_model for model assumptions and model performance summaries with performance::model_performance; to test performance between models use compare_performance (can be for glm, lm, lmerMod at the same time)
# library(performance)
# check_model(rec_h_bin7)
# model_performance(rec_h_bin7_600)
# compare_performance(rec_h_bin7_600, rec_h_bin7_600_test)

 
tab_model(rec_h_bin2_600_int2, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))


```

#### Save bin house
```{r}
save(rec_h_bin7_600, file = "rec_h_bin7_600.rda")

```




##       House beta - Second part of the hurdle, Poisson  ##


## Load data house models
```{r}

load("processdata/recruits_2021_02_01.rda")
names(recruits)

recruits_house <- recruits

str(recruits_house)

```



#### Prepare data set for second part of the hurdle
```{r}
recruits_house <- recruits_house[recruits_house$recruitment > 0, ]
str(recruits_house)
summary(recruits_house$recruitment)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)
```


#### Scale variables if necessary
```{r}
recruits_house$disthus <- scale(recruits_house$disthus)
recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

recruits_house$disthus_400 <- scale(recruits_house$disthus_400)
recruits_house$disthus_350 <- scale(recruits_house$disthus_350)
recruits_house$disthus_300 <- scale(recruits_house$disthus_300)
recruits_house$disthus_200 <- scale(recruits_house$disthus_200)
recruits_house$disthus_150 <- scale(recruits_house$disthus_150)

summary(recruits_house$disthus_600)
```

#### Checking for the cut for disthus
```{r}
library(lme4)


beta_cuthouse1 <- glmer(recruitment ~ disthus + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse2 <- glmer(recruitment ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) # BEST

beta_cuthouse3 <- glmer(recruitment ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

beta_cuthouse4 <- glmer(recruitment ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse5 <- glmer(recruitment ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse6 <- glmer(recruitment ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse7 <- glmer(recruitment ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse8 <- glmer(recruitment ~ disthus_150 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3, beta_cuthouse4, beta_cuthouse5, beta_cuthouse6, beta_cuthouse7, beta_cuthouse8)
# candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod2 19 1145906      0.00     1      1 -572933.9
# Mod3 19 1145926     20.57     0      1 -572944.2
# Mod4 19 1147153   1247.54     0      1 -573557.7
# Mod5 19 1147952   2045.83     0      1 -573956.8
# Mod6 19 1148915   3008.94     0      1 -574438.4
# Mod7 19 1150628   4722.33     0      1 -575295.1
# Mod8 19 1150812   4906.55     0      1 -575387.2
# Mod1 19 1153297   7391.18     0      1 -576629.5

```


#### Save recruits_house_beta dataset
```{r}
## !! Use recruits_house - all variables needed are in there !! ##



```


## Beta house models

#### Load models
```{r}
load("processdata/recruits_house.rda")
str(recruits_house)
names(recruits_house)
```


#### Prepare data set for second part of the hurdle
```{r}
recruits_house <- recruits_house[recruits_house$recruitment > 0, ]
str(recruits_house)
summary(recruits_house$recruitment)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)
```


#### Scale variables if necessary
```{r}
recruits_house$disthus <- scale(recruits_house$disthus)

# recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

# recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

# summary(recruits_house$disthus_600)
```


#### Beta house models
```{r}
library(lme4)
rec_h_beta1 <- glmer(recruitment ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)
summary(rec_h_beta1)

rec_h_beta2 <- glmer(recruitment ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

rec_h_beta3 <- glmer(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

rec_h_beta3_int <- glmer(recruitment ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

rec_h_beta3_int2 <- glmer(recruitment ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + tretetthet9 +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) # BEST


summary(rec_h_beta3)


library(glmmTMB)
rec_h_beta3_TMB <- glmmTMB(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
                           family = poisson,
                           data = recruits_house) # Same as rec_h_beta3, just run with glmmTMB
summary(rec_h_beta3_TMB)
save(rec_h_beta3_TMB, file = "output/recruitment_2021_02_02/rec_h_beta3_TMB.rda")



# Model with interaction for discussion
rec_h_beta3_TMB_int <- glmmTMB(recruitment ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
                           family = poisson,
                           data = recruits_house)
summary(rec_h_beta3_TMB_int)
save(rec_h_beta3_TMB_int, file = "output/recruitment_2021_02_02/rec_h_beta3_TMB_int.rda")


rec_h_beta3_TMB_int2 <- glmmTMB(recruitment ~ disthus_600*beitetrykk9 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + tretetthet9 +  ac + (1|KOMNR),
                           family = poisson,
                           data = recruits_house)

```


#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_rec_house_beta <- list(rec_h_beta1, rec_h_beta2, rec_h_beta3, rec_h_beta3_int, rec_h_beta3_int2)
# candmods_rec_house_beta <- list(rec_h_beta4, rec_h_beta7)
aictab(candmods_rec_house_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod5 20 1015966       0.0     1      1 -507963.1
# Mod4 19 1142979  127013.3     0      1 -571470.7
# Mod3 18 1143746  127779.9     0      1 -571855.0
# Mod2 18 1145904  129937.7     0      1 -572933.9
# Mod1 19 1145905  129939.3     0      1 -572933.7


candmods_rec_house_beta <- list(rec_h_beta3_TMB_int, rec_h_beta3_TMB_int2)
aictab(candmods_rec_house_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod2 20 1015966       0.0     1      1 -507963.1
# Mod1 19 1142979  127013.3     0      1 -571470.7


```


#### Summary beta house
```{r}
summary(rec_h_beta3_TMB_int2)
# vcov(rec_h_bin5)
```



#### Plots beta house
```{r}
library(sjPlot)
plot_model(rec_h_beta3_TMB, type = "diag")
plot_model(rec_h_beta3_TMB, type = "eff")
plot_model(rec_h_beta3_TMB, type = "est")

p_rec_h_beta3_TMB_int <- plot_model(rec_h_beta3_TMB_int, type = "int", mdrt.values = "minmax")


# library(performance)
# # check_model(rec_h_beta7_TMB)
# model_performance(rec_h_beta3_TMB)

tab_model(rec_h_beta3_TMB_int2, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))


library(ggplot2)
library(gridExtra)
rec_h_bin_grid <- grid.arrange(p_rec_h_bin2_600_int, p_rec_h_beta3_TMB_int,
          nrow = 1, ncol = 2)
ggsave(file = "rec_h_bin_grid.pdf", rec_h_bin_grid, width = 20, height = 10, units = "cm") #, dpi = 150

```


#### Save house beta model
```{r}
save(rec_h_beta3_TMB_int2, file = "output/recruitment_2021_02_02/rec_h_beta3_TMB_int2.rda")
```








###################################################################################
##     ROADS                ROADS                 ROADS                       #####
###################################################################################

##       Roads Bin - First part of a hurdle model to account for zero inflation  ##


## Bin roads

#### Load data roads models
```{r}
load("processdata/recruits_roads_2021_02_03.rda")

recruits_roads$recruitment_bin <- ifelse(recruits_roads$recruitment == 0, 0, 1)
str(recruits_roads)
names(recruits_roads)

```



#### Scale variables if necessary
```{r}
# recruits_roads$distvei2 <- scale(recruits_roads$distvei2)

# recruits_roads$beitetrykk9 <- scale(recruits_roads$beitetrykk9)
recruits_roads$helling <- scale(recruits_roads$helling)
recruits_roads$HOH <- scale(recruits_roads$HOH)
recruits_roads$tretetthet9 <- scale(recruits_roads$tretetthet9)
# recruits_roads$moose_density <- scale(recruits_roads$moose_density)

# summary(recruits_roads$beitetrykk9)
```



## Bin roads models
```{r}
library(lme4)
rec_r_bin1 <- glmer(recruitment_bin ~ distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_roads)
summary(rec_r_bin1) 

rec_r_bin2 <- glmer(recruitment_bin ~ distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_roads)

rec_r_bin3  <- glmer(recruitment_bin ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_roads)

rec_r_bin3_2  <- glmer(recruitment_bin ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 +  ac + (1|KOMNR),
family = binomial(),
data = recruits_roads)  # BEST



# Model with interaction for discussion
rec_r_bin3_int  <- glmer(recruitment_bin ~ distvei2*beitetrykk9 + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_roads)

summary(rec_r_bin3_int)
save(rec_r_bin3_int, file = "output/recruitment_2021_02_02/rec_r_bin3_int.rda")
```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_rec_roads_bin <- list(rec_r_bin1, rec_r_bin2, rec_r_bin3, rec_r_bin3_int, rec_r_bin3_2)
aictab(candmods_rec_roads_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt       LL
# Mod5 19 13169.07      0.00     1      1 -6565.54
# Mod3 18 13860.14    691.07     0      1 -6912.07
# Mod1 18 13874.89    705.81     0      1 -6919.44
# Mod4 19 13875.85    706.77     0      1 -6918.92
# Mod2 15 13884.79    715.71     0      1 -6927.39

save(rec_r_bin3_2, file = "output/recruitment_2021_02_02/rec_r_bin3_2.rda")


candmods_rec_roads_bin <- list(rec_r_bin3, rec_r_bin3_int)
aictab(candmods_rec_roads_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt       LL
# Mod1 18 13860.14       0.0     1      1 -6912.07
# Mod2 19 13875.85      15.7     0      1 -6918.92 # Interaction model worse

```

#### Summary bin roads
```{r}
summary(rec_r_bin3_2)
# vcov(rec_h_bin5)
```



#### Plots bin roads
```{r}
# library(performance)
# # check_model(rec_r_bin3)
# model_performance(rec_r_bin3)

library(sjPlot)
plot_model(rec_r_bin3_2, type = "diag")
plot_model(rec_r_bin3_2, type = "eff")
plot_model(rec_r_bin3_2, type = "est")

p_rec_r_bin3_int <- plot_model(rec_r_bin3_int, type = "int", mdrt.values = "minmax")
# "minmax" -> (default) minimum and maximum values (lower and upper bounds) of the moderator are used to plot the interaction between independent variable and moderator(s).
# "meansd" -> uses the mean value of the moderator as well as one standard deviation below and above mean value to plot the effect of the moderator on the independent variable (following the convention suggested by Cohen and Cohen and popularized by Aiken and West (1991), i.e. using the mean, the value one standard deviation above, and the value one standard deviation below the mean as values of the moderator, see Grace-Martin K: 3 Tips to Make Interpreting Moderation Effects Easier).
# "quart" -> calculates and uses the quartiles (lower, median and upper) of the moderator value.


tab_model(rec_r_bin3_2, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

```

#### Save bin roads
```{r}
save(rec_r_bin3_2, file = "output/recruitment_2021_02_02/rec_r_bin3_2.rda")

```




##       Roads beta - Second part of the hurdle, Poisson  ##


## Load data road models
```{r}
load("processdata/recruits_roads_2021_02_03.rda")

names(recruits_roads)
```


#### Prepare data set for second part of the hurdle
```{r}
recruits_roads <- recruits_roads[recruits_roads$recruitment > 0, ]
str(recruits_roads)
summary(recruits_roads$recruitment)

is.integer(recruits_roads$recruitment)
recruits_roads$recruitment <- as.integer(recruits_roads$recruitment)
is.integer(recruits_roads$recruitment)

```


#### Scale variables if necessary
```{r}

# recruits_roads$distvei2 <- scale(recruits_roads$distvei2)

# recruits_roads$beitetrykk9 <- scale(recruits_roads$beitetrykk9)
recruits_roads$helling <- scale(recruits_roads$helling)
recruits_roads$HOH <- scale(recruits_roads$HOH)
recruits_roads$tretetthet9 <- scale(recruits_roads$tretetthet9)
# recruits_roads$moose_density <- scale(recruits_roads$moose_density)


```


## Beta roads models
```{r}
library(lme4)
rec_r_beta1 <- glmer(recruitment ~ distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_roads)
summary(rec_r_beta1)

rec_r_beta2 <- glmer(recruitment ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_roads)

rec_r_beta2_2 <- glmer(recruitment ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 +  ac + (1|KOMNR),
family = poisson(),
data = recruits_roads)


library(glmmTMB)
rec_r_beta2_TMB <- glmmTMB(recruitment ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson,
data = recruits_roads)# BEST; Same as rec_r_beta2, just run with glmmTMB

summary(rec_r_beta2_TMB)
save(rec_r_beta2_TMB, file = "output/recruitment_2021_02_02/rec_r_beta2_TMB.rda")

rec_r_beta2_2_TMB <- glmmTMB(recruitment ~ log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 +  ac + (1|KOMNR),
family = poisson,
data = recruits_roads)


# Model with interaction for discussion
rec_r_beta2_TMB_int <- glmmTMB(recruitment ~ distvei2*beitetrykk9 + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson,
data = recruits_roads)

summary(rec_r_beta2_TMB_int)
save(rec_r_beta2_TMB_int, file = "output/recruitment_2021_02_02/rec_r_beta2_TMB_int.rda")
```




#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_rec_roads_beta <- list(rec_r_beta1, rec_r_beta2, rec_r_beta2_2)
aictab(candmods_rec_roads_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod3 19 426135.8      0.00     1      1 -213048.9
# Mod2 18 483623.1  57487.35     0      1 -241793.6
# Mod1 18 484601.0  58465.18     0      1 -242282.5

candmods_rec_roads_beta <- list(rec_r_beta2_TMB, rec_r_beta2_TMB_int)
aictab(candmods_rec_roads_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 18 483623.1      0.00     1      1 -241793.6
# Mod2 19 484459.5    836.42     0      1 -242210.8 # interaction model worse!

```


#### Summary beta roads
```{r}
summary(rec_r_beta2_2_TMB)
# vcov(rec_h_bin5)
```



#### Plots beta roads
```{r}
# library(effects)
# 
# plot(allEffects(rec_r_beta4))
# plot(effect("skogkategori", rec_r_beta4))

# library(performance)
# # check_model(rec_r_beta4)
# model_performance(rec_r_beta4_TMB)
# compare_performance(rec_r_beta4_TMB, rec_r_beta5_TMB, rank = T, bayesfactor = F)



library(sjPlot)
plot_model(rec_r_beta2_2_TMB, type = "diag")
plot_model(rec_r_beta2_2_TMB, type = "eff") 
plot_model(rec_r_beta2_2_TMB, type = "est")


p_rec_r_beta2_TMB_int <- plot_model(rec_r_beta2_TMB_int, type = "int", mdrt.values = "minmax")
# "minmax" -> (default) minimum and maximum values (lower and upper bounds) of the moderator are used to plot the interaction between independent variable and moderator(s).
# "meansd" -> uses the mean value of the moderator as well as one standard deviation below and above mean value to plot the effect of the moderator on the independent variable (following the convention suggested by Cohen and Cohen and popularized by Aiken and West (1991), i.e. using the mean, the value one standard deviation above, and the value one standard deviation below the mean as values of the moderator, see Grace-Martin K: 3 Tips to Make Interpreting Moderation Effects Easier).
# "quart" -> calculates and uses the quartiles (lower, median and upper) of the moderator value.


tab_model(rec_r_beta2_2_TMB, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))


library(ggplot2)
library(gridExtra)
rec_r_grid <- grid.arrange(p_rec_r_bin3_int, p_rec_r_beta2_TMB_int,
          nrow = 1, ncol = 2)
ggsave(file = "rec_r_grid.tiff", rec_r_grid, width = 20, height = 10, units = "cm", dpi = 150) #, dpi = 150

```


#### Save roads beta model
```{r}
save(rec_r_beta2_2_TMB, file = "output/recruitment_2021_02_02/rec_r_beta2_2_TMB.rda")
```




## Figures bin
```{r}
library(sjPlot)
library(ggplot2)

## House
load("processdata/recruits_house_2021_02_02.rda")
str(recruits_house) # Check if variables are scales - if not scale
# load("output/recruitment_2021_02_02/rec_h_bin2_600_int.rda")
load("output/recruitment_2021_03_25/rec_h_bin2_600_int2.rda")
recruits_house$recruitment_bin <- ifelse(recruits_house$recruitment == 0, 0, 1)
summary(rec_h_bin2_600_int2)

# scale
recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
# recruits_house$moose_density <- scale(recruits_house$moose_density)
library(plyr)
levels(recruits_house$skogkategori)
recruits_house$skogkategori <- mapvalues(recruits_house$skogkategori, from =  c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog", "Lavproduktiv_ungskog", "Uproduktiv_skog"), to = c("Lowproductive_old_forest",  "Highproductive_old_forest", "Highproductive_young_forest", "Lowproductive_young_forest", "Unproductive_forest"))
levels(recruits_house$skogkategori)

levels(recruits_house$kant)
recruits_house$kant <- mapvalues(recruits_house$kant, from =  c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant"), to = c("Edge >20m", "Edge <10m", "Edge 10-20m", "Edge information missing"))
levels(recruits_house$kant)

levels(recruits_house$treartgruppe9)
recruits_house$treartgruppe9 <- mapvalues(recruits_house$treartgruppe9, from =  c("FURU", "LAUV", "ROS"), to = c("Pine", "Deciduous trees", "RAW"))
levels(recruits_house$treartgruppe9)



p_rec_bin1 <- plot_model(rec_h_bin2_600_int2, type = c("int"), transform = NULL, terms = c("disthus_600 [all]"))
p_rec_bin1 <- p_rec_bin1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Recruitment\n")

p_rec_bin2 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("skogkategori"))
p_rec_bin2 <- p_rec_bin2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 40, vjust = 1, hjust = 0.8)) +
  xlab("Forest category\n") +
  ylab("Recruitment\n") 


p_rec_bin3 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p_rec_bin3 <- p_rec_bin3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 15, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Recruitment\n") 


p_rec_bin4 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("kant"))
p_rec_bin4 <- p_rec_bin4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 25, vjust = 1, hjust = 0.8)) +
  xlab("Edge\n") +
  ylab("Recruitment\n")


p_rec_bin5 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("helling [all]"))
p_rec_bin5 <- p_rec_bin5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Recruitment\n")


p_rec_bin6 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
p_rec_bin6 <- p_rec_bin6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Recruitment\n")


p_rec_bin7 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("moose_density [all]"))
p_rec_bin7 <- p_rec_bin7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Moose density\n") +
  ylab("Recruitment\n")


p_rec_bin8 <- plot_model(rec_h_bin2_600_int2, type = c("eff"), transform = NULL, terms = c("beitetrykk9 [all]"))
p_rec_bin8 <- p_rec_bin8 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Browsing pressure\n") +
  ylab("Recruitment\n")



## Roads
load("processdata/recruits_roads_2021_02_03.rda")
str(recruits_roads) # Check if variables are scales - if not scale
# load("output/recruitment_2021_02_02/rec_r_bin3.rda")
load("output/recruitment_2021_03_25/rec_r_bin3_2.rda")

recruits_roads$recruitment_bin <- ifelse(recruits_roads$recruitment == 0, 0, 1)
is.integer(recruits_roads$recruitment)
recruits_roads$recruitment <- as.integer(recruits_roads$recruitment)
is.integer(recruits_roads$recruitment)
# Scale
# recruits_roads$beitetrykk9 <- scale(recruits_roads$beitetrykk9)
recruits_roads$helling <- scale(recruits_roads$helling)
recruits_roads$HOH <- scale(recruits_roads$HOH)
recruits_roads$tretetthet9 <- scale(recruits_roads$tretetthet9)



p_rec_bin1_road <- plot_model(rec_r_bin3_2, type = c("eff"), transform = NULL, terms = c("distvei2"))
p_rec_bin1_road <- p_rec_bin1_road + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to road\n") +
  ylab("Recruitment\n")


# p_rec_bin2 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("skogkategori"))
# p_rec_bin2 <- p_rec_bin2 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 12, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30"),
#     axis.text.x = element_text(angle = 40, vjust = 1, hjust = 1)) +
#   xlab("Forest category\n") +
#   ylab("Recruitment\n") 
# 
# 
# p_rec_bin3 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
# p_rec_bin3 <- p_rec_bin3 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Tree species group\n") +
#   ylab("Recruitment\n") 
# 
# 
# p_rec_bin4 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("kant"))
# p_rec_bin4 <- p_rec_bin4 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 12, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30"),
#     axis.text.x = element_text(angle = 20, vjust = 1, hjust = 1)) +
#   xlab("Edge\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_bin5 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("helling [all]"))
# p_rec_bin5 <- p_rec_bin5 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Slope\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_bin6 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
# p_rec_bin6 <- p_rec_bin6 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Altitude\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_bin7 <- plot_model(rec_r_bin32, type = c("eff"), transform = NULL, terms = c("moose_density [all]"))
# p_rec_bin7 <- p_rec_bin7 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Moose density\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_bin8 <- plot_model(rec_r_bin3, type = c("eff"), transform = NULL, terms = c("beitetrykk9 [all]"))
# p_rec_bin8 <- p_rec_bin8 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Browsing pressure\n") +
#   ylab("Recruitment\n")




library(gridExtra)
rec_eff_plot_1_bin <- grid.arrange(p_rec_bin1, p_rec_bin1_road, p_rec_bin2, p_rec_bin3,
          nrow = 2, ncol = 2)

rec_eff_plot_2_bin <- grid.arrange(p_rec_bin8, p_rec_bin4, p_rec_bin5, p_rec_bin6, p_rec_bin7,
          nrow = 3, ncol = 2)

# type = "pred" = Predicted values (marginal effects) for specific model terms. See ggpredict for details.
# type = "eff" = Similar to type = "pred", however, discrete predictors are held constant at their proportions (not reference level). See ggeffect for details. ggeffect() computes marginal effects by internally calling Effect. Effect constructs an "eff" object for a term (usually a high-order term), absorbing the lower-order terms marginal to the term in question, and averaging over other terms in the model.

ggsave(file = "rec_eff_plot_1_bin.tiff", rec_eff_plot_1_bin, width = 30, height = 30, units = "cm", dpi = 150) #, dpi = 150
ggsave(file = "rec_eff_plot_2_bin.tiff", rec_eff_plot_2_bin, width = 30, height = 30, units = "cm", dpi = 150) #, dpi = 150

```



## Figures beta
```{r}
library(sjPlot)
library(ggplot2)

## House
load("processdata/recruits_house_2021_02_02.rda")
str(recruits_house)
# load("output/recruitment_2021_02_02/rec_h_beta3_TMB_int.rda")
load("output/recruitment_2021_03_25/rec_h_beta3_TMB_int2.rda")

# recruits_house <- recruits_house[ ,-c(7)]
recruits_house$recruitment_bin <- ifelse(recruits_house$recruitment == 0, 0, 1)
# scale
# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)
# recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
# recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)

library(plyr)
levels(recruits_house$skogkategori)
recruits_house$skogkategori <- mapvalues(recruits_house$skogkategori, from =  c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog", "Lavproduktiv_ungskog", "Uproduktiv_skog"), to = c("Lowproductive_old_forest",  "Highproductive_old_forest", "Highproductive_young_forest", "Lowproductive_young_forest", "Unproductive_forest"))
levels(recruits_house$skogkategori)

levels(recruits_house$kant)
recruits_house$kant <- mapvalues(recruits_house$kant, from =  c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant"), to = c("Edge >20m", "Edge <10m", "Edge 10-20m", "Edge information missing"))
levels(recruits_house$kant)

levels(recruits_house$treartgruppe9)
recruits_house$treartgruppe9 <- mapvalues(recruits_house$treartgruppe9, from =  c("FURU", "LAUV", "ROS"), to = c("Pine", "Deciduous trees", "RAW"))
levels(recruits_house$treartgruppe9)



p_rec_beta1 <- plot_model(rec_h_beta3_TMB_int2, type = c("int"), transform = NULL, terms = c("disthus_600"))
p_rec_beta1 <- p_rec_beta1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("Recruitment\n")

p_rec_beta2 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("skogkategori"))
p_rec_beta2 <- p_rec_beta2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30, vjust = 1, hjust = 0.8)) +
  xlab("Forest category\n") +
  ylab("Recruitment\n") 


p_rec_beta3 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p_rec_beta3 <- p_rec_beta3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 15, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("Recruitment\n") 


p_rec_beta4 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("kant"))
p_rec_beta4 <- p_rec_beta4 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
    axis.text.x = element_text(angle = 30, vjust = 1, hjust = 0.8)) +
  xlab("Edge\n") +
  ylab("Recruitment\n")


p_rec_beta5 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("helling [all]"))
p_rec_beta5 <- p_rec_beta5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("Recruitment\n")


p_rec_beta6 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
p_rec_beta6 <- p_rec_beta6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("Recruitment\n")


p_rec_beta7 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("beitetrykk9 [all]"))
p_rec_beta7 <- p_rec_beta7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Browsing pressure\n") +
  ylab("Recruitment\n")


# p_rec_beta8 <- plot_model(rec_h_beta3_TMB_int2, type = c("eff"), transform = NULL, terms = c("moose_density"))
# p_rec_beta8 <- p_rec_beta8 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Moose density\n") +
#   ylab("Recruitment\n")

# p_rec_test <- plot_model(rec_h_beta7_TMB, type = c("eff"), transform = NULL, terms = c("disthus_600"))
# p_rec_test <- p_rec_test + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Distance to houses\n") +
#   ylab("Recruitment\n")


## Roads
load("processdata/recruits_roads_2021_02_03.rda")
str(recruits_roads)
# load("output/recruitment_2021_02_02/rec_r_beta2_TMB.rda")
load("output/recruitment_2021_03_25/rec_r_beta2_2_TMB.rda")
summary(rec_r_beta2_2_TMB)

recruits_roads$recruitment_bin <- ifelse(recruits_roads$recruitment == 0, 0, 1)
is.integer(recruits_roads$recruitment)
recruits_roads$recruitment <- as.integer(recruits_roads$recruitment)
is.integer(recruits_roads$recruitment)
# Scale
# recruits_roads$beitetrykk9 <- scale(recruits_roads$beitetrykk9)
recruits_roads$helling <- scale(recruits_roads$helling)
recruits_roads$HOH <- scale(recruits_roads$HOH)
recruits_roads$tretetthet9 <- scale(recruits_roads$tretetthet9)



p_rec_beta1_road <- plot_model(rec_r_beta2_2_TMB, type = c("eff"), transform = NULL, terms = c("distvei2  [0:500]"))
p_rec_beta1_road <- p_rec_beta1_road + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to road\n") +
  ylab("Recruitment\n")


# p_rec_beta2 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("skogkategori"))
# p_rec_beta2 <- p_rec_beta2 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 12, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30"),
#     axis.text.x = element_text(angle = 20, vjust = 1, hjust = 1)) +
#   xlab("Forest category\n") +
#   ylab("Recruitment\n") 
# 
# 
# p_rec_beta3 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
# p_rec_beta3 <- p_rec_beta3 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Tree species group\n") +
#   ylab("Recruitment\n") 
# 
# 
# p_rec_beta4 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("kant"))
# p_rec_beta4 <- p_rec_beta4 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 12, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30"),
#     axis.text.x = element_text(angle = 20, vjust = 1, hjust = 1)) +
#   xlab("Edge\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_beta5 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("helling [all]"))
# p_rec_beta5 <- p_rec_beta5 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Slope\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_beta6 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("HOH [all]"))
# p_rec_beta6 <- p_rec_beta6 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Altitude\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_beta7 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("beitetrykk9 [all]"))
# p_rec_beta7 <- p_rec_beta7 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Browsing pressure\n") +
#   ylab("Recruitment\n")
# 
# 
# p_rec_beta8 <- plot_model(rec_r_beta2_TMB, type = c("eff"), transform = NULL, terms = c("moose_density"))
# p_rec_beta8 <- p_rec_beta8 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("Moose density\n") +
#   ylab("Recruitment\n")



library(gridExtra)
rec_eff_plot_1_beta <- grid.arrange(p_rec_beta1, p_rec_beta1_road, p_rec_beta2, p_rec_beta3, p_rec_beta7, p_rec_beta4, p_rec_beta5, p_rec_beta6,
          nrow = 4, ncol = 2)

# rec_eff_plot_2_beta <- grid.arrange(p_rec_beta7, p_rec_beta4, p_rec_beta5, p_rec_beta6,
#           nrow = 3, ncol = 2)


ggsave(file = "rec_eff_plot_1_beta.pdf", rec_eff_plot_1_beta, width = 30, height = 45, units = "cm") #, dpi = 150
# ggsave(file = "rec_eff_plot_2_beta.pdf", rec_eff_plot_2_beta, width = 30, height = 30, units = "cm") #, dpi = 150

```
