==> 09_recruitment_hurdle_remake.Rmd <==
---
title: "09_recruitment_hurdle_remake"
author: "Anne"
date: "12/04/2021"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


########################################################################
## Generell test for zero inflation for recruitment response variable ##
#######################################################################
```{r}
load("processdata/recruits_2021_02_01.rda")

library(ggplot2)
library(ggpubr)

ggplot(recruits, aes(recruitment)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

str(recruits$recruitment)
summary(recruits$recruitment)



#### Test for zero-inflation

# Use glmmTMB package to run glm
library("glmmTMB")
m_test_zi <- glm(recruitment ~ disthus + distvei2 + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac, family = poisson, data = recruits)

library("performance")
check_zeroinflation(m_test_zi)

# # Check for zero-inflation
# 
#    Observed zeros: 13066
#   Predicted zeros: 4
#             Ratio: 0.00
# 
# Model is underfitting zeros (probable zero-inflation).



```


##       BINARY - First part of hurdle model to account for zero inflation  ##


## Bin threshold distance to house

#### Load data house threshold
```{r}
load("processdata/recruits_2021_04_19.rda")
names(recruits)

recruits_house <- recruits

# load("processdata/recruits_house.rda")

names(recruits_house)
```


#### Scale variables if necessary
```{r}

recruits_house$disthus <- scale(recruits_house$disthus)

recruits_house$distvei2 <- scale(recruits_house$distvei2)

recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

recruits_house$disthus_400 <- scale(recruits_house$disthus_400)
recruits_house$disthus_350 <- scale(recruits_house$disthus_350)
recruits_house$disthus_300 <- scale(recruits_house$disthus_300)
recruits_house$disthus_200 <- scale(recruits_house$disthus_200)
recruits_house$disthus_150 <- scale(recruits_house$disthus_150)

summary(recruits_house$disthus)
```


#### Checking for the cut for disthus
```{r}
library(lme4)

bin_cuthouse1 <- glmer(factor(recruitment > 0) ~ disthus + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse2 <- glmer(factor(recruitment > 0) ~ disthus_600 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse3 <- glmer(factor(recruitment > 0) ~ disthus_500 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # BEST

bin_cuthouse4 <- glmer(factor(recruitment > 0) ~ disthus_400 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse5 <- glmer(factor(recruitment > 0) ~ disthus_350 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse6 <- glmer(factor(recruitment > 0) ~ disthus_300 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse7 <- glmer(factor(recruitment > 0) ~ disthus_200 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cuthouse8 <- glmer(factor(recruitment > 0) ~ disthus_150 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cuthouse1, bin_cuthouse2, bin_cuthouse3, bin_cuthouse4, bin_cuthouse5, bin_cuthouse6, bin_cuthouse7, bin_cuthouse8)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod3 20 22798.08      0.00  0.48   0.48 -11379.04
# Mod2 20 22798.23      0.14  0.45   0.92 -11379.11
# Mod4 20 22801.88      3.79  0.07   1.00 -11380.94
# Mod5 20 22808.16     10.08  0.00   1.00 -11384.08
# Mod6 20 22815.91     17.83  0.00   1.00 -11387.95
# Mod7 20 22829.85     31.77  0.00   1.00 -11394.92
# Mod8 20 22843.69     45.61  0.00   1.00 -11401.84
# Mod1 20 22868.69     70.61  0.00   1.00 -11414.35

```









## Bin threshold distance to road

#### Load data road threshold
```{r}
load("processdata/recruits_2021_04_19.rda")
names(recruits)

recruits_house <- recruits

# load("processdata/recruits_house.rda")

names(recruits_house)
str(recruits_house)
```


#### Scale variables if necessary
```{r}

recruits_house$disthus <- scale(recruits_house$disthus)

recruits_house$distvei2 <- scale(recruits_house$distvei2)

recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$distvei_600 <- scale(recruits_house$distvei_600)

recruits_house$distvei_500 <- scale(recruits_house$distvei_500)

recruits_house$distvei_400 <- scale(recruits_house$distvei_400)
recruits_house$distvei_350 <- scale(recruits_house$distvei_350)
recruits_house$distvei_300 <- scale(recruits_house$distvei_300)
recruits_house$distvei_200 <- scale(recruits_house$distvei_200)
recruits_house$distvei_150 <- scale(recruits_house$distvei_150)
recruits_house$distvei_100 <- scale(recruits_house$distvei_100)
recruits_house$distvei_50 <- scale(recruits_house$distvei_50)

summary(recruits_house$disthus)
```


#### Checking for the cut for distvei
```{r}
library(lme4)

bin_cutroad1 <- glmer(factor(recruitment > 0) ~ disthus + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cutroad2 <- glmer(factor(recruitment > 0) ~ distvei_600 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cutroad3 <- glmer(factor(recruitment > 0) ~ distvei_500 + beitetrykk9 + disthus + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) 

bin_cutroad4 <- glmer(factor(recruitment > 0) ~ distvei_400 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # BEST

bin_cutroad5 <- glmer(factor(recruitment > 0) ~ distvei_350 + beitetrykk9*skogkategori + disthus*skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cutroad6 <- glmer(factor(recruitment > 0) ~ distvei_300 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cutroad7 <- glmer(factor(recruitment > 0) ~ distvei_200 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

bin_cutroad8 <- glmer(factor(recruitment > 0) ~ distvei_150 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

```

#### AIC selection threshold roads
```{r}
library(AICcmodavg)

candmods_bin_cuthouse <- list(bin_cutroad1, bin_cutroad2, bin_cutroad3, bin_cutroad4, bin_cutroad5, bin_cutroad6, bin_cutroad7, bin_cutroad8)
aictab(candmods_bin_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
# Model selection based on AICc:
# 
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod4 20 22780.32      0.00  0.48   0.48 -11370.16
# Mod3 20 22780.73      0.41  0.39   0.87 -11370.36
# Mod6 20 22784.41      4.09  0.06   0.93 -11372.21
# Mod5 19 22784.97      4.65  0.05   0.97 -11373.49
# Mod2 20 22787.23      6.91  0.02   0.99 -11373.62
# Mod7 20 22787.96      7.64  0.01   1.00 -11373.98
# Mod8 20 22795.27     14.95  0.00   1.00 -11377.63
# Mod1 20 22868.69     88.37  0.00   1.00 -11414.35

```


```{r}
str(recruits_house)
recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

distveilog_bin <- glmer(factor(recruitment > 0) ~ disthus_500 + skogkategori + log(distvei2+1) + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # BEST

bin_cutroad4 <- glmer(factor(recruitment > 0) ~ distvei_400 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)


candmods_bin_cutroad <- list(distveilog_bin, bin_cutroad4)
aictab(candmods_bin_cutroad, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 20 22759.27      0.00     1      1 -11359.64
# Mod2 20 22780.32     21.05     0      1 -11370.16

save(distveilog_bin, file = "output/log_or_thresh/distveilog_bin.rda")
save(bin_cutroad4, file = "output/log_or_thresh/bin_cutroad4.rda")

```


#### Save recruits_house dataset
```{r}
load("processdata/recruits_2021_04_19.rda")
str(recruits)
names(recruits)
recruits_house <- recruits[, -c(4,20:33)]
names(recruits_house)
save(recruits_house, file = "processdata/recruits_house_2021_04_19.rda")

```



## Bin house models

#### Load data
```{r}
# load("processdata/recruits_house_2021_02_02.rda")
load("processdata/recruits_house_2021_04_19.rda")
str(recruits_house)
recruits_house$recruitment_bin <- ifelse(recruits_house$recruitment == 0, 0, 1)

```

#### Scale variables if necessary
```{r}

# recruits_house$disthus <- scale(recruits_house$disthus)

# recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
# recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)

# recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
# recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

summary(recruits_house$helling)
```



```{r}
library(lme4)
rec_h_bin1 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # disthus_500 is best, but adjust to zero-trunc model with disthus_600
rec_h_bin1_1 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # without browsing pressure

## test for interactions ----
rec_h_bin2 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1) + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) 

rec_h_bin3 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1) + skogkategori + beitetrykk9*treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

rec_h_bin4 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1)*skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

rec_h_bin5 <- glmer(recruitment_bin ~ disthus_600*skogkategori + log(distvei2+1) + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

rec_h_bin6 <- glmer(recruitment_bin ~ disthus_600*skogkategori + log(distvei2+1) + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # best

rec_h_bin7 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1)*skogkategori + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) # Take this so the interaction is the same as for the beta model

rec_h_bin8 <- glmer(recruitment_bin ~ disthus_600*skogkategori + log(distvei2+1) + beitetrykk9*treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house) 

rec_h_bin9 <- glmer(recruitment_bin ~ disthus_600 + log(distvei2+1)*skogkategori + beitetrykk9*treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)


rec_h_bin10 <- glmer(recruitment_bin ~ disthus_600*skogkategori + log(distvei2+1) + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + ac + (1|KOMNR),
family = binomial(),
data = recruits_house)

```

#### AIC selection bin models
```{r}
library(AICcmodavg)

candmods_rec_house_bin <- list(rec_h_bin1, rec_h_bin2, rec_h_bin3, rec_h_bin4, rec_h_bin5, rec_h_bin6, rec_h_bin7, rec_h_bin8, rec_h_bin9)
aictab(candmods_rec_house_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod6 28 22731.23      0.00  0.50   0.50 -11337.61
# Mod7 28 22731.62      0.39  0.41   0.92 -11337.81
# Mod2 24 22734.83      3.61  0.08   1.00 -11343.42
# Mod8 26 22743.94     12.71  0.00   1.00 -11345.97
# Mod9 26 22744.84     13.62  0.00   1.00 -11346.42
# Mod3 22 22747.01     15.79  0.00   1.00 -11351.51
# Mod5 24 22756.59     25.37  0.00   1.00 -11354.30
# Mod4 24 22757.67     26.45  0.00   1.00 -11354.84
# Mod1 20 22759.77     28.55  0.00   1.00 -11359.89

candmods_rec_house_bin <- list(rec_h_bin6, rec_h_bin10)
aictab(candmods_rec_house_bin, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 28 22731.23       0.0  0.67   0.67 -11337.61
# Mod2 27 22732.62       1.4  0.33   1.00 -11339.31



save(rec_h_bin1, file = "output/2021_04_28/rec_h_bin1.rda")
save(rec_h_bin1_1, file = "output/2021_04_28/rec_h_bin1_1.rda")
save(rec_h_bin2, file = "output/2021_04_28/rec_h_bin2.rda")
save(rec_h_bin3, file = "output/2021_04_28/rec_h_bin3.rda")
save(rec_h_bin4, file = "output/2021_04_28/rec_h_bin4.rda")
save(rec_h_bin5, file = "output/2021_04_28/rec_h_bin5.rda")
save(rec_h_bin6, file = "output/2021_04_28/rec_h_bin6.rda")
save(rec_h_bin7, file = "output/2021_04_28/rec_h_bin7.rda")
save(rec_h_bin8, file = "output/2021_04_28/rec_h_bin8.rda")
save(rec_h_bin9, file = "output/2021_04_28/rec_h_bin9.rda")
save(rec_h_bin10, file = "output/2021_04_28/rec_h_bin10.rda")

summary(rec_h_bin6)
summary(rec_h_bin1)
```

#### Summary bin house
```{r}
summary(rec_h_bin6)
```



#### Plots bin house
```{r}
library(sjPlot)
plot_model(rec_h_bin6, type = "diag")
plot_model(rec_h_bin6, type = "eff")
plot_model(rec_h_bin6, type = "est")

tab_model(rec_h_bin6, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))


plot_model(rec_h_bin6, type = "eff", terms = c("disthus_600", "skogkategori"))
plot_model(rec_h_bin6, type = "eff", terms = c("beitetrykk9", "skogkategori"))


```

#### Save bin house
```{r}
save(rec_h_bin6, file = "output/output/2021_04_28/rec_h_bin6.rda")

```






##       ZERO-TRUNCATED - Second part of the hurdle, Poisson  ##


## Load data house threshold
```{r}

load("processdata/recruits_2021_04_19.rda")
names(recruits)

recruits_house <- recruits

str(recruits_house)

```



#### Prepare data set for second part of the hurdle
```{r}
recruits_house <- recruits_house[recruits_house$recruitment > 0, ]
str(recruits_house)
summary(recruits_house$recruitment)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)
```


#### Scale variables if necessary
```{r}
recruits_house$disthus <- scale(recruits_house$disthus)
recruits_house$distvei2 <- scale(recruits_house$distvei2)

recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

recruits_house$disthus_400 <- scale(recruits_house$disthus_400)
recruits_house$disthus_350 <- scale(recruits_house$disthus_350)
recruits_house$disthus_300 <- scale(recruits_house$disthus_300)
recruits_house$disthus_200 <- scale(recruits_house$disthus_200)
recruits_house$disthus_150 <- scale(recruits_house$disthus_150)

summary(recruits_house$disthus_600)
```

#### Checking for the threshold for disthus
```{r}
library(lme4)


beta_cuthouse1 <- glmer(recruitment ~ disthus + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse2 <- glmer(recruitment ~ disthus_600 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

beta_cuthouse3 <- glmer(recruitment ~ disthus_500 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house) # BEST

beta_cuthouse4 <- glmer(recruitment ~ disthus_400 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse5 <- glmer(recruitment ~ disthus_350 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse6 <- glmer(recruitment ~ disthus_300 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse7 <- glmer(recruitment ~ disthus_200 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cuthouse8 <- glmer(recruitment ~ disthus_150 + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

```

#### AIC selection threshold house
```{r}
library(AICcmodavg)

candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3, beta_cuthouse4, beta_cuthouse5, beta_cuthouse6, beta_cuthouse7, beta_cuthouse8)
# candmods_beta_cuthouse <- list(beta_cuthouse1, beta_cuthouse2, beta_cuthouse3)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
# Model selection based on AICc:
# 
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# 600  19 1145906      0.00     1      1 -572933.9
# 500  19 1145926     20.57     0      1 -572944.2
# 400  19 1147153   1247.54     0      1 -573557.7
# 350  19 1147952   2045.83     0      1 -573956.8
# 300  19 1148915   3008.94     0      1 -574438.4
# 200  19 1150628   4722.33     0      1 -575295.1
# 150  19 1150812   4906.55     0      1 -575387.2
# full 19 1153297   7391.18     0      1 -576629.5

```







## Load data road threshold
```{r}

load("processdata/recruits_2021_04_19.rda")
names(recruits)

recruits_house <- recruits

str(recruits_house)

```



#### Prepare data set for second part of the hurdle
```{r}
recruits_house <- recruits_house[recruits_house$recruitment > 0, ]
str(recruits_house)
summary(recruits_house$recruitment)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)
```


#### Scale variables if necessary
```{r}
recruits_house$disthus <- scale(recruits_house$disthus)
recruits_house$distvei2 <- scale(recruits_house$distvei2)

recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
recruits_house$moose_density <- scale(recruits_house$moose_density)

recruits_house$distvei_600 <- scale(recruits_house$distvei_600)

recruits_house$distvei_500 <- scale(recruits_house$distvei_500)

recruits_house$distvei_400 <- scale(recruits_house$distvei_400)
recruits_house$distvei_350 <- scale(recruits_house$distvei_350)
recruits_house$distvei_300 <- scale(recruits_house$distvei_300)
recruits_house$distvei_200 <- scale(recruits_house$distvei_200)
recruits_house$distvei_150 <- scale(recruits_house$distvei_150)
recruits_house$distvei_100 <- scale(recruits_house$distvei_100)
recruits_house$distvei_50 <- scale(recruits_house$distvei_50)

# summary(recruits_house$disthus_600)
```

#### Checking for the threshold for distvei
```{r}
library(lme4)


beta_cutroad1 <- glmer(recruitment ~ disthus + skogkategori + distvei2 + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad2 <- glmer(recruitment ~ distvei_600 + disthus + skogkategori + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) # BEST

beta_cutroad3 <- glmer(recruitment ~ distvei_500 + disthus + skogkategori + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

beta_cutroad4 <- glmer(recruitment ~ distvei_400 + disthus + skogkategori + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad5 <- glmer(recruitment ~ distvei_350 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad6 <- glmer(recruitment ~ distvei_300 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad7 <- glmer(recruitment ~ distvei_200 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad8 <- glmer(recruitment ~ distvei_150 + beitetrykk9 + skogkategori + disthus + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

```

#### AIC selection threshold roads
```{r}
library(AICcmodavg)

candmods_beta_cuthouse <- list(beta_cutroad1, beta_cutroad2, beta_cutroad3, beta_cutroad4, beta_cutroad5, beta_cutroad6, beta_cutroad7, beta_cutroad8)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

# Model selection based on AICc:
# 
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod2 19 1149435      0.00     1      1 -574698.5
# Mod3 19 1149528     92.58     0      1 -574744.8
# Mod4 19 1149978    543.38     0      1 -574970.2
# Mod8 19 1150002    566.82     0      1 -574981.9
# Mod5 19 1150241    805.76     0      1 -575101.4
# Mod7 19 1150421    986.39     0      1 -575191.7
# Mod6 19 1150859   1424.14     0      1 -575410.6
# Mod1 19 1153297   3861.93     0      1 -576629.5

```


```{r}
recruits_house$disthus_600 <- scale(recruits_house$disthus_600)

beta_cutroad_log <- glmer(recruitment ~ disthus_500 + skogkategori + log(distvei2+1) + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + tretetthet9 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

beta_cutroad2 <- glmer(recruitment ~ distvei_600 + disthus + skogkategori + beitetrykk9 + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density +  ac + (1|KOMNR),
family = poisson(),
data = recruits_house)


candmods_beta_cuthouse <- list(beta_cutroad_log, beta_cutroad2)
aictab(candmods_beta_cuthouse, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod1 19 1016400       0.0     1      1 -508181.1
# Mod2 19 1149435  133034.9     0      1 -574698.5


save(beta_cutroad2_log, file = "output/log_or_thresh/beta_cutroad2_log.rda")
save(beta_cutroad2, file = "output/log_or_thresh/beta_cutroad2.rda")

```






## Zero-truncated house data

#### Load models
```{r}
load("processdata/recruits_house_2021_04_19.rda")
str(recruits_house)
names(recruits_house)
```


#### Prepare data set for second part of the hurdle
```{r}
recruits_house <- recruits_house[recruits_house$recruitment > 0, ]
str(recruits_house)
summary(recruits_house$recruitment)

is.integer(recruits_house$recruitment)
recruits_house$recruitment <- as.integer(recruits_house$recruitment)
is.integer(recruits_house$recruitment)
```


#### Scale variables if necessary
```{r}
# recruits_house$disthus <- scale(recruits_house$disthus)

# recruits_house$distvei2 <- scale(recruits_house$distvei2)

# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
# recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)

# recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
# recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

# summary(recruits_house$disthus_600)
```


#### Zero-truncated house models
```{r}
library(lme4)
rec_h_beta1 <- glmer(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)
summary(rec_h_beta1)

rec_h_beta1_1 <- glmer(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)



## test for interactions ----
rec_h_beta2 <- glmer(recruitment ~ disthus_600 + log(distvei2+1) + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

rec_h_beta3 <- glmer(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + beitetrykk9*treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

rec_h_beta4 <- glmer(recruitment ~ disthus_600 + log(distvei2+1)*skogkategori + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house) 

rec_h_beta5 <- glmer(recruitment ~ disthus_600*skogkategori + log(distvei2+1) + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

rec_h_beta6 <- glmer(recruitment ~ disthus_600*skogkategori + log(distvei2+1) + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)

rec_h_beta7 <- glmer(recruitment ~ disthus_600 + log(distvei2+1)*skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)


rec_h_beta8 <- glmer(recruitment ~ disthus_600 + log(distvei2+1)*skogkategori + treartgruppe9 + beitetrykk9*skogkategori + kant + helling + HOH + I(HOH^2) + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house) # best

rec_h_beta9 <- glmer(recruitment ~ disthus_600 + log(distvei2+1)*skogkategori + beitetrykk9*treartgruppe9 + kant + helling + HOH + I(HOH^2) + moose_density + tretetthet9 + behandling1011 + ac + (1|KOMNR),
family = poisson(),
data = recruits_house)



library(glmmTMB)
rec_h_beta4_TMB <- glmmTMB(recruitment ~ disthus_600 + log(distvei2+1) + skogkategori + treartgruppe9 + beitetrykk9 + kant + helling + HOH + I(HOH^2) + tretetthet9 + ac + (1|KOMNR),
                           family = poisson,
                           data = recruits_house) # Same as rec_h_beta4, just run with glmmTMB

summary(rec_h_beta4_TMB_500)
save(rec_h_beta4_TMB_500, file = "output/recruitment_2021_04_12/rec_h_beta4_TMB_500.rda")




```


#### AIC selection beta models
```{r}
library(AICcmodavg)

candmods_rec_house_beta <- list(rec_h_beta1, rec_h_beta2, rec_h_beta3, rec_h_beta4, rec_h_beta5, rec_h_beta6, rec_h_beta7, rec_h_beta8, rec_h_beta9)
candmods_rec_house_beta <- list(rec_h_beta4, rec_h_beta8)
aictab(candmods_rec_house_beta, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)
#       K     AIC Delta_AIC AICWt Cum.Wt        LL
# Mod8 28 1010872      0.00  0.73   0.73 -505408.0
# Mod4 29 1010874      2.00  0.27   1.00 -505408.0
# Mod5 29 1012012   1139.94  0.00   1.00 -505977.0
# Mod2 25 1012098   1226.03  0.00   1.00 -506024.0
# Mod9 27 1015134   4262.26  0.00   1.00 -507540.1
# Mod7 25 1015232   4359.91  0.00   1.00 -507591.0
# Mod6 25 1016053   5180.52  0.00   1.00 -508001.3
# Mod3 23 1016103   5230.58  0.00   1.00 -508028.3
# Mod1 21 1016191   5318.84  0.00   1.00 -508074.4



save(rec_h_beta1, file = "output/2021_04_28/rec_h_beta1.rda")
save(rec_h_beta1_1, file = "output/2021_04_28/rec_h_beta1_1.rda")
save(rec_h_beta2, file = "output/2021_04_28/rec_h_beta2.rda")
save(rec_h_beta3, file = "output/2021_04_28/rec_h_beta3.rda")
save(rec_h_beta4, file = "output/2021_04_28/rec_h_beta4.rda")
save(rec_h_beta5, file = "output/2021_04_28/rec_h_beta5.rda")
save(rec_h_beta6, file = "output/2021_04_28/rec_h_beta6.rda")
save(rec_h_beta7, file = "output/2021_04_28/rec_h_beta7.rda")
save(rec_h_beta8, file = "output/2021_04_28/rec_h_beta8.rda")
save(rec_h_beta9, file = "output/2021_04_28/rec_h_beta9.rda")

```


#### Summary beta house
```{r}
summary(rec_h_beta8)

```



#### Plots beta house
```{r}
library(sjPlot)
plot_model(rec_h_beta8, type = "diag")
plot_model(rec_h_beta8, type = "eff")
plot_model(rec_h_beta8, type = "est")

tab_model(rec_h_beta8, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))


plot_model(rec_h_beta8, type = "eff", terms = "disthus_600")
plot_model(rec_h_beta8, type = "eff", terms = c("distvei2 [0:600]", "skogkategori"))
plot_model(rec_h_beta8, type = "eff", terms = c("beitetrykk9", "skogkategori"))

```


#### Save house beta model
```{r}
save(rec_h_beta8, file = "output/2021_04_28/rec_h_beta8.rda")
```







## Delta AIC comparison figure
```{r}
library(readr)
delta_aic_comparison <- read_csv("figs_tabs/delta_aic_comparison.csv", na = "NA")
str(delta_aic_comparison)

delta_aic_comparison <- delta_aic_comparison[-c(9), ]
str(delta_aic_comparison)

library(ggplot2)
overview <- ggplot() + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  labs(x="Threshold cut-off\n", y = "delta AIC\n")+
   theme_classic() +
  scale_color_manual(name = "Model", 
                     values = c("Browsing pressure\nbinary\n" = "#56B4E9", "Browsing pressure\nzero-truncated\n" = "#0072B2", "recruitment\nbinary\n" = "#b37c00", "recruitment\nzero-truncated" = "#E69F00")) +
  scale_x_continuous(breaks = c(0, 100, 200, 300, 400, 500, 600, 3000)) +
  scale_y_continuous(breaks = c(0, 500, 1000, 2000, 3000, 4000, 5000, 7500))


overview2 <- ggplot() + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) + 
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) +
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) + 
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) +
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  
  labs(x="Threshold cut-off\n", y = "delta AIC\n")+
   theme_classic() +
  scale_color_manual(name = "Model", 
                     values = c("Browsing pressure\nbinary\n" = "#56B4E9", "Browsing pressure\nzero-truncated\n" = "#0072B2", "recruitment\nbinary\n" = "#b37c00", "recruitment\nzero-truncated" = "#E69F00")) +
  scale_x_continuous(breaks = c(0, 100, 200, 300, 400, 500, 600, 3000)) +
  scale_y_continuous(breaks = c(0, 500, 1000, 2000, 3000, 4000, 5000, 7500))




detail <- ggplot() + 
  geom_vline(xintercept = 500, linetype="solid", color = "grey70", size=1) + 
  geom_hline(yintercept = 20, linetype="solid", color = "grey70", size=1) + 
  geom_hline(yintercept = 10, linetype="dashed", color = "grey80", size=1) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  labs(x="Threshold cut-off\n", y = "delta AIC\n")+
   theme_classic() +
  scale_color_manual(name = "Model", 
                     values = c("Browsing pressure\nbinary\n" = "#56B4E9", "Browsing pressure\nzero-truncated\n" = "#0072B2", "recruitment\nbinary\n" = "#b37c00", "recruitment\nzero-truncated" = "#E69F00")) +
  scale_x_continuous(breaks = c(0, 100, 200, 300, 400, 500, 600, 3000)) +
  scale_y_continuous(breaks = c(0, 5, 10, 20, 30, 40, 50))+ 
  coord_cartesian(xlim = c(0, 600), ylim = c(0, 50))


detail2 <- ggplot() + 
  geom_vline(xintercept = 500, linetype="solid", color = "grey70", size=1) + 
  geom_hline(yintercept = 20, linetype="solid", color = "grey70", size=1) + 
  geom_hline(yintercept = 10, linetype="dashed", color = "grey80", size=1) + 
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) +
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_bin, color = "Browsing pressure\nbinary\n"), show.legend = T) + 
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) + 
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_bp_zt, color = "Browsing pressure\nzero-truncated\n"), show.legend = T) +
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_bin, color = "recruitment\nbinary\n"), show.legend = T) + 
  
  geom_point(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  geom_line(data = delta_aic_comparison, aes(x = threshold, y = delta_aic_rec_zt, color = "recruitment\nzero-truncated"), show.legend = T) +
  
  labs(x="Threshold cut-off\n", y = "delta AIC\n")+
   theme_classic() +
  scale_color_manual(name = "Model", 
                     values = c("Browsing pressure\nbinary\n" = "#56B4E9", "Browsing pressure\nzero-truncated\n" = "#0072B2", "recruitment\nbinary\n" = "#b37c00", "recruitment\nzero-truncated" = "#E69F00")) +
  scale_x_continuous(breaks = c(0, 100, 200, 300, 400, 500, 600, 3000)) +
  scale_y_continuous(breaks = c(0, 5, 10, 20, 30, 40, 50))+ 
  coord_cartesian(xlim = c(100, 600), ylim = c(0, 50))



```
