==> 09_recruitment_treedensity.Rmd <==
---
title: "09_recruitment_treedensity"
author: "Anne"
date: "26/03/2021"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



## Load data and make some adjustments
```{r}
load("processdata/recruits_house_2021_02_02.rda")
str(recruits_house)

# scale
# recruits_house$beitetrykk9 <- scale(recruits_house$beitetrykk9)
recruits_house$helling <- scale(recruits_house$helling)
recruits_house$HOH <- scale(recruits_house$HOH)
# recruits_house$tretetthet9 <- scale(recruits_house$tretetthet9)
# recruits_house$moose_density <- scale(recruits_house$moose_density)
recruits_house$disthus_600 <- scale(recruits_house$disthus_600)
# recruits_house$disthus_500 <- scale(recruits_house$disthus_500)

summary(recruits_house$tretetthet9)
hist(recruits_house$tretetthet9)
recruits_house$tretetthet9_trunc <- ifelse(recruits_house$tretetthet9 >= 5000, 5000, recruits_house$tretetthet9)

recruits_house$tretetthet9_trunc <- as.integer(recruits_house$tretetthet9_trunc)

names(recruits_house)



library(plyr)

levels(recruits_house$skogkategori)
recruits_house$skogkategori <- mapvalues(recruits_house$skogkategori, from =  c("Lavproduktiv_eldreskog", "Høyproduktiv_eldreskog", "Høyproduktiv_ungskog", "Lavproduktiv_ungskog", "Uproduktiv_skog"), to = c("Lowproductive_old_forest",  "Highproductive_old_forest", "Highproductive_young_forest", "Lowproductive_young_forest", "Unproductive_forest"))
levels(recruits_house$skogkategori)

levels(recruits_house$kant)
recruits_house$kant <- mapvalues(recruits_house$kant, from =  c("kant >20m", "kant <10m", "kant 10-20m", "mangler kant"), to = c("Edge >20m", "Edge <10m", "Edge 10-20m", "Edge information missing"))
levels(recruits_house$kant)

levels(recruits_house$treartgruppe9)
recruits_house$treartgruppe9 <- mapvalues(recruits_house$treartgruppe9, from =  c("FURU", "LAUV", "ROS"), to = c("Pine", "Deciduous trees", "RAW"))
levels(recruits_house$treartgruppe9)

save(recruits_house, file = "processdata/recruits_house_treedensity_2021_04_06.rda")
```


## Models
```{r}
library(lme4)
library(lmerTest)

lmer1 <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + HOH + treartgruppe9 + beitetrykk9 + moose_density + ac + (1|KOMNR),
            data = recruits_house, REML = F)

lmer2 <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + treartgruppe9 + beitetrykk9 + moose_density + ac + (1|KOMNR),
            data = recruits_house, REML = F) # 2nd best, with one point difference

lmer3 <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + HOH + treartgruppe9 + beitetrykk9 + ac + (1|KOMNR),
            data = recruits_house, REML = F)

lmer4 <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + treartgruppe9 + beitetrykk9 + ac + (1|KOMNR),
            data = recruits_house, REML = F) # BEST


summary(lmer1)

lmer4_reml <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + treartgruppe9 + beitetrykk9 + ac + (1|KOMNR),
            data = recruits_house, REML = T)
summary(lmer4_reml)

save(lmer4_reml, file = "lmer4_reml.rda") # in output/treedensity
```


## AIC selection
```{r}
library(AICcmodavg)

candmods_treedensity <- list(lmer1, lmer2, lmer3, lmer4)
aictab(candmods_treedensity, modnames = NULL,
                       second.ord = F, nobs = NULL, sort = TRUE)

#       K      AIC Delta_AIC AICWt Cum.Wt        LL
# Mod4 16 64835.47      0.00  0.37   0.37 -32401.74
# Mod2 17 64836.00      0.52  0.28   0.65 -32401.00
# Mod3 17 64836.74      1.26  0.20   0.85 -32401.37
# Mod1 18 64837.29      1.82  0.15   1.00 -32400.65

```


## Test assumptions
```{r}
library(sjPlot)

plot_model(lmer4_reml, type = "diag")


tab_model(lmer4_reml, transform = NULL,
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, p.threshold = c(0.05, 0.01, 0.001))

```

## Effect plots
```{r}
library(ggplot2)
library(sjPlot)
library(sjlabelled)
# lmer4_reml <- lmer(log(tretetthet9_trunc+1) ~ disthus_600 + skogkategori + kant + helling + treartgruppe9 + beitetrykk9 + ac + (1|KOMNR),
#             data = recruits_house, REML = T)

load("output/treedensity/lmer4_reml.rda")

p_treedens_disthus <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, terms = c("disthus_600"))
p_treedens_disthus <- p_treedens_disthus + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to houses\n") +
  ylab("log Tree density\n")


p_treedens_skogkategori <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, 
                                      terms = c("skogkategori"))
p_treedens_skogkategori <- p_treedens_skogkategori + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
     axis.text.x = element_text(angle = 45, vjust = 1, hjust = 0.8)) +
  xlab("Forest category\n") +
  ylab("log Tree density\n")

p_treedens_kant <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, terms = c("kant"))
p_treedens_kant <- p_treedens_kant + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 12, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
     axis.text.x = element_text(angle = 45, vjust = 1, hjust = 0.8)) +
  xlab("Edge\n") +
  ylab("log Tree density\n")

p_treedens_helling <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, terms = c("helling"))
p_treedens_helling <- p_treedens_helling + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("log Tree density\n")

p_treedens_treart <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p_treedens_treart <- p_treedens_treart + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30"),
     axis.text.x = element_text(angle = 45, vjust = 1, hjust = 0.8)) +
  xlab("Tree species group\n") +
  ylab("log Tree density\n")

p_treedens_beite <- plot_model(lmer4_reml, type = c("eff"), transform = NULL, terms = c("beitetrykk9"))
p_treedens_beite <- p_treedens_beite + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Browsing pressure\n") +
  ylab("log Tree density\n")


library(gridExtra)
treedensity_grid <- grid.arrange(p_treedens_beite, p_treedens_disthus, p_treedens_helling, p_treedens_kant, p_treedens_skogkategori, p_treedens_treart,
          nrow = 3, ncol = 2)


ggsave(file = "treedensity_grid.tiff", treedensity_grid, width = 30, height = 30, units = "cm", dpi = 150)

```
