==> 09_treedensity-height_hurdle.Rmd <==
---
title: "treedensity-height_hurdle"
author: "anne"
date: "3 mai 2019"
output: word_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```



#### Modelling spatial autocorrelation (https://onlinelibrary.wiley.com/doi/full/10.1111/j.2007.0906-7590.05171.x)
Change the projection and make coordinates and a spatial point data frame to make a shapefile

```{r}
library(tidyr)
browse9.5m$UTMmerge <- paste(browse9.5m$UTM_OV_33, browse9.5m$UTM_SN_33, sep="_")
sum(duplicated(browse9.5m$UTMmerge))
```


```{r}

library(sp)
library(rgdal)

crs <- CRS("+proj=utm +zone=33 +ellps=GRS80 +units=m +no_defs")
coords <- cbind(browse9.5m$UTM_OV_33, browse9.5m$UTM_SN_33)
data_spdf <- SpatialPointsDataFrame(coords, browse9.5m, proj4string = crs)
# plot(data_spdf)
writeOGR(data_spdf, dsn = "output", layer = "spac", driver = "ESRI Shapefile")

```


```{r}

library(spdep)
library(ape)

# spdep package:

# prepare neighbour lists for spatial autocorrelation analysis
# nb.list <- dnearneigh(as.matrix(browse9.3[browse9.3$browspres_merged, c("UTM_OV_33", "UTM_SN_33")]), 0, 5000)
# nb.weights <- nb2listw(nb.list, zero.policy = TRUE)
# 
# moran.plot(resid_m1abin, nb.weights, zero.policy = TRUE, spChk = NULL, labels = TRUE, xlab = NULL, ylab = NULL, quiet = NULL)
# 
# resid_m1abin <- resid(m1abin, type = "pearson")
# resid_m1bbin <- resid(m1bbin, type = "pearson")
# 
# moran.test(browse9.3$browspres_merged, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)
# 
# moran.test(resid_m1bbin, nb.weights, zero.policy = TRUE, randomisation=TRUE,
#  alternative="greater", rank = FALSE, na.action=na.fail, spChk=NULL,
#  adjust.n=TRUE, drop.EI2=FALSE)


# Make a matrix of coordinates - use when chunk above was not run
coords <- as.matrix(cbind(browse9.5m$UTM_OV_33, browse9.5m$UTM_SN_33))

# compute the autocovariate based on the above distance and weight
ac <- autocov_dist(browse9.5m$browspres_merged, coords, nbs = 5000, type="inverse", zero.policy = TRUE)
browse9.5m$ac <- ac


save(browse9.5m, file = "browse9.5m.rda")


```




#### gam

```{r}
library(ggplot2)
ggplot(browse9.6, aes(x = disthus, y = browspres_merged)) +
    geom_point() +
    stat_smooth(method = "glm", color = "green") +
  stat_smooth(color = "blue")

# +
#   scale_x_continuous(breaks=c(100,150, 200, 250, 300, 400, 500, 1000, 1500, 2000, 3000))

```

```{r}
library(mgcv)



gam1 <- gam(proplt ~ disthus + distvei2 + skogkategori + treartgruppe9 + kant + helling + HOH + browspres_merged,
             data = browse9.6,
             method = "REML")

gam2 <- gam(proplt ~ s(disthus) + s(distvei2) + skogkategori + treartgruppe9 + kant + helling + HOH + browspres_merged,
             data = browse9.5m,
             method = "REML")

```

```{r}
plot.gam(gam1, all.terms = TRUE)
plot.gam(gam2, all.terms = TRUE)
# threshold: disthus ~700/900m, distvei2 ~200/300m
```










###################################################################################
##     HOUSE                HOUSE                 HOUSE                       #####
###################################################################################

##       House Bin - First part of a hurdle model to account for zero inflation  ##

##### Checking for the cut for disthus (browse9.5m/6) 
```{r}
# load("processdata/browse9.5sc.rda")
load("processdata/browse9.5m.rda")
load("processdata/browse9.5msc.rda")
```

           "distvei2",
```{r}
library(lme4)

pvars <- c("disthus_400", "HOH", "ac", "disthus_500", "disthus_300", "disthus") # "disthus","disthus_500", "disthus_1000", "disthus_900", "disthus_800", "disthus_700", "disthus_600", "disthus_160", "disthus_300", "disthus_200", "disthus_175", "disthus_350", "helling", "browspres_merged"
browse9.5msc <- browse9.5m
browse9.5msc[pvars] <- lapply(browse9.5msc[pvars],scale)

save(browse9.5msc, file = "browse9.5msc.rda")
load("processdata/browse9.5msc.rda")

m1bin_house <- glmer(factor(proplt > 0) ~ disthus + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m2bin_house <- glmer(factor(proplt > 0) ~ disthus_1000 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m3bin_house <- glmer(factor(proplt > 0) ~ disthus_900 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m4bin_house <- glmer(factor(proplt > 0) ~ disthus_800 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m5bin_house <- glmer(factor(proplt > 0) ~ disthus_700 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m6bin_house <- glmer(factor(proplt > 0) ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m7bin_house <- glmer(factor(proplt > 0) ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m8bin_house <- glmer(factor(proplt > 0) ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc) # best!

m9bin_house <- glmer(factor(proplt > 0) ~ disthus_160 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m10bin_house <- glmer(factor(proplt > 0) ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc) 

m11bin_house <- glmer(factor(proplt > 0) ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m12bin_house <- glmer(factor(proplt > 0) ~ disthus_175 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m13bin_house <- glmer(factor(proplt > 0) ~ disthus_350 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

# ------------ log selection -----------

pvars <- c("disthus_400",
           "helling","HOH", "browspres_merged", "ac")
browse9.5msc <- browse9.5m
browse9.5msc[pvars] <- lapply(browse9.5msc[pvars],scale)

# "disthus_400", "disthus", "distvei2",

m0abin_house <- glmer(factor(proplt > 0) ~ log(disthus+1) + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac +  (1|KOMNR),
            family = binomial(),
            data = browse9.5msc)

m0bbin_house <- glmer(factor(proplt > 0) ~ log(disthus+1) + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac +  (1|KOMNR),
            family = binomial(),
            data = browse9.5msc)

m0cbin_house <- glmer(factor(proplt > 0) ~ disthus + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac +  (1|KOMNR),
            family = binomial(),
            data = browse9.5msc)

m0dbin_house <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc) # BEST


```


##### House bin model selection
```{r}
library(lme4)
load("processdata/browse9.5msc.rda")

m8bin_house <- glmer(factor(proplt > 0) ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

mhusbin <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc) # best

mhusabin <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

m8bbin_house <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.5msc)

AIC(m8bin_house, mhusbin, m8bbin_house)

save(mhusbin, file = "mhusbin.rda")
save(mhusabin, file = "mhusabin.rda")

```

###### AIC selection threshold distance to house, log or not and house model selection
```{r}

AIC(m7bin_house, m8bin_house, m10bin_house) # cut 400 is the thing!

AIC(m8bin_house, mhusbin, m8bbin_house)



library(AICcmodavg)


candmods_R_hbin <- list(m8bin_house, mhusbin, m8bbin_house)

glmmbin_tab <- aictab(candmods_R_hbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

```



##       House Beta - Second part of a hurdle model  ##

##### Checking for the cut for disthus (browse9.5mpos)  
```{r}
library(glmmTMB)


browse9.5mpos <- browse9.5m[browse9.5m$proplt > 0, ]
browse9.5mpos$proplt <- ifelse(browse9.5mpos$proplt == 1, 0.99, browse9.5mpos$proplt)

pvars <- c("disthus","disthus_500", "disthus_400", "disthus_1000", "disthus_900", "disthus_800", "disthus_700", "disthus_600", "disthus_160", "disthus_300", "disthus_200", "disthus_175", "disthus_350",
           "distvei2",
           "HOH", "ac") #"helling",, "browspres_merged"
browse9.5mpossc <- browse9.5mpos
browse9.5mpossc[pvars] <- lapply(browse9.5mpossc[pvars],scale)

save(browse9.5mpossc, file = "browse9.5mpossc.rda")
load("processdata/browse9.5mpossc.rda")


m1beta_house <- glmmTMB(proplt ~ disthus_1000 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m2beta_house <- glmmTMB(proplt ~ disthus_900 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m3beta_house <- glmmTMB(proplt ~ disthus_800 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m4beta_house <- glmmTMB(proplt ~ disthus_700 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m5beta_house <- glmmTMB(proplt ~ disthus_600 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m6beta_house <- glmmTMB(proplt ~ disthus_500 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m7beta_house <- glmmTMB(proplt ~ disthus_400 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m8beta_house<- glmmTMB(proplt ~ disthus_300 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m10beta_house <- glmmTMB(proplt ~ disthus_200 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) # best

m11beta_house <- glmmTMB(proplt ~ disthus_160 + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m12beta_house <- glmmTMB(proplt ~ disthus + distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())



# ------ log selection ------
pvars <- c("disthus_200",
           "HOH", "ac") 
browse9.5mpossc <- browse9.5mpos
browse9.5mpossc[pvars] <- lapply(browse9.5mpossc[pvars],scale)

# "helling",, "browspres_merged"

m0abeta_house <- glmmTMB(proplt ~ distvei2 + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) 

m0bbeta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) # Best!

m0cbeta_house <- glmmTMB(proplt ~ disthus_200 + log(distvei2+1) + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) 

m0dbeta_house <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) 

save(m0bbeta_house, file = "m0bbeta_house.rda")

```


```{r}
AIC(m6beta, m7beta, m8beta, m10beta, m11beta, m12beta)
# all are the same (also not significant) -> I suppose that means that disthus is not important?
AIC(m10beta, m0bbeta)
# m0bbeta best!

# library(AICcmodavg)
library(bblme)

# candmods_R_hbin <- list(m1bin, m7bin, m8bin, m10bin)
# 
# glmmbin_tab <- aictab(candmods_R_hbin, modnames = NULL,
#                        second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

ICtab(m1beta_house, m7beta_house, m8beta_house, m10beta_house, m11beta_house, type = "AIC", weights = TRUE, delta = TRUE)
AIC(m1beta_house, m7beta_house, m8beta_house, m10beta_house, m11beta_house)

ICtab(m0abeta_house, m0bbeta_house, m0cbeta_house, m0dbeta_house, type = "AIC", weights = TRUE, delta = TRUE)
AIC(m0abeta_house, m0bbeta_house, m0cbeta_house, m0dbeta_house)

```




##### House beta Model selection (browse9.5mpossc)
```{r}
library(glmmTMB)
load("processdata/browse9.5m.rda")
load("processdata/browse9.5mpos.rda")
load("processdata/browse9.5mpossc.rda")

browse9.5mpos <- browse9.5m[browse9.5m$proplt > 0, ]
browse9.5mpos$proplt <- ifelse(browse9.5mpos$proplt == 1, 0.99, browse9.5mpos$proplt)


pvars <- c("HOH", "ac") # "helling", "browspres_merged"
browse9.5mpossc <- browse9.5mpos
browse9.5mpossc[pvars] <- lapply(browse9.5mpossc[pvars],scale)

save(browse9.5mpossc, file = "browse9.5mpossc.rda")


m12beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m13beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) # best

m14beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m15beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) # 2nd best

m16beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) # 3rd best, but simplest dAIC 0.8 in comp. to m13 (0.0)

m16abeta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())  # 3rd best 

m17beta_house <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + kant + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family())

m18beta_house <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.5mpossc,
            family = beta_family()) 

```

###### AIC selection beta house
```{r}
save(m16abeta_house, file = "m16abeta_house.rda")

AIC(m12beta_house, m13beta_house, m14beta_house, m15beta_house, m16beta_house, m16abeta_house, m17beta_house, m18beta_house)

# library(AICcmodavg)
library(bbmle)

# candmods_R_rbin <- list(m14bin, m15bin, m16cbin, m17bin, m18bin)
# 
# glmmbeta_tab <- aictab(candmods_R_rbin, modnames = NULL,
#                        second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

ICtab(m12beta_house, m13beta_house, m14beta_house, m15beta_house, m16beta_house, m16abeta_house, m17beta_house, m18beta_house, type = "AIC", weights = TRUE, delta = TRUE)
AIC(m12beta, m13beta, m14beta, m15beta, m16abeta)


ICtab(m12beta_house, m13beta_house, m15beta_house, m16beta_house, type = "AIC", weights = TRUE, delta = TRUE)
AIC(m12beta_house, m13beta_house, m15beta_house, m16beta_house)

library(effects)
plot(allEffects(m16beta))
plot(effect("skogkategori", m16beta))

save(m16beta, file = "m16beta.rda")

```
              dAIC df weight
m13beta_house   0.0 14 0.262 
m15beta_house   0.2 13 0.235 
m16beta_house   0.8 12 0.175 
m16abeta_house  0.8 12 0.175 
m12beta_house   3.5 17 0.045 
m14beta_house   3.7 16 0.042 
m18beta_house   3.9 11 0.037 
m17beta_house   4.3 15 0.031 

```{r}
browse9.5mpossc <- browse9.5mpossc[,c(1,3,7,15:17,19:20,25:26,40)]
save(browse9.5mpossc, file = "browse9.5mpossc.rda")
```

```{r}
library(sjPlot)
load("output/m16abeta_house.rda")

tab_model(m16abeta_house, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))


```




###################################################################################
##     ROADS                ROADS                 ROADS                       #####
###################################################################################

##       Roads Bin - First part of a hurdle model to account for zero inflation  ##

## Roads bin model selection
```{r}
library(lme4)

load("processdata/browse9.6.rda")
load("processdata/browse9.6sc.rda")

pvars <- c("HOH","ac" ) #  "helling", "browspres_merged", "ac" 
browse9.6sc <- browse9.6
browse9.6sc[pvars] <- lapply(browse9.6sc[pvars],scale)

save(browse9.6sc, file = "browse9.6sc.rda")


m1bin_roads <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.6sc) 

m2bin_roads <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.6sc) # best

m3bin_roads <- glmer(factor(proplt > 0) ~ skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
family = binomial(),
data = browse9.6sc)


```


##### AIC roads bin model selection
```{r}
save(m2bin_roads, file = "m2bin_roads.rda")

AIC(m1bin_roads, m2bin_roads, m3bin_roads)


library(AICcmodavg)
# library(bblme)

candmods_R_rbin <- list(m1bin_roads, m2bin_roads, m3bin_roads)

glmmbeta_tab <- aictab(candmods_R_rbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE, c.hat = 1)

# ICtab(m4abeta, m2beta, m3beta, m4beta, type = "AIC", weights = TRUE, delta = TRUE)
# AIC(m4abeta, m2beta, m3beta, m4beta)



library(sjPlot)

tab_model(m2bin_roads, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))


```
    K     AICc Delta_AICc AICcWt Cum.Wt       LL
Mod2 13 10902.98        0.0   0.55   0.55 -5438.47
Mod1 16 10903.38        0.4   0.45   1.00 -5435.66
Mod3 12 10917.68       14.7   0.00   1.00 -5446.82

```{r}
browse9.6sc <- browse9.6sc[,c(1,3,7,15:17,19:20,25:26,33,40)]
browse9.6sc <- browse9.6sc[,c(1:10,12)]
save(browse9.6sc, file = "browse9.6sc.rda")
```


##       Roads Beta - Second part of a hurdle model  ##

## Roads Beta model selection
```{r}
library(glmmTMB)

# browse9.6sc$proplt <- ifelse(browse9.6sc$proplt == 1, 0.99, browse9.6sc$proplt)

load("processdata/browse9.6sc.rda")
browse9.6scpos <- browse9.6sc[browse9.6sc$proplt > 0, ]
save(browse9.6scpos, file = "browse9.6scpos.rda")

m1betaroad <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.6scpos,
            family = beta_family())

m2betaroad <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + kant + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.6scpos,
            family = beta_family())

m3betaroad <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.6scpos,
            family = beta_family()) # Best

m4betaroad <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.6scpos,
            family = beta_family())

m5betaroad <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
            data = browse9.6scpos,
            family = beta_family())


```

##### AIC selection roads beta 
```{r}
save(m3betaroad, file = "m3betaroad.rda")

AIC(m1betaroad, m2betaroad, m3betaroad, m4betaroad, m5betaroad)

library(bbmle)
ICtab(m1betaroad, m2betaroad, m3betaroad, m4betaroad, m5betaroad, type = "AIC", weights = TRUE, delta = TRUE, base = TRUE)
```

           df       AIC
m1betaroad 17 -599.5613
m2betaroad 16 -601.0889
m3betaroad 13 -603.6964
m4betaroad 14 -602.1769
m5betaroad 12 -603.3233

           dAIC df weight
m3betaroad  0.0 13 0.371 
m5betaroad  0.4 12 0.308 
m4betaroad  1.5 14 0.174 
m2betaroad  2.6 16 0.101 
m1betaroad  4.1 17 0.047 




###############################################################
## Tables and figures manuscript                        ######
##############################################################

## Tables appendix

```{r}
# load("output/m16abeta.rda")
```


Table of estimates:
```{r}
library(sjPlot)

tab_model(m16cbin, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))


tab_model(m3betaroad, transform = NULL, 
          show.intercept = TRUE, show.est = TRUE, show.ci = 0.95, show.se = TRUE, show.std = NULL, show.p = TRUE, show.stat = FALSE, show.df = FALSE, show.zeroinf = FALSE, show.r2 = TRUE, show.icc = TRUE, show.adj.icc = FALSE, show.re.var = TRUE, show.aic = TRUE, show.dev = FALSE, show.obs = TRUE, p.threshold = c(0.05, 0.01, 0.001))


```

## figures
```{r}
library(sjPlot)
library(ggplot2)

load("processdata/browse9.6scpos.rda")


p1hus <- plot_model(model, type = c("eff"), terms = c("disthus_400"))
p1hus <- p1hus + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to house\n") +
  ylab("% Recruits\n")

p1 <- plot_model(model, type = c("eff"), terms = c("distvei2 [0:500]"))
p1 <- p1 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Distance to roads\n") +
  ylab("% Recruits\n")

p2 <- plot_model(m3betaroad, type = c("eff"), transform = NULL, terms = c("skogkategori"))
p2 <- p2 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Forest category\n") +
  ylab("% Recruits\n") +
  label_angle(angle.x = 45, base.theme = theme_sjplot())

p3 <- plot_model(m3betaroad, type = c("eff"), transform = NULL, terms = c("treartgruppe9"))
p3 <- p3 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Tree species group\n") +
  ylab("% Recruits\n")

# p4 <- plot_model(m2bin_roads, type = c("eff"), transform = NULL, terms = c("kant"))
# pb3 + theme_sjplot()

p5 <- plot_model(m3betaroad, type = c("eff"), transform = NULL, terms = c("helling [n=1]"))
p5 <- p5 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Slope\n") +
  ylab("% Recruits\n")

p6 <- plot_model(m3betaroad, type = c("eff"), transform = NULL, terms = c("browspres_merged"))
p6 <- p6 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Browsing pressure\n") +
  ylab("% Recruits\n")

p7 <- plot_model(m3betaroad, type = c("eff"), transform = NULL, terms = c("HOH"))
p7 <- p7 + theme_light() +
  theme(plot.title = element_blank(),
    axis.text = element_text(family = "sans", size = 20, color = "grey30"),
        axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
  xlab("Altitude\n") +
  ylab("% Recruits\n")

library(gridExtra)
grid.arrange(p6, p2, p3, p5, p7,
          nrow = 3, ncol = 2)
# bin: p1hus, p1, p6, p2, p3, p5, p7
```





########################################################
## Additional post-hocs                            ###
######################################################


##### Testing for kant in total/ indirect effect
```{r}
# BIN
library(lme4)
load("processdata/browse9.6sc.rda")

# m2bin_roads <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + (1|KOMNR),
# family = binomial(),
# data = browse9.6sc)

bin_kant <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + kant + helling + HOH + I(HOH^2) +  ac + (1|KOMNR), 
                  family = binomial(), 
                  data = browse9.6sc)

summary(bin_kant)

# BETA
library(glmmTMB)
load("processdata/browse9.6scpos.rda")

# m3betaroad <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + (1|KOMNR),
#             data = browse9.6scpos,
#             family = beta_family())

beta_kant <- glmmTMB(proplt ~  skogkategori + treartgruppe9 + HOH + I(HOH^2) + kant + (1|KOMNR), data = browse9.6scpos, family = beta_family())

summary(beta_kant)

```



##### Testing for a regional effect/ interaction region:distvei
```{r}
##### House #############################################
## bin
library(lme4)
load("processdata/browse9.5m_regionsc.rda")
load("output/mhusabin_reg2.rda")
load("output/mhusabin.rda")

mhusabin_reg <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + region + (1|KOMNR),
family = binomial(),
data = browse9.5m_regionsc)

mhusabin_reg2 <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + browspres_merged + region + disthus_400:region + (1|KOMNR),
family = binomial(),
data = browse9.5m_regionsc)


save(mhusabin_reg, file = "mhusabin_reg.rda")
save(mhusabin_reg2, file = "mhusabin_reg2.rda")

summary(mhusabin_reg)
summary(mhusabin_reg2)

# AIC ----
AIC(mhusabin, mhusabin_reg)
#             df      AIC
# mhusabin     17 14213.03
# mhusabin_reg 30 14182.70

library(AICcmodavg)
candmods <- list(mhusabin, mhusabin_reg, mhusabin_reg2)

glmmbeta_tab <- aictab(candmods, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt       LL
# Mod2 30 14182.70      0.00     1      1 -7061.35
# Mod1 17 14213.03     30.32     0      1 -7089.51

# Plots
library(sjPlot)
tab_model(mhusabin_reg2)

plot_model(mhusabin_reg, type = c("eff"), transform = NULL, terms = c("region"))

plot_model(mhusabin_reg2, type = c("eff"), transform = NULL, terms = c("disthus_400", "region [Sor-Trondelag, Sogn og Fjordane]"))


## beta
library(glmmTMB)
load("processdata/browse9.5m_regionsc.rda")
load("processdata/browse9.5m_regionscpos.rda")

load("output/m16beta_house_reg.rda")
load("output/m16abeta_house.rda")

m16beta_house_reg <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + region + (1|KOMNR),
            data = browse9.5m_regionscpos,
            family = beta_family()) 

save(m16beta_house_reg, file = "m16beta_house_reg.rda")

summary(m16beta_house_reg)


# AIC ----
AIC(m16abeta_house, m16beta_house_reg)
#                   df       AIC
# m16beta_house     12 -809.5930
# m16beta_house_reg 25 -872.8844
library(bbmle)
ICtab(m16abeta_house, m16beta_house_reg, m16beta_house_reg2, type = "AIC", weights = TRUE, delta = TRUE)
#                   dAIC df weight
# m16beta_house_reg  0.0 25 1     
# m16abeta_house    63.3 12 <0.001

# Plots
library(sjPlot)
tab_model(m16beta_house_reg)

plot_model(m16beta_house_reg, type = c("eff"), transform = NULL, terms = c("region"))




##### Roads ###########################################
## bin
library(lme4)
load("processdata/browse9.6_regionsc.rda")
load("output/m2bin_roads.rda")
load("output/m2bin_roads_reg.rda")

m2bin_roads_reg <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) + ac + region + (1|KOMNR),
family = binomial(),
data = browse9.6_regionsc)

m2bin_roads_reg2 <- glmer(factor(proplt > 0) ~ log(distvei2+1) + browspres_merged + region + log(distvei2+1):region + (1|KOMNR),
family = binomial(),
data = browse9.6_regionsc)

save(m2bin_roads_reg, file = "m2bin_roads_reg.rda")
save(m2bin_roads_reg2, file = "m2bin_roads_reg2.rda")

summary(m2bin_roads_reg)
summary(m2bin_roads_reg2)

# AIC ----
AIC(m2bin_roads, m2bin_roads_reg)
#                 df      AIC
# m2bin_roads     13 10902.94
# m2bin_roads_reg 26 10885.19

library(AICcmodavg)
candmods <- list(m2bin_roads, m2bin_roads_reg)

glmmbeta_tab <- aictab(candmods, modnames = NULL,
                       second.ord = FALSE, nobs = NULL, sort = TRUE)
#       K      AIC Delta_AIC AICWt Cum.Wt       LL
# Mod2 26 10885.19      0.00     1      1 -5416.59
# Mod1 13 10902.94     17.75     0      1 -5438.47

# Plots
library(sjPlot)
tab_model(m2bin_roads_reg2)

plot_model(m2bin_roads_reg, type = c("eff"), transform = NULL, terms = c("region"))

plot_model(m2bin_roads_reg2, type = c("eff"), transform = NULL, terms = c("distvei2 [0:500]", "region [Sor-Trondelag]"))

## beta
library(glmmTMB)
load("processdata/browse9.6_regionscpos.rda")
load("output/m3betaroad_reg.rda")


m3betaroad_reg <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + region + (1|KOMNR),
            data = browse9.6_regionscpos,
            family = beta_family())


save(m3betaroad_reg, file = "m3betaroad_reg.rda")

summary(m3betaroad_reg)

# AIC ----
AIC(m3betaroad, m3betaroad_reg)
#                df       AIC
# m3betaroad     13 -603.6964
# m3betaroad_reg 26 -671.3960

library(bbmle)
ICtab(m3betaroad, m3betaroad_reg, type = "AIC", weights = TRUE, delta = TRUE)
#                dAIC df weight
# m3betaroad_reg  0.0 26 1     
# m3betaroad     67.7 13 <0.001

# Plots
library(sjPlot)
tab_model(m3betaroad_reg)
plot_model(m3betaroad_reg, type = c("eff"), transform = NULL, terms = c("region"))

```



#### Adding moose_density to the model
```{r}
library(lme4)
library(glmmTMB)

##### Houses
## BIN
load("output_copy/mhusabin.rda")
rh_bin <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + moose_density + (1|KOMNR),
family = binomial(),
data = browse9.5msc_test)

rh_bin_int <- glmer(factor(proplt > 0) ~ disthus_400 + log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + kant + helling + HOH + I(HOH^2) +  ac + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
family = binomial(),
data = browse9.5msc_test)


library(AICcmodavg)
candmods_BP_hbin <- list(rh_bin, mhusabin, rh_bin_int)
aictab(candmods_BP_hbin, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)
#       K     AICc Delta_AICc AICcWt Cum.Wt       LL
# Mod1 18 14206.75       0.00   0.70   0.70 -7085.35
# Mod3 19 14208.68       1.92   0.27   0.97 -7085.30
# Mod2 17 14213.08       6.33   0.03   1.00 -7089.51

summary(rh_bin) # moose density negative **


# library(sjPlot)
# library(ggplot2)
# p1 <- plot_model(bh_bin, type = c("eff"), transform = NULL, terms = c("moose_density"),
#            grid = TRUE)
# p1 <- p1 + theme_light() +
#   theme(plot.title = element_blank(),
#     axis.text = element_text(family = "sans", size = 20, color = "grey30"),
#         axis.title = element_text(family = "sans", size = 20, color = "grey30")) +
#   xlab("5 year mean moose density\n") +
#   ylab("Browsing pressure (%)\n")


## BETA
load("output_copy/m16abeta_house.rda")
rh_beta <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + moose_density + (1|KOMNR),
            data = browse9.5msc_test_pos,
            family = beta_family())

rh_beta_int <- glmmTMB(proplt ~ log(distvei2+1) + skogkategori + treartgruppe9 + HOH + I(HOH^2) + browspres_merged + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
            data = browse9.5msc_test_pos,
            family = beta_family())


library(bbmle)
AIC(rh_beta, m16abeta_house, rh_beta_int)
ICtab(rh_beta, m16abeta_house, rh_beta_int, type = "AIC", weights = TRUE, delta = TRUE)

summary(rh_beta) # moose density not sign.




##### Roads
library(lme4)
library(glmmTMB)


## BIN
load("output_copy/m2bin_roads.rda")
rr_bin <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + moose_density + (1|KOMNR),
family = binomial(),
data = browse9.6sc_test)

rr_bin_int <- glmer(factor(proplt > 0) ~ log(distvei2+1) + skogkategori + treartgruppe9 + browspres_merged + helling + HOH + I(HOH^2) +  ac + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
family = binomial(),
data = browse9.6sc_test)


library(AICcmodavg)
candmods_BP_hbeta <- list(rr_bin, m2bin_roads, rr_bin_int)
aictab(candmods_BP_hbeta, modnames = NULL,
                       second.ord = TRUE, nobs = NULL, sort = TRUE)
#       K     AICc Delta_AICc AICcWt Cum.Wt       LL
# Mod1 14 10898.69       0.00   0.62   0.62 -5435.32
# Mod3 15 10900.05       1.36   0.31   0.93 -5435.00
# Mod2 13 10902.98       4.29   0.07   1.00 -5438.47

summary(rr_bin) # moose density negativ *


## BETA
load("output_copy/m3betaroad.rda")
rr_beta <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + moose_density + (1|KOMNR),
            data = browse9.6sc_test_pos,
            family = beta_family())

rr_beta_int <- glmmTMB(proplt ~ skogkategori + treartgruppe9 + ac + helling + HOH + I(HOH^2) + browspres_merged + moose_density + moose_density*log(distvei2+1) + (1|KOMNR),
            data = browse9.6sc_test_pos,
            family = beta_family())


library(bbmle)
AIC(rr_beta, m3betaroad, rr_beta_int)
ICtab(rr_beta, m3betaroad, rr_beta_int, type = "AIC", weights = TRUE, delta = TRUE)

summary(rr_beta) # moose density not sign.
```





## Test of model
```{r}
library(ggplot2)
library(ggpubr)
ggplot(recruits_test_rep, aes(proplt)) +
  geom_histogram(fill = "#0073C2FF") +
  theme_pubclean()

str(recruits_test_rep$proplt)
summary(recruits_test_rep$proplt)

library(glmmTMB)



#### Test for zero-inflation

# Use glmmTMB package to run glm
library("glmmTMB")
m_test_zi <- glm(proplt ~ , family = poisson, data = recruits_test_rep)

library("performance")
check_zeroinflation(m_test_zi)




## Poisson
full_zipo <- glmmTMB(tree_layer_c ~ method + years_since + (1|site_id_new.x), data = plots_full_richness, ziformula = ~1, family = poisson)

## Negative binomial "NB2" (variance = µ(1 + µ/k))
full_zibinom <- update(full_zipo,family = nbinom2)
## Negative binomial "NB1" (variance = φµ)
full_zibinom1 <- update(full_zipo,family = nbinom1)

## Fit updated to hurdle model
full_hnzibinom1 <- update(full_zibinom1, ziformula=~., data=plots_full_richness, family=list(family="truncated_nbinom1", link="log"))

full_hnzibinom <- update(full_zibinom, ziformula=~., data=plots_full_richness, family=list(family="truncated_nbinom2", link="log"))

```
